refactor: Use kebab-case filenames in atoms/src
Renames camelCase modules (colorNames, domCore, className, linkText,
tagName, executeScript, findElement, localStorage, sessionStorage) to
kebab-case for naming consistency across atoms/src, and updates all
import specifiers accordingly.
