+++ atoms/click.js
@@ -1,3 +1,3 @@
-function(){"use strict";var AtomExport=(()=>{var ee=Object.defineProperty;var Ge=Object.getOwnPropertyDescriptor;var je=Object.getOwnPropertyNames;var $e=Object.prototype.hasOwnProperty;var ze=(n,e)=>{for(var t in e)ee(n,t,{get:e[t],enumerable:!0})},Je=(n,e,t,o)=>{if(e&&typeof e=="object"||typeof e=="function")for(let s of je(e))!$e.call(n,s)&&s!==t&&ee(n,s,{get:()=>e[s],enumerable:!(o=Ge(e,s))||o.enumerable});return n};var qe=n=>Je(ee({},"__esModule",{value:!0}),n);var Vt={};ze(Vt,{default:()=>We});var we={aliceblue:"#f0f8ff",antiquewhite:"#faebd7",aqua:"#00ffff",aquamarine:"#7fffd4",azure:"#f0ffff",beige:"#f5f5dc",bisque:"#ffe4c4",black:"#000000",blanchedalmond:"#ffebcd",blue:"#0000ff",blueviolet:"#8a2be2",brown:"#a52a2a",burlywood:"#deb887",cadetblue:"#5f9ea0",chartreuse:"#7fff00",chocolate:"#d2691e",coral:"#ff7f50",cornflowerblue:"#6495ed",cornsilk:"#fff8dc",crimson:"#dc143c",cyan:"#00ffff",darkblue:"#00008b",darkcyan:"#008b8b",darkgoldenrod:"#b8860b",darkgray:"#a9a9a9",darkgreen:"#006400",darkgrey:"#a9a9a9",darkkhaki:"#bdb76b",darkmagenta:"#8b008b",darkolivegreen:"#556b2f",darkorange:"#ff8c00",darkorchid:"#9932cc",darkred:"#8b0000",darksalmon:"#e9967a",darkseagreen:"#8fbc8f",darkslateblue:"#483d8b",darkslategray:"#2f4f4f",darkslategrey:"#2f4f4f",darkturquoise:"#00ced1",darkviolet:"#9400d3",deeppink:"#ff1493",deepskyblue:"#00bfff",dimgray:"#696969",dimgrey:"#696969",dodgerblue:"#1e90ff",firebrick:"#b22222",floralwhite:"#fffaf0",forestgreen:"#228b22",fuchsia:"#ff00ff",gainsboro:"#dcdcdc",ghostwhite:"#f8f8ff",gold:"#ffd700",goldenrod:"#daa520",gray:"#808080",green:"#008000",greenyellow:"#adff2f",grey:"#808080",honeydew:"#f0fff0",hotpink:"#ff69b4",indianred:"#cd5c5c",indigo:"#4b0082",ivory:"#fffff0",khaki:"#f0e68c",lavender:"#e6e6fa",lavenderblush:"#fff0f5",lawngreen:"#7cfc00",lemonchiffon:"#fffacd",lightblue:"#add8e6",lightcoral:"#f08080",lightcyan:"#e0ffff",lightgoldenrodyellow:"#fafad2",lightgray:"#d3d3d3",lightgreen:"#90ee90",lightgrey:"#d3d3d3",lightpink:"#ffb6c1",lightsalmon:"#ffa07a",lightseagreen:"#20b2aa",lightskyblue:"#87cefa",lightslategray:"#778899",lightslategrey:"#778899",lightsteelblue:"#b0c4de",lightyellow:"#ffffe0",lime:"#00ff00",limegreen:"#32cd32",linen:"#faf0e6",magenta:"#ff00ff",maroon:"#800000",mediumaquamarine:"#66cdaa",mediumblue:"#0000cd",mediumorchid:"#ba55d3",mediumpurple:"#9370db",mediumseagreen:"#3cb371",mediumslateblue:"#7b68ee",mediumspringgreen:"#00fa9a",mediumturquoise:"#48d1cc",mediumvioletred:"#c71585",midnightblue:"#191970",mintcream:"#f5fffa",mistyrose:"#ffe4e1",moccasin:"#ffe4b5",navajowhite:"#ffdead",navy:"#000080",oldlace:"#fdf5e6",olive:"#808000",olivedrab:"#6b8e23",orange:"#ffa500",orangered:"#ff4500",orchid:"#da70d6",palegoldenrod:"#eee8aa",palegreen:"#98fb98",paleturquoise:"#afeeee",palevioletred:"#db7093",papayawhip:"#ffefd5",peachpuff:"#ffdab9",peru:"#cd853f",pink:"#ffc0cb",plum:"#dda0dd",powderblue:"#b0e0e6",purple:"#800080",red:"#ff0000",rosybrown:"#bc8f8f",royalblue:"#4169e1",saddlebrown:"#8b4513",salmon:"#fa8072",sandybrown:"#f4a460",seagreen:"#2e8b57",seashell:"#fff5ee",sienna:"#a0522d",silver:"#c0c0c0",skyblue:"#87ceeb",slateblue:"#6a5acd",slategray:"#708090",slategrey:"#708090",snow:"#fffafa",springgreen:"#00ff7f",steelblue:"#4682b4",tan:"#d2b48c",teal:"#008080",thistle:"#d8bfd8",tomato:"#ff6347",turquoise:"#40e0d0",violet:"#ee82ee",wheat:"#f5deb3",white:"#ffffff",whitesmoke:"#f5f5f5",yellow:"#ffff00",yellowgreen:"#9acd32"};var Ze=new Set(["backgroundColor","borderTopColor","borderRightColor","borderBottomColor","borderLeftColor","color","outlineColor"]),Qe=/#([0-9a-fA-F])([0-9a-fA-F])([0-9a-fA-F])/,et=/^#(?:[0-9a-f]{3}){1,2}$/i,tt=/^(?:rgba)?\((\d{1,3}),\s?(\d{1,3}),\s?(\d{1,3}),\s?(0|1|0\.\d*)\)$/i,nt=/^(?:rgb)?\((0|[1-9]\d{0,2}),\s?(0|[1-9]\d{0,2}),\s?(0|[1-9]\d{0,2})\)$/i;function ot(n){n=n.toLowerCase();let e=we[n];if(!e&&(e=n.charAt(0)==="#"?n:`#${n}`,e.length===4&&(e=e.replace(Qe,"#$1$1$2$2$3$3")),!et.test(e)))return null;let t=parseInt(e.substring(1,3),16),o=parseInt(e.substring(3,5),16),s=parseInt(e.substring(5,7),16);return[t,o,s,1]}function st(n){let e=n.match(tt);if(e){let t=Number(e[1]),o=Number(e[2]),s=Number(e[3]),i=Number(e[4]);if(t>=0&&t<=255&&o>=0&&o<=255&&s>=0&&s<=255&&i>=0&&i<=1)return[t,o,s,i]}return null}function rt(n){let e=n.match(nt);if(e){let t=Number(e[1]),o=Number(e[2]),s=Number(e[3]);if(t>=0&&t<=255&&o>=0&&o<=255&&s>=0&&s<=255)return[t,o,s,1]}return null}function ve(n,e){if(!Ze.has(n))return e;let t=st(e)||rt(e)||ot(e);return t?`rgba(${t.join(", ")})`:e}var it={15:"element not selectable",11:"element not visible",31:"unknown error",30:"unknown error",24:"invalid cookie domain",29:"invalid element coordinates",12:"invalid element state",32:"invalid selector",51:"invalid selector",52:"invalid selector",17:"javascript error",405:"unsupported operation",34:"move target out of bounds",27:"no such alert",7:"no such element",8:"no such frame",23:"no such window",28:"script timeout",33:"session not created",10:"stale element reference",21:"timeout",25:"unable to set cookie",26:"unexpected alert open",13:"unknown error",9:"unknown command"};function te(n){return it[n]??"unknown error"}var p=class extends Error{code;state;isAutomationError=!0;constructor(e,t=""){super(t),this.code=e,this.state=te(e);let o=this.state.replace(/(^|\s+)[a-z]/g,i=>i.toUpperCase().trimStart()).replace(/\s+/g,"");this.name=/Error$/.test(o)?o:`${o}Error`;let s=new Error(this.message);s.name=this.name,this.stack=s.stack||""}};var lt=/[;]+(?=(?:(?:[^"]*"){2})*[^"]*$)(?=(?:(?:[^']*'){2})*[^']*$)(?=(?:[^()]*\([^()]*\))*[^()]*$)/;function ut(n){let e=n.split(lt),t=[];for(let s of e){let i=s.indexOf(":");if(i>0){let a=[s.slice(0,i),s.slice(i+1)];t.push(a[0].toLowerCase(),":",a[1],";")}}let o=t.join("");return o.charAt(o.length-1)===";"?o:`${o};`}function Se(n,e){if(e=e.toLowerCase(),e==="style")return ut(n.style.cssText);let t=n.getAttributeNode(e);return t&&t.specified?t.value:null}function x(n,e){return n[e]}function c(n,e){return n instanceof HTMLFormElement?n.nodeType===Node.ELEMENT_NODE&&(!e||e==="FORM"):!!n&&n.nodeType===Node.ELEMENT_NODE&&(!e||n.tagName.toUpperCase()===e)}function K(n){return n.nodeType===Node.DOCUMENT_NODE?n:n.ownerDocument}function ne(n,e,t=!1){let o=t?n:n.parentNode;for(;o;){if(e(o))return o;o=o.parentNode}return null}function k(n){if(c(n,"OPTION"))return!0;if(c(n,"INPUT")){let e=n.type.toLowerCase();return e==="checkbox"||e==="radio"}return!1}function F(n){if(!k(n))throw new p(15,"Element is not selectable");let e=n.type?.toLowerCase();return!!x(n,e==="checkbox"||e==="radio"?"checked":"selected")}function Re(n,e){if(!n)throw new p(32,"No selector specified");n=n.trim();let t;try{t=e.querySelector(n)}catch{throw new p(32,"An invalid or illegal selector was specified")}return t&&t.nodeType===Node.ELEMENT_NODE?t:null}var w=class{x;y;constructor(e=0,t=0){this.x=e,this.y=t}},H=class n{x;y;constructor(e,t){this.x=e,this.y=t}magnitude(){return Math.hypot(this.x,this.y)}scale(e,t=e){return this.x*=e,this.y*=t,this}subtract(e){return this.x-=e.x,this.y-=e.y,this}rotate(e){let t=Math.cos(e),o=Math.sin(e),s=this.x*t-this.y*o,i=this.y*t+this.x*o;return this.x=s,this.y=i,this}static fromCoordinate(e){return new n(e.x,e.y)}static sum(e,t){return new n(e.x+t.x,e.y+t.y)}static difference(e,t){return new n(e.x-t.x,e.y-t.y)}},B=class{width;height;constructor(e,t){this.width=e,this.height=t}},V=class{top;right;bottom;left;constructor(e,t,o,s){this.top=e,this.right=t,this.bottom=o,this.left=s}},g=class n{left;top;width;height;constructor(e,t,o,s){this.left=e,this.top=t,this.width=o,this.height=s}toBox(){let e=this.left+this.width,t=this.top+this.height;return new V(this.top,e,t,this.left)}intersects(e){return n.intersects(this,e)}static intersects(e,t){return e.left<=t.left+t.width&&t.left<=e.left+e.width&&e.top<=t.top+t.height&&t.top<=e.top+e.height}};function W(n,e,t){return Math.min(Math.max(n,e),t)}var ct=typeof ShadowRoot=="function";function Et(n){return n.replace(/-([a-z])/g,(e,t)=>t.toUpperCase())}function $(n){let e="document"in n?n.document:K(n);try{let t=e?.activeElement;return t&&t.nodeName?t:null}catch{return null}}function R(n){return z(n,!0)&&Ie(n)&&!ft(n)}function ft(n){return N(n,"pointer-events")==="none"}var dt=["A","AREA","BUTTON","INPUT","LABEL","SELECT","TEXTAREA"];function Ce(n){return dt.some(e=>c(n,e))||Se(n,"tabindex")!=null&&Number(x(n,"tabIndex"))>=0||se(n)}var mt=["BUTTON","INPUT","OPTGROUP","OPTION","SELECT","TEXTAREA"];function Ie(n){if(!mt.some(o=>c(n,o)))return!0;if(x(n,"disabled"))return!1;if(n.parentNode&&n.parentNode.nodeType===Node.ELEMENT_NODE&&c(n,"OPTGROUP")||c(n,"OPTION"))return Ie(n.parentNode);let t=n;for(;t;){let o=t.parentNode;if(o&&c(o,"FIELDSET")&&x(o,"disabled")){if(!c(t,"LEGEND"))return!1;let s=t.previousElementSibling;for(;s;){if(c(s,"LEGEND"))return!1;s=s.previousElementSibling}}t=o}return!0}var ht=new Set(["text","search","tel","url","email","password","number"]);function pt(n){if(c(n,"TEXTAREA"))return!0;if(c(n,"INPUT")){let e=n.type.toLowerCase();return ht.has(e)}return Me(n)}function Tt(n){return c(n,"INPUT")&&n.type.toLowerCase()==="file"}function A(n,e){return c(n,"INPUT")&&n.type.toLowerCase()===e}function Me(n){let e=n;if(e.contentEditable===void 0)return!1;if(e.isContentEditable!==void 0)return e.isContentEditable;function t(o){if(o.contentEditable==="inherit"){let s=I(o);return s?t(s):!1}return o.contentEditable==="true"}return t(e)}function se(n){return(pt(n)||Tt(n)||A(n,"range")||A(n,"date")||A(n,"month")||A(n,"week")||A(n,"time")||A(n,"datetime-local")||A(n,"color"))&&!x(n,"readOnly")}function I(n){let e=n.parentNode;for(;e&&e.nodeType!==Node.ELEMENT_NODE&&e.nodeType!==Node.DOCUMENT_NODE&&e.nodeType!==Node.DOCUMENT_FRAGMENT_NODE;)e=e.parentNode;return e&&c(e)?e:null}function bt(n,e){let t=n.ownerDocument.defaultView;if(t?.getComputedStyle){let o=t.getComputedStyle(n,null);if(o)return o[e]||o.getPropertyValue(e)||""}return""}function _e(n,e){let t=n.style,o=t[e];if(o===void 0&&typeof t.getPropertyValue=="function"&&(o=t.getPropertyValue(e)),o!=="inherit")return o!==void 0?o:null;let s=I(n);return s?_e(s,e):null}function N(n,e){let t=Et(e);(t==="float"||t==="cssFloat"||t==="styleFloat")&&(t="cssFloat");let o=bt(n,t)||_e(n,t);return o===null?null:ve(t,o)}function oe(n,e,t){if(!c(n))throw new Error("Argument to isShown must be of type Element");if(c(n,"BODY"))return!0;if(c(n,"OPTION")||c(n,"OPTGROUP")){let E=n.parentNode;for(;E&&!c(E,"SELECT");)E=E.parentNode;return!!E&&oe(E,!0,t)}let o=Ae(n);if(o)return!!o.image&&o.rect.width>0&&o.rect.height>0&&oe(o.image,e,t);if(c(n,"INPUT")&&n.type.toLowerCase()==="hidden"||c(n,"NOSCRIPT"))return!1;let s=N(n,"visibility");if(s==="collapse"||s==="hidden"||!t(n)||!e&&Le(n)===0)return!1;function i(E){let d=C(E);if(d.height>0&&d.width>0)return!0;if(c(E,"PATH")&&(d.height>0||d.width>0)){let T=N(E,"stroke-width");return!!T&&parseInt(T,10)>0}let y=N(E,"visibility");return y==="collapse"||y==="hidden"||!t(E)?!1:N(E,"overflow")!=="hidden"&&[...E.childNodes].some(T=>{if(T.nodeType===Node.TEXT_NODE){let v=T.nodeValue??"";return!(/^[\s]*$/.test(v)&&/[\n\r\t]/.test(v))}return c(T)&&i(T)})}if(!i(n))return!1;function a(E){return L(E)==="hidden"&&[...E.childNodes].every(d=>!c(d)||a(d)||!i(d))}return!a(n)}function z(n,e=!1){function t(o){if(c(o)){let i=o;if(N(i,"display")==="none"||N(i,"content-visibility")==="hidden")return!1}let s=gt(o);if(ct&&s instanceof ShadowRoot){if(s.host.shadowRoot&&s.host.shadowRoot!==s)return!1;s=s.host}return s&&(s.nodeType===Node.DOCUMENT_NODE||s.nodeType===Node.DOCUMENT_FRAGMENT_NODE)?!0:s&&c(s,"DETAILS")&&!s.open&&!c(o,"SUMMARY")?!1:!!s&&t(s)}return oe(n,e,t)}function J(n){let e=n.scrollingElement||n.documentElement,t=n.defaultView;return new w(t.pageXOffset||e.scrollLeft,t.pageYOffset||e.scrollTop)}function L(n,e){let t=q(n,e),o=K(n),s=o.documentElement,i=o.body,a=N(s,"overflow"),E=!1;function d(m,b){if(m===s)return!0;let h=N(m,"display");return!(h?.startsWith("inline")||h==="contents"||b==="absolute"&&N(m,"position")==="static")}function y(m){let b=N(m,"position");if(b==="fixed")return E=!0,m===s?null:s;let h=I(m);for(;h&&!d(h,b);)h=I(h);return h}function T(m){let b=m;if(a==="visible"){if(m===s&&i)b=i;else if(m===i)return{x:"visible",y:"visible"}}let h={x:N(b,"overflow-x")??"",y:N(b,"overflow-y")??""};return m===s&&(h.x=h.x==="visible"?"auto":h.x,h.y=h.y==="visible"?"auto":h.y),h}function v(m){return m===s?J(o):new w(m.scrollLeft,m.scrollTop)}for(let m=y(n);m;m=y(m)){let b=T(m);if(b.x==="visible"&&b.y==="visible")continue;let h=C(m);if(h.width===0||h.height===0)return"hidden";let Ne=t.right<h.left,ge=t.bottom<h.top;if(Ne&&b.x==="hidden"||ge&&b.y==="hidden")return"hidden";if(Ne&&b.x!=="visible"||ge&&b.y!=="visible"){let U=v(m),Xe=t.right<h.left-U.x,Ye=t.bottom<h.top-U.y;return Xe&&b.x!=="visible"||Ye&&b.x!=="visible"||L(m)==="hidden"?"hidden":"scroll"}let Oe=t.left>=h.left+h.width,ye=t.top>=h.top+h.height;if(Oe&&b.x==="hidden"||ye&&b.y==="hidden")return"hidden";if(Oe&&b.x!=="visible"||ye&&b.y!=="visible"){if(E){let U=v(m);if(t.left>=s.scrollWidth-U.x||t.right>=s.scrollHeight-U.y)return"hidden"}return L(m)==="hidden"?"hidden":"scroll"}}return"none"}function C(n){let e=Ae(n);if(e)return e.rect;if(c(n,"HTML")){let o=K(n),s=o.defaultView,i=o.compatMode==="CSS1Compat"?o.documentElement:o.body;return new g(0,0,i.clientWidth||s.innerWidth,i.clientHeight||s.innerHeight)}let t;try{t=n.getBoundingClientRect()}catch{return new g(0,0,0,0)}return new g(t.left,t.top,t.right-t.left,t.bottom-t.top)}function Ae(n){let e=c(n,"MAP");if(!e&&!c(n,"AREA"))return null;let t=e?n:n.parentNode&&c(n.parentNode,"MAP")?n.parentNode:null,o=null,s=null,i=t&&t.name;if(t&&i){let a=K(t),E=`*[usemap="#${i}"]`;if(o=Re(E,a),o&&(s=C(o),!e&&n.shape.toLowerCase()!=="default")){let d=Nt(n),y=Math.min(Math.max(d.left,0),s.width),T=Math.min(Math.max(d.top,0),s.height),v=Math.min(d.width,s.width-y),m=Math.min(d.height,s.height-T);s=new g(y+s.left,T+s.top,v,m)}}return{image:o,rect:s||new g(0,0,0,0)}}function Nt(n){let e=n,t=e.shape.toLowerCase(),o=e.coords.split(",");if(t==="rect"&&o.length===4){let s=Number(o[0]),i=Number(o[1]);return new g(s,i,Number(o[2])-s,Number(o[3])-i)}else if(t==="circle"&&o.length===3){let s=Number(o[0]),i=Number(o[1]),a=Number(o[2]);return new g(s-a,i-a,2*a,2*a)}else if(t==="poly"&&o.length>2){let s=Number(o[0]),i=Number(o[1]),a=s,E=i;for(let d=2;d+1<o.length;d+=2)s=Math.min(s,Number(o[d])),a=Math.max(a,Number(o[d])),i=Math.min(i,Number(o[d+1])),E=Math.max(E,Number(o[d+1]));return new g(s,i,a-s,E-i)}return new g(0,0,0,0)}function q(n,e){let t=C(n).toBox();if(e){let o=e instanceof g?e:new g(e.x,e.y,1,1);t.left=W(t.left+o.left,t.left,t.right),t.top=W(t.top+o.top,t.top,t.bottom),t.right=W(t.left+o.width,t.left,t.right),t.bottom=W(t.top+o.height,t.top,t.bottom)}return t}function Le(n){let e=1,t=N(n,"opacity");t&&(e=Number(t));let o=I(n);return o&&(e*=Le(o)),e}function gt(n){let e=n.parentNode;if(e&&e.shadowRoot&&n.assignedSlot!==void 0)return n.assignedSlot?n.assignedSlot.parentNode:null;let t=n;if(t.getDestinationInsertionPoints){let o=t.getDestinationInsertionPoints();if(o.length>0)return o[o.length-1]}return e}var O=class{type;bubbles;cancelable;constructor(e,t,o){this.type=e,this.bubbles=t,this.cancelable=o}create(e,t){let s=("ownerDocument"in e?e.ownerDocument:e.document).createEvent("HTMLEvents");return s.initEvent(this.type,this.bubbles,this.cancelable),s}toString(){return this.type}},S=class extends O{create(e,t){let o=t,s="ownerDocument"in e?e.ownerDocument:e.document,i=s.defaultView,a=s.createEvent("MouseEvents"),E=1;return this.type==="mousewheel"&&(a.wheelDelta=o.wheelDelta),a.initMouseEvent(this.type,this.bubbles,this.cancelable,i,E,o.clientX,o.clientY,o.clientX,o.clientY,o.ctrlKey,o.altKey,o.shiftKey,o.metaKey,o.button,o.relatedTarget),a}},X=class extends O{create(e,t){let o=t,i=("ownerDocument"in e?e.ownerDocument:e.document).createEvent("Events");return i.initEvent(this.type,this.bubbles,this.cancelable),i.altKey=o.altKey,i.ctrlKey=o.ctrlKey,i.metaKey=o.metaKey,i.shiftKey=o.shiftKey,i.keyCode=o.charCode||o.keyCode,i.charCode=this.type==="keypress"?i.keyCode:0,i}};function re(n,e){return e.map(t=>new Touch({identifier:t.identifier,target:n,screenX:t.screenX,screenY:t.screenY,clientX:t.clientX,clientY:t.clientY,pageX:t.pageX,pageY:t.pageY}))}var Y=class extends O{create(e,t){let o=t,s="ownerDocument"in e?e.ownerDocument:e.document,i=s.defaultView,a=typeof TouchEvent.prototype.initTouchEvent=="function"?"INIT_TOUCH_EVENT":"TOUCH_EVENT_CTOR",E=re(e,o.changedTouches),d=o.touches===o.changedTouches?E:re(e,o.touches),y=o.targetTouches===o.changedTouches?E:re(e,o.targetTouches);if(a==="INIT_TOUCH_EVENT"){let T=s.createEvent("TouchEvent");return T.initTouchEvent(this.type,this.bubbles,this.cancelable,i,1,0,0,o.clientX,o.clientY,o.ctrlKey,o.altKey,o.shiftKey,o.metaKey,d,y,E,o.scale,o.rotation),T.relatedTarget=o.relatedTarget,T}return new TouchEvent(this.type,{touches:d,targetTouches:y,changedTouches:E,bubbles:this.bubbles,cancelable:this.cancelable,ctrlKey:o.ctrlKey,shiftKey:o.shiftKey,altKey:o.altKey,metaKey:o.metaKey})}},f={BLUR:new O("blur",!1,!1),CHANGE:new O("change",!0,!1),FOCUS:new O("focus",!1,!1),FOCUSIN:new O("focusin",!0,!1),FOCUSOUT:new O("focusout",!0,!1),INPUT:new O("input",!0,!1),ORIENTATIONCHANGE:new O("orientationchange",!1,!1),SELECT:new O("select",!0,!1),SUBMIT:new O("submit",!0,!0),TEXTINPUT:new O("textInput",!0,!0),CLICK:new S("click",!0,!0),CONTEXTMENU:new S("contextmenu",!0,!0),DBLCLICK:new S("dblclick",!0,!0),MOUSEDOWN:new S("mousedown",!0,!0),MOUSEMOVE:new S("mousemove",!0,!1),MOUSEOUT:new S("mouseout",!0,!0),MOUSEOVER:new S("mouseover",!0,!0),MOUSEUP:new S("mouseup",!0,!0),MOUSEWHEEL:new S("mousewheel",!0,!0),KEYDOWN:new X("keydown",!0,!0),KEYPRESS:new X("keypress",!0,!0),KEYUP:new X("keyup",!0,!0),TOUCHEND:new Y("touchend",!0,!0),TOUCHMOVE:new Y("touchmove",!0,!0),TOUCHSTART:new Y("touchstart",!0,!0)};function M(n,e,t){let o=e.create(n,t);return"isTrusted"in o||(o.isTrusted=!1),n.dispatchEvent(o)}var ie=class{pressedModifiers=0;isPressed(e){return(this.pressedModifiers&e)!==0}setPressed(e,t){t?this.pressedModifiers|=e:this.pressedModifiers&=~e}isShiftPressed(){return this.isPressed(1)}isControlPressed(){return this.isPressed(2)}isAltPressed(){return this.isPressed(4)}isMetaPressed(){return this.isPressed(8)}},le=class{fireHtmlEvent(e,t){return M(e,t)}fireKeyboardEvent(e,t,o){return M(e,t,o)}fireMouseEvent(e,t,o){return M(e,t,o)}fireTouchEvent(e,t,o){return M(e,t,o)}};function Ot(n){return c(n,"FORM")}var D=class{el;select=null;modifiersState;eventEmitter;constructor(e,t){this.el=document.documentElement;let o=$(this.el);o&&this.setElement(o),this.modifiersState=e||new ie,this.eventEmitter=t||new le}getElement(){return this.el}setElement(e){this.el=e,c(e,"OPTION")?this.select=ne(e,t=>c(t,"SELECT")):this.select=null}fireHtmlEvent(e){return this.eventEmitter.fireHtmlEvent(this.el,e)}fireKeyboardEvent(e,t){return this.eventEmitter.fireKeyboardEvent(this.el,e,t)}fireMouseEvent(e,t,o,s=null,i=0,a=!1){if(!a&&!R(this.el))return!1;if(s&&!(e===f.MOUSEOVER||e===f.MOUSEOUT))throw new p(12,`Event type does not allow related target: ${e}`);let E={clientX:t.x,clientY:t.y,button:o,altKey:this.modifiersState.isAltPressed(),ctrlKey:this.modifiersState.isControlPressed(),shiftKey:this.modifiersState.isShiftPressed(),metaKey:this.modifiersState.isMetaPressed(),wheelDelta:i,relatedTarget:s},d=this.select?this.getTargetOfOptionMouseEvent(e):this.el;return d?this.eventEmitter.fireMouseEvent(d,e,E):!0}fireTouchEvent(e,t,o,s,i){let a={touches:[],targetTouches:[],changedTouches:[],altKey:this.modifiersState.isAltPressed(),ctrlKey:this.modifiersState.isControlPressed(),shiftKey:this.modifiersState.isShiftPressed(),metaKey:this.modifiersState.isMetaPressed(),relatedTarget:null,scale:0,rotation:0,clientX:o.x,clientY:o.y},E=J(this.el.ownerDocument),d=(y,T)=>{let v={identifier:y,screenX:T.x,screenY:T.y,clientX:T.x,clientY:T.y,pageX:T.x+E.x,pageY:T.y+E.y};a.changedTouches.push(v),(e===f.TOUCHSTART||e===f.TOUCHMOVE)&&(a.touches.push(v),a.targetTouches.push(v))};return d(t,o),s!==void 0&&i!==void 0&&d(s,i),this.eventEmitter.fireTouchEvent(this.el,e,a)}getTargetOfOptionMouseEvent(e){let t=this.select;switch(e){case f.CLICK:case f.MOUSEUP:return t.multiple?this.el:this.select;default:return t.multiple?this.el:null}}clickElement(e,t,o=!1){if(!o&&!R(this.el))return;let s=!this.select&&k(this.el),i=s&&F(this.el);this.fireMouseEvent(f.CLICK,e,t,null,0,o)&&s&&this.toggleRadioButtonOrCheckbox(i)}maybeToggleOption(){if(!this.select||!R(this.el))return;let e=this.select,t=F(this.el);t&&!e.multiple||(this.el.selected=!t,e.multiple||M(e,f.CHANGE))}toggleRadioButtonOrCheckbox(e){}focusOnElement(){let e=ne(this.el,o=>!!o&&c(o)&&Ce(o),!0);e=e||this.el;let t=$(e);return e===t?!1:(t&&typeof t.blur=="function"&&(c(t,"BODY")||t.blur()),typeof e.focus=="function"?(e.focus(),!0):!1)}submitForm(e){if(!Ot(e))throw new p(12,"Element is not a form, so could not submit.");M(e,f.SUBMIT)&&(c(e.submit)?HTMLFormElement.prototype.submit.call(e):e.submit())}};function yt(n){return navigator.userAgent.includes(n)}function ue(){return yt("Macintosh")}var ce=class{code;character;shiftChar;constructor(e,t,o){this.code=e,this.character=t||null,this.shiftChar=o||this.character}},ae=new Map;function r(n,e,t){let o=new ce(n,e,t);return e&&(!ae.has(e)||t)&&(ae.set(e,{key:o,shift:!1}),t&&ae.set(t,{key:o,shift:!0})),o}var l={BACKSPACE:r(8),TAB:r(9),ENTER:r(13),SHIFT:r(16),CONTROL:r(17),ALT:r(18),PAUSE:r(19),CAPS_LOCK:r(20),ESC:r(27),SPACE:r(32," "),PAGE_UP:r(33),PAGE_DOWN:r(34),END:r(35),HOME:r(36),LEFT:r(37),UP:r(38),RIGHT:r(39),DOWN:r(40),PRINT_SCREEN:r(44),INSERT:r(45),DELETE:r(46),ZERO:r(48,"0",")"),ONE:r(49,"1","!"),TWO:r(50,"2","@"),THREE:r(51,"3","#"),FOUR:r(52,"4","$"),FIVE:r(53,"5","%"),SIX:r(54,"6","^"),SEVEN:r(55,"7","&"),EIGHT:r(56,"8","*"),NINE:r(57,"9","("),A:r(65,"a","A"),B:r(66,"b","B"),C:r(67,"c","C"),D:r(68,"d","D"),E:r(69,"e","E"),F:r(70,"f","F"),G:r(71,"g","G"),H:r(72,"h","H"),I:r(73,"i","I"),J:r(74,"j","J"),K:r(75,"k","K"),L:r(76,"l","L"),M:r(77,"m","M"),N:r(78,"n","N"),O:r(79,"o","O"),P:r(80,"p","P"),Q:r(81,"q","Q"),R:r(82,"r","R"),S:r(83,"s","S"),T:r(84,"t","T"),U:r(85,"u","U"),V:r(86,"v","V"),W:r(87,"w","W"),X:r(88,"x","X"),Y:r(89,"y","Y"),Z:r(90,"z","Z"),META:r(91),META_RIGHT:r(ue()?93:92),CONTEXT_MENU:r(ue()?0:null),NUM_ZERO:r(96,"0"),NUM_ONE:r(97,"1"),NUM_TWO:r(98,"2"),NUM_THREE:r(99,"3"),NUM_FOUR:r(100,"4"),NUM_FIVE:r(101,"5"),NUM_SIX:r(102,"6"),NUM_SEVEN:r(103,"7"),NUM_EIGHT:r(104,"8"),NUM_NINE:r(105,"9"),NUM_MULTIPLY:r(106,"*"),NUM_PLUS:r(107,"+"),NUM_MINUS:r(109,"-"),NUM_PERIOD:r(110,"."),NUM_DIVISION:r(111,"/"),NUM_LOCK:r(144),F1:r(112),F2:r(113),F3:r(114),F4:r(115),F5:r(116),F6:r(117),F7:r(118),F8:r(119),F9:r(120),F10:r(121),F11:r(122),F12:r(123),EQUALS:r(187,"=","+"),SEPARATOR:r(108,","),HYPHEN:r(189,"-","_"),COMMA:r(188,",","<"),PERIOD:r(190,".",">"),SLASH:r(191,"/","?"),BACKTICK:r(192,"`","~"),OPEN_BRACKET:r(219,"[","{"),BACKSLASH:r(220,"\\","|"),CLOSE_BRACKET:r(221,"]","}"),SEMICOLON:r(186,";",":"),APOSTROPHE:r(222,"'",'"')};var St=[l.ALT,l.CONTROL,l.META,l.SHIFT],vn=new Map([[l.SHIFT.code,1],[l.CONTROL.code,2],[l.ALT.code,4],[l.META.code,8]]);var Rt=3,P=new Map([[f.CLICK,[0,1,2,null]],[f.CONTEXTMENU,[null,null,2,null]],[f.MOUSEUP,[0,1,2,null]],[f.MOUSEOUT,[0,1,2,0]],[f.MOUSEMOVE,[0,1,2,0]]]);P.set(f.DBLCLICK,P.get(f.CLICK));P.set(f.MOUSEDOWN,P.get(f.MOUSEUP));P.set(f.MOUSEOVER,P.get(f.MOUSEOUT));var G=class extends D{buttonPressed=null;elementPressed=null;clientXY=new w(0,0);nextClickIsDoubleClick=!1;hasEverInteracted=!1;constructor(e,t,o){super(t,o),e&&(typeof e.buttonPressed=="number"&&(this.buttonPressed=e.buttonPressed),e.elementPressed&&c(e.elementPressed)&&(this.elementPressed=e.elementPressed),this.clientXY=new w(e.clientXY.x,e.clientXY.y),this.nextClickIsDoubleClick=!!e.nextClickIsDoubleClick,this.hasEverInteracted=!!e.hasEverInteracted,e.element&&c(e.element)&&this.setElement(e.element))}fireMousedown(){return c(this.getElement(),"OPTION")||c(this.getElement(),"SELECT")?!0:this.fireMouseEventInternal(f.MOUSEDOWN)}pressButton(e){if(this.buttonPressed!==null)throw new p(13,"Cannot press more than one button or an already pressed button.");this.buttonPressed=e,this.elementPressed=this.getElement(),this.fireMousedown()&&this.focusOnElement()}releaseButton(e=!1){if(this.buttonPressed===null)throw new p(13,"Cannot release a button when no button is pressed.");this.maybeToggleOption();let t=R(this.getElement());this.fireMouseEventInternal(f.MOUSEUP,void 0,void 0,e);try{this.buttonPressed===0&&this.getElement()===this.elementPressed?(this.clickElement(this.clientXY,this.getButtonValue(f.CLICK),t),this.maybeDoubleClickElement()):this.buttonPressed===2&&this.fireMouseEventInternal(f.CONTEXTMENU)}catch{}this.buttonPressed=null,this.elementPressed=null}maybeDoubleClickElement(){this.nextClickIsDoubleClick&&this.fireMouseEventInternal(f.DBLCLICK),this.nextClickIsDoubleClick=!this.nextClickIsDoubleClick}move(e,t){let o=R(e),s=C(e);this.clientXY.x=t.x+s.left,this.clientXY.y=t.y+s.top;let i=this.getElement();if(e!==i){try{i&&i.ownerDocument.defaultView?.closed&&(i=null)}catch{i=null}if(i){let a=i===document.documentElement||i===document.body;i=!this.hasEverInteracted&&a?null:i,this.fireMouseEventInternal(f.MOUSEOUT,e)}this.setElement(e),this.fireMouseEventInternal(f.MOUSEOVER,i,void 0,o)}this.fireMouseEventInternal(f.MOUSEMOVE,void 0,void 0,o),this.nextClickIsDoubleClick=!1}scroll(e){if(e===0)throw new p(13,"Must scroll a non-zero number of ticks.");let t=e>0?-120:120;for(let o=0;o<Math.abs(e);o++)this.fireMouseEventInternal(f.MOUSEWHEEL,void 0,t)}fireMouseEventInternal(e,t,o,s){return this.hasEverInteracted=!0,this.fireMouseEvent(e,this.clientXY,this.getButtonValue(e),t??null,o??0,s??!1)}getButtonValue(e){let t=P.get(e);if(!t)return 0;let o=this.buttonPressed===null?Rt:this.buttonPressed,s=t[o];if(s===null)throw new p(13,"Event does not permit the specified mouse button.");return s}getState(){return{buttonPressed:this.buttonPressed,elementPressed:this.elementPressed,clientXY:{x:this.clientXY.x,y:this.clientXY.y},nextClickIsDoubleClick:this.nextClickIsDoubleClick,hasEverInteracted:this.hasEverInteracted,element:this.getElement()}}};function Mt(n){if(!z(n,!0))throw new p(11,"Element is not currently visible and may not be manipulated")}function Ee(n,e,t,o){let s=_t(n,e),i=t||new G;i.move(n,s),i.pressButton(0),i.releaseButton(o)}function _t(n,e){if(Mt(n),de(n,e),e)return H.fromCoordinate(e);let t=fe(n);return new H(t.width/2,t.height/2)}function De(n){let e=n,t=e.offsetWidth,o=e.offsetHeight;if(!t&&!o&&e.getBoundingClientRect){let i=e.getBoundingClientRect();return new B(i.right-i.left,i.bottom-i.top)}return new B(t,o)}function Pe(n){if(N(n,"display")!=="none")return De(n);let e=n.style,t=e.display,o=e.visibility,s=e.position;e.visibility="hidden",e.position="absolute",e.display="inline";let i=De(n);return e.display=t,e.position=s,e.visibility=o,i}function fe(n){let e=Pe(n),t=n.offsetParent;return e.width>0&&e.height>0||!t?e:fe(t)}function de(n,e){let t=L(n,e);if(t!=="scroll")return t==="none";if(n.scrollIntoView&&(n.scrollIntoView(),L(n,e)==="none"))return!0;let o=q(n,e);for(let s=I(n);s;s=I(s))At(o,s);return L(n,e)==="none"}function At(n,e){let t=C(e),o=Lt(e),s=n.left-t.left-o.left,i=n.top-t.top-o.top,a=e.clientWidth+n.left-n.right,E=e.clientHeight+n.top-n.bottom;e.scrollLeft+=Math.min(s,Math.max(s-a,0)),e.scrollTop+=Math.min(i,Math.max(i-E,0))}function Lt(n){let e=n.ownerDocument.defaultView?.getComputedStyle(n),t=parseFloat(e?.borderLeftWidth||"0"),o=parseFloat(e?.borderRightWidth||"0"),s=parseFloat(e?.borderTopWidth||"0"),i=parseFloat(e?.borderBottomWidth||"0");return new V(s,o,i,t)}var he="ELEMENT",pe="element-6066-11e4-a52e-4f735466cecf",Te="WINDOW";function Dt(n){return typeof n=="object"&&n!==null||typeof n=="function"}function Pt(n){return Array.isArray(n)?!0:typeof n=="object"&&n!==null&&typeof n.length=="number"}function xt(n){function e(t,o){if(t==null)return null;switch(typeof t){case"string":case"number":case"boolean":return t;case"function":return t.toString();case"object":break;default:return null}let s=t;if(o.includes(s))throw new p(17,"Recursive object cannot be transferred");if("nodeType"in s&&(s.nodeType===Node.ELEMENT_NODE||s.nodeType===Node.DOCUMENT_NODE)){let a=Ue(s);return{[he]:a,[pe]:a}}if("document"in s)return{[Te]:Ue(s)};if(Array.isArray(t))return t.map(a=>e(a,o));if(o=[...o,s],Pt(t))return Array.from(t).map(a=>e(a,o));let i={};for(let a of Object.keys(s))i[a]=e(s[a],o);return i}return e(n,[])}function Q(n,e){if(Array.isArray(n))return n.map(t=>Q(t,e));if(Dt(n)){if(typeof n=="function")return n;let t=n;if(he in t)return j(t[he],e);if(pe in t)return j(t[pe],e);if(Te in t)return j(t[Te],e);let o={};for(let s of Object.keys(t))o[s]=Q(t[s],e);return o}return n}function Ke(n){return{status:0,value:xt(n)}}function ke(n){let e="code"in n&&typeof n.code=="number"?n.code:13;return{status:e,value:{message:n.message,error:te(e)}}}var xe="$wdc_",Ut=":wdc:";function be(n){let e=n||document,t=e[xe];return t||(t=e[xe]={},t.nextId=Date.now()),t.nextId||(t.nextId=Date.now()),t}var Kt=200;function Fe(n,e){if("setInterval"in e)return!!e.closed;let t=e;for(;t;){if(t===n.documentElement)return!1;t.host&&t.nodeType===Node.DOCUMENT_FRAGMENT_NODE&&(t=t.host),t=t.parentNode}return!0}function kt(n){let e=be(n);for(let t of Object.keys(e)){let o=e[t];!o||typeof o!="object"||Fe(n,o)&&delete e[t]}}function Ue(n){let e=n.ownerDocument,t=be(e);Object.keys(t).length-1>=Kt&&kt(e||document);let o;for(let s of Object.keys(t))if(t[s]===n){o=s;break}return o||(o=Ut+t.nextId++,t[o]=n),o}function j(n,e){n=decodeURIComponent(n);let t=e||document,o=be(t);if(!(n in o))throw new p(10,"Element does not exist in cache");let s=o[n];if(!Fe(t,s))return s;throw delete o[n],"setInterval"in s?new p(23,"Window has been closed."):new p(10,"Element is no longer attached to the DOM")}var u={NULL:"\uE000",CANCEL:"\uE001",HELP:"\uE002",BACK_SPACE:"\uE003",TAB:"\uE004",CLEAR:"\uE005",RETURN:"\uE006",ENTER:"\uE007",SHIFT:"\uE008",CONTROL:"\uE009",ALT:"\uE00A",PAUSE:"\uE00B",ESCAPE:"\uE00C",SPACE:"\uE00D",PAGE_UP:"\uE00E",PAGE_DOWN:"\uE00F",END:"\uE010",HOME:"\uE011",ARROW_LEFT:"\uE012",LEFT:"\uE012",ARROW_UP:"\uE013",UP:"\uE013",ARROW_RIGHT:"\uE014",RIGHT:"\uE014",ARROW_DOWN:"\uE015",DOWN:"\uE015",INSERT:"\uE016",DELETE:"\uE017",SEMICOLON:"\uE018",EQUALS:"\uE019",NUMPAD0:"\uE01A",NUMPAD1:"\uE01B",NUMPAD2:"\uE01C",NUMPAD3:"\uE01D",NUMPAD4:"\uE01E",NUMPAD5:"\uE01F",NUMPAD6:"\uE020",NUMPAD7:"\uE021",NUMPAD8:"\uE022",NUMPAD9:"\uE023",MULTIPLY:"\uE024",ADD:"\uE025",SEPARATOR:"\uE026",SUBTRACT:"\uE027",DECIMAL:"\uE028",DIVIDE:"\uE029",F1:"\uE031",F2:"\uE032",F3:"\uE033",F4:"\uE034",F5:"\uE035",F6:"\uE036",F7:"\uE037",F8:"\uE038",F9:"\uE039",F10:"\uE03A",F11:"\uE03B",F12:"\uE03C",COMMAND:"\uE03D",META:"\uE03D"};var io=new Map([[u.NULL,null],[u.BACK_SPACE,l.BACKSPACE],[u.TAB,l.TAB],[u.RETURN,l.ENTER],[u.ENTER,l.ENTER],[u.SHIFT,l.SHIFT],[u.CONTROL,l.CONTROL],[u.ALT,l.ALT],[u.PAUSE,l.PAUSE],[u.ESCAPE,l.ESC],[u.SPACE,l.SPACE],[u.PAGE_UP,l.PAGE_UP],[u.PAGE_DOWN,l.PAGE_DOWN],[u.END,l.END],[u.HOME,l.HOME],[u.LEFT,l.LEFT],[u.UP,l.UP],[u.RIGHT,l.RIGHT],[u.DOWN,l.DOWN],[u.INSERT,l.INSERT],[u.DELETE,l.DELETE],[u.SEMICOLON,l.SEMICOLON],[u.EQUALS,l.EQUALS],[u.NUMPAD0,l.NUM_ZERO],[u.NUMPAD1,l.NUM_ONE],[u.NUMPAD2,l.NUM_TWO],[u.NUMPAD3,l.NUM_THREE],[u.NUMPAD4,l.NUM_FOUR],[u.NUMPAD5,l.NUM_FIVE],[u.NUMPAD6,l.NUM_SIX],[u.NUMPAD7,l.NUM_SEVEN],[u.NUMPAD8,l.NUM_EIGHT],[u.NUMPAD9,l.NUM_NINE],[u.MULTIPLY,l.NUM_MULTIPLY],[u.ADD,l.NUM_PLUS],[u.SUBTRACT,l.NUM_MINUS],[u.DECIMAL,l.NUM_PERIOD],[u.DIVIDE,l.NUM_DIVISION],[u.SEPARATOR,l.SEPARATOR],[u.F1,l.F1],[u.F2,l.F2],[u.F3,l.F3],[u.F4,l.F4],[u.F5,l.F5],[u.F6,l.F6],[u.F7,l.F7],[u.F8,l.F8],[u.F9,l.F9],[u.F10,l.F10],[u.F11,l.F11],[u.F12,l.F12],[u.META,l.META]]);function Ve(n){return n?j(n.WINDOW):window}function Bt(n,e,t){let o;try{let s=Ve(t),i=Q(e,s.document),a=n.apply(null,i);o=Ke(a)}catch(s){o=ke(s)}return JSON.stringify(o)}function We(n,e){return Bt(Ee,[n],e)}return qe(Vt);})();
+function(){"use strict";var AtomExport=(()=>{var ee=Object.defineProperty;var Ge=Object.getOwnPropertyDescriptor;var je=Object.getOwnPropertyNames;var $e=Object.prototype.hasOwnProperty;var ze=(n,e)=>{for(var t in e)ee(n,t,{get:e[t],enumerable:!0})},Je=(n,e,t,o)=>{if(e&&typeof e=="object"||typeof e=="function")for(let s of je(e))!$e.call(n,s)&&s!==t&&ee(n,s,{get:()=>e[s],enumerable:!(o=Ge(e,s))||o.enumerable});return n};var qe=n=>Je(ee({},"__esModule",{value:!0}),n);var Vt={};ze(Vt,{default:()=>We});var Ze={15:"element not selectable",11:"element not visible",31:"unknown error",30:"unknown error",24:"invalid cookie domain",29:"invalid element coordinates",12:"invalid element state",32:"invalid selector",51:"invalid selector",52:"invalid selector",17:"javascript error",405:"unsupported operation",34:"move target out of bounds",27:"no such alert",7:"no such element",8:"no such frame",23:"no such window",28:"script timeout",33:"session not created",10:"stale element reference",21:"timeout",25:"unable to set cookie",26:"unexpected alert open",13:"unknown error",9:"unknown command"};function te(n){return Ze[n]??"unknown error"}var p=class extends Error{code;state;isAutomationError=!0;constructor(e,t=""){super(t),this.code=e,this.state=te(e);let o=this.state.replace(/(^|\s+)[a-z]/g,i=>i.toUpperCase().trimStart()).replace(/\s+/g,"");this.name=/Error$/.test(o)?o:`${o}Error`;let s=new Error(this.message);s.name=this.name,this.stack=s.stack||""}};var Qe=/[;]+(?=(?:(?:[^"]*"){2})*[^"]*$)(?=(?:(?:[^']*'){2})*[^']*$)(?=(?:[^()]*\([^()]*\))*[^()]*$)/;function et(n){let e=n.split(Qe),t=[];for(let s of e){let i=s.indexOf(":");if(i>0){let a=[s.slice(0,i),s.slice(i+1)];t.push(a[0].toLowerCase(),":",a[1],";")}}let o=t.join("");return o.charAt(o.length-1)===";"?o:`${o};`}function we(n,e){if(e=e.toLowerCase(),e==="style")return et(n.style.cssText);let t=n.getAttributeNode(e);return t&&t.specified?t.value:null}function x(n,e){return n[e]}function c(n,e){return n instanceof HTMLFormElement?n.nodeType===Node.ELEMENT_NODE&&(!e||e==="FORM"):!!n&&n.nodeType===Node.ELEMENT_NODE&&(!e||n.tagName.toUpperCase()===e)}function K(n){return n.nodeType===Node.DOCUMENT_NODE?n:n.ownerDocument}function ne(n,e,t=!1){let o=t?n:n.parentNode;for(;o;){if(e(o))return o;o=o.parentNode}return null}function k(n){if(c(n,"OPTION"))return!0;if(c(n,"INPUT")){let e=n.type.toLowerCase();return e==="checkbox"||e==="radio"}return!1}function F(n){if(!k(n))throw new p(15,"Element is not selectable");let e=n.type?.toLowerCase();return!!x(n,e==="checkbox"||e==="radio"?"checked":"selected")}var ve={aliceblue:"#f0f8ff",antiquewhite:"#faebd7",aqua:"#00ffff",aquamarine:"#7fffd4",azure:"#f0ffff",beige:"#f5f5dc",bisque:"#ffe4c4",black:"#000000",blanchedalmond:"#ffebcd",blue:"#0000ff",blueviolet:"#8a2be2",brown:"#a52a2a",burlywood:"#deb887",cadetblue:"#5f9ea0",chartreuse:"#7fff00",chocolate:"#d2691e",coral:"#ff7f50",cornflowerblue:"#6495ed",cornsilk:"#fff8dc",crimson:"#dc143c",cyan:"#00ffff",darkblue:"#00008b",darkcyan:"#008b8b",darkgoldenrod:"#b8860b",darkgray:"#a9a9a9",darkgreen:"#006400",darkgrey:"#a9a9a9",darkkhaki:"#bdb76b",darkmagenta:"#8b008b",darkolivegreen:"#556b2f",darkorange:"#ff8c00",darkorchid:"#9932cc",darkred:"#8b0000",darksalmon:"#e9967a",darkseagreen:"#8fbc8f",darkslateblue:"#483d8b",darkslategray:"#2f4f4f",darkslategrey:"#2f4f4f",darkturquoise:"#00ced1",darkviolet:"#9400d3",deeppink:"#ff1493",deepskyblue:"#00bfff",dimgray:"#696969",dimgrey:"#696969",dodgerblue:"#1e90ff",firebrick:"#b22222",floralwhite:"#fffaf0",forestgreen:"#228b22",fuchsia:"#ff00ff",gainsboro:"#dcdcdc",ghostwhite:"#f8f8ff",gold:"#ffd700",goldenrod:"#daa520",gray:"#808080",green:"#008000",greenyellow:"#adff2f",grey:"#808080",honeydew:"#f0fff0",hotpink:"#ff69b4",indianred:"#cd5c5c",indigo:"#4b0082",ivory:"#fffff0",khaki:"#f0e68c",lavender:"#e6e6fa",lavenderblush:"#fff0f5",lawngreen:"#7cfc00",lemonchiffon:"#fffacd",lightblue:"#add8e6",lightcoral:"#f08080",lightcyan:"#e0ffff",lightgoldenrodyellow:"#fafad2",lightgray:"#d3d3d3",lightgreen:"#90ee90",lightgrey:"#d3d3d3",lightpink:"#ffb6c1",lightsalmon:"#ffa07a",lightseagreen:"#20b2aa",lightskyblue:"#87cefa",lightslategray:"#778899",lightslategrey:"#778899",lightsteelblue:"#b0c4de",lightyellow:"#ffffe0",lime:"#00ff00",limegreen:"#32cd32",linen:"#faf0e6",magenta:"#ff00ff",maroon:"#800000",mediumaquamarine:"#66cdaa",mediumblue:"#0000cd",mediumorchid:"#ba55d3",mediumpurple:"#9370db",mediumseagreen:"#3cb371",mediumslateblue:"#7b68ee",mediumspringgreen:"#00fa9a",mediumturquoise:"#48d1cc",mediumvioletred:"#c71585",midnightblue:"#191970",mintcream:"#f5fffa",mistyrose:"#ffe4e1",moccasin:"#ffe4b5",navajowhite:"#ffdead",navy:"#000080",oldlace:"#fdf5e6",olive:"#808000",olivedrab:"#6b8e23",orange:"#ffa500",orangered:"#ff4500",orchid:"#da70d6",palegoldenrod:"#eee8aa",palegreen:"#98fb98",paleturquoise:"#afeeee",palevioletred:"#db7093",papayawhip:"#ffefd5",peachpuff:"#ffdab9",peru:"#cd853f",pink:"#ffc0cb",plum:"#dda0dd",powderblue:"#b0e0e6",purple:"#800080",red:"#ff0000",rosybrown:"#bc8f8f",royalblue:"#4169e1",saddlebrown:"#8b4513",salmon:"#fa8072",sandybrown:"#f4a460",seagreen:"#2e8b57",seashell:"#fff5ee",sienna:"#a0522d",silver:"#c0c0c0",skyblue:"#87ceeb",slateblue:"#6a5acd",slategray:"#708090",slategrey:"#708090",snow:"#fffafa",springgreen:"#00ff7f",steelblue:"#4682b4",tan:"#d2b48c",teal:"#008080",thistle:"#d8bfd8",tomato:"#ff6347",turquoise:"#40e0d0",violet:"#ee82ee",wheat:"#f5deb3",white:"#ffffff",whitesmoke:"#f5f5f5",yellow:"#ffff00",yellowgreen:"#9acd32"};var tt=new Set(["backgroundColor","borderTopColor","borderRightColor","borderBottomColor","borderLeftColor","color","outlineColor"]),nt=/#([0-9a-fA-F])([0-9a-fA-F])([0-9a-fA-F])/,ot=/^#(?:[0-9a-f]{3}){1,2}$/i,st=/^(?:rgba)?\((\d{1,3}),\s?(\d{1,3}),\s?(\d{1,3}),\s?(0|1|0\.\d*)\)$/i,rt=/^(?:rgb)?\((0|[1-9]\d{0,2}),\s?(0|[1-9]\d{0,2}),\s?(0|[1-9]\d{0,2})\)$/i;function it(n){n=n.toLowerCase();let e=ve[n];if(!e&&(e=n.charAt(0)==="#"?n:`#${n}`,e.length===4&&(e=e.replace(nt,"#$1$1$2$2$3$3")),!ot.test(e)))return null;let t=parseInt(e.substring(1,3),16),o=parseInt(e.substring(3,5),16),s=parseInt(e.substring(5,7),16);return[t,o,s,1]}function lt(n){let e=n.match(st);if(e){let t=Number(e[1]),o=Number(e[2]),s=Number(e[3]),i=Number(e[4]);if(t>=0&&t<=255&&o>=0&&o<=255&&s>=0&&s<=255&&i>=0&&i<=1)return[t,o,s,i]}return null}function ut(n){let e=n.match(rt);if(e){let t=Number(e[1]),o=Number(e[2]),s=Number(e[3]);if(t>=0&&t<=255&&o>=0&&o<=255&&s>=0&&s<=255)return[t,o,s,1]}return null}function Se(n,e){if(!tt.has(n))return e;let t=lt(e)||ut(e)||it(e);return t?`rgba(${t.join(", ")})`:e}function Re(n,e){if(!n)throw new p(32,"No selector specified");n=n.trim();let t;try{t=e.querySelector(n)}catch{throw new p(32,"An invalid or illegal selector was specified")}return t&&t.nodeType===Node.ELEMENT_NODE?t:null}var w=class{x;y;constructor(e=0,t=0){this.x=e,this.y=t}},H=class n{x;y;constructor(e,t){this.x=e,this.y=t}magnitude(){return Math.hypot(this.x,this.y)}scale(e,t=e){return this.x*=e,this.y*=t,this}subtract(e){return this.x-=e.x,this.y-=e.y,this}rotate(e){let t=Math.cos(e),o=Math.sin(e),s=this.x*t-this.y*o,i=this.y*t+this.x*o;return this.x=s,this.y=i,this}static fromCoordinate(e){return new n(e.x,e.y)}static sum(e,t){return new n(e.x+t.x,e.y+t.y)}static difference(e,t){return new n(e.x-t.x,e.y-t.y)}},B=class{width;height;constructor(e,t){this.width=e,this.height=t}},V=class{top;right;bottom;left;constructor(e,t,o,s){this.top=e,this.right=t,this.bottom=o,this.left=s}},g=class n{left;top;width;height;constructor(e,t,o,s){this.left=e,this.top=t,this.width=o,this.height=s}toBox(){let e=this.left+this.width,t=this.top+this.height;return new V(this.top,e,t,this.left)}intersects(e){return n.intersects(this,e)}static intersects(e,t){return e.left<=t.left+t.width&&t.left<=e.left+e.width&&e.top<=t.top+t.height&&t.top<=e.top+e.height}};function W(n,e,t){return Math.min(Math.max(n,e),t)}var ct=typeof ShadowRoot=="function";function Et(n){return n.replace(/-([a-z])/g,(e,t)=>t.toUpperCase())}function $(n){let e="document"in n?n.document:K(n);try{let t=e?.activeElement;return t&&t.nodeName?t:null}catch{return null}}function R(n){return z(n,!0)&&Ie(n)&&!ft(n)}function ft(n){return N(n,"pointer-events")==="none"}var dt=["A","AREA","BUTTON","INPUT","LABEL","SELECT","TEXTAREA"];function Ce(n){return dt.some(e=>c(n,e))||we(n,"tabindex")!=null&&Number(x(n,"tabIndex"))>=0||se(n)}var mt=["BUTTON","INPUT","OPTGROUP","OPTION","SELECT","TEXTAREA"];function Ie(n){if(!mt.some(o=>c(n,o)))return!0;if(x(n,"disabled"))return!1;if(n.parentNode&&n.parentNode.nodeType===Node.ELEMENT_NODE&&c(n,"OPTGROUP")||c(n,"OPTION"))return Ie(n.parentNode);let t=n;for(;t;){let o=t.parentNode;if(o&&c(o,"FIELDSET")&&x(o,"disabled")){if(!c(t,"LEGEND"))return!1;let s=t.previousElementSibling;for(;s;){if(c(s,"LEGEND"))return!1;s=s.previousElementSibling}}t=o}return!0}var ht=new Set(["text","search","tel","url","email","password","number"]);function pt(n){if(c(n,"TEXTAREA"))return!0;if(c(n,"INPUT")){let e=n.type.toLowerCase();return ht.has(e)}return Me(n)}function Tt(n){return c(n,"INPUT")&&n.type.toLowerCase()==="file"}function A(n,e){return c(n,"INPUT")&&n.type.toLowerCase()===e}function Me(n){let e=n;if(e.contentEditable===void 0)return!1;if(e.isContentEditable!==void 0)return e.isContentEditable;function t(o){if(o.contentEditable==="inherit"){let s=I(o);return s?t(s):!1}return o.contentEditable==="true"}return t(e)}function se(n){return(pt(n)||Tt(n)||A(n,"range")||A(n,"date")||A(n,"month")||A(n,"week")||A(n,"time")||A(n,"datetime-local")||A(n,"color"))&&!x(n,"readOnly")}function I(n){let e=n.parentNode;for(;e&&e.nodeType!==Node.ELEMENT_NODE&&e.nodeType!==Node.DOCUMENT_NODE&&e.nodeType!==Node.DOCUMENT_FRAGMENT_NODE;)e=e.parentNode;return e&&c(e)?e:null}function bt(n,e){let t=n.ownerDocument.defaultView;if(t?.getComputedStyle){let o=t.getComputedStyle(n,null);if(o)return o[e]||o.getPropertyValue(e)||""}return""}function _e(n,e){let t=n.style,o=t[e];if(o===void 0&&typeof t.getPropertyValue=="function"&&(o=t.getPropertyValue(e)),o!=="inherit")return o!==void 0?o:null;let s=I(n);return s?_e(s,e):null}function N(n,e){let t=Et(e);(t==="float"||t==="cssFloat"||t==="styleFloat")&&(t="cssFloat");let o=bt(n,t)||_e(n,t);return o===null?null:Se(t,o)}function oe(n,e,t){if(!c(n))throw new Error("Argument to isShown must be of type Element");if(c(n,"BODY"))return!0;if(c(n,"OPTION")||c(n,"OPTGROUP")){let E=n.parentNode;for(;E&&!c(E,"SELECT");)E=E.parentNode;return!!E&&oe(E,!0,t)}let o=Ae(n);if(o)return!!o.image&&o.rect.width>0&&o.rect.height>0&&oe(o.image,e,t);if(c(n,"INPUT")&&n.type.toLowerCase()==="hidden"||c(n,"NOSCRIPT"))return!1;let s=N(n,"visibility");if(s==="collapse"||s==="hidden"||!t(n)||!e&&Le(n)===0)return!1;function i(E){let d=C(E);if(d.height>0&&d.width>0)return!0;if(c(E,"PATH")&&(d.height>0||d.width>0)){let T=N(E,"stroke-width");return!!T&&parseInt(T,10)>0}let y=N(E,"visibility");return y==="collapse"||y==="hidden"||!t(E)?!1:N(E,"overflow")!=="hidden"&&[...E.childNodes].some(T=>{if(T.nodeType===Node.TEXT_NODE){let v=T.nodeValue??"";return!(/^[\s]*$/.test(v)&&/[\n\r\t]/.test(v))}return c(T)&&i(T)})}if(!i(n))return!1;function a(E){return L(E)==="hidden"&&[...E.childNodes].every(d=>!c(d)||a(d)||!i(d))}return!a(n)}function z(n,e=!1){function t(o){if(c(o)){let i=o;if(N(i,"display")==="none"||N(i,"content-visibility")==="hidden")return!1}let s=gt(o);if(ct&&s instanceof ShadowRoot){if(s.host.shadowRoot&&s.host.shadowRoot!==s)return!1;s=s.host}return s&&(s.nodeType===Node.DOCUMENT_NODE||s.nodeType===Node.DOCUMENT_FRAGMENT_NODE)?!0:s&&c(s,"DETAILS")&&!s.open&&!c(o,"SUMMARY")?!1:!!s&&t(s)}return oe(n,e,t)}function J(n){let e=n.scrollingElement||n.documentElement,t=n.defaultView;return new w(t.pageXOffset||e.scrollLeft,t.pageYOffset||e.scrollTop)}function L(n,e){let t=q(n,e),o=K(n),s=o.documentElement,i=o.body,a=N(s,"overflow"),E=!1;function d(m,b){if(m===s)return!0;let h=N(m,"display");return!(h?.startsWith("inline")||h==="contents"||b==="absolute"&&N(m,"position")==="static")}function y(m){let b=N(m,"position");if(b==="fixed")return E=!0,m===s?null:s;let h=I(m);for(;h&&!d(h,b);)h=I(h);return h}function T(m){let b=m;if(a==="visible"){if(m===s&&i)b=i;else if(m===i)return{x:"visible",y:"visible"}}let h={x:N(b,"overflow-x")??"",y:N(b,"overflow-y")??""};return m===s&&(h.x=h.x==="visible"?"auto":h.x,h.y=h.y==="visible"?"auto":h.y),h}function v(m){return m===s?J(o):new w(m.scrollLeft,m.scrollTop)}for(let m=y(n);m;m=y(m)){let b=T(m);if(b.x==="visible"&&b.y==="visible")continue;let h=C(m);if(h.width===0||h.height===0)return"hidden";let Ne=t.right<h.left,ge=t.bottom<h.top;if(Ne&&b.x==="hidden"||ge&&b.y==="hidden")return"hidden";if(Ne&&b.x!=="visible"||ge&&b.y!=="visible"){let U=v(m),Xe=t.right<h.left-U.x,Ye=t.bottom<h.top-U.y;return Xe&&b.x!=="visible"||Ye&&b.x!=="visible"||L(m)==="hidden"?"hidden":"scroll"}let Oe=t.left>=h.left+h.width,ye=t.top>=h.top+h.height;if(Oe&&b.x==="hidden"||ye&&b.y==="hidden")return"hidden";if(Oe&&b.x!=="visible"||ye&&b.y!=="visible"){if(E){let U=v(m);if(t.left>=s.scrollWidth-U.x||t.right>=s.scrollHeight-U.y)return"hidden"}return L(m)==="hidden"?"hidden":"scroll"}}return"none"}function C(n){let e=Ae(n);if(e)return e.rect;if(c(n,"HTML")){let o=K(n),s=o.defaultView,i=o.compatMode==="CSS1Compat"?o.documentElement:o.body;return new g(0,0,i.clientWidth||s.innerWidth,i.clientHeight||s.innerHeight)}let t;try{t=n.getBoundingClientRect()}catch{return new g(0,0,0,0)}return new g(t.left,t.top,t.right-t.left,t.bottom-t.top)}function Ae(n){let e=c(n,"MAP");if(!e&&!c(n,"AREA"))return null;let t=e?n:n.parentNode&&c(n.parentNode,"MAP")?n.parentNode:null,o=null,s=null,i=t&&t.name;if(t&&i){let a=K(t),E=`*[usemap="#${i}"]`;if(o=Re(E,a),o&&(s=C(o),!e&&n.shape.toLowerCase()!=="default")){let d=Nt(n),y=Math.min(Math.max(d.left,0),s.width),T=Math.min(Math.max(d.top,0),s.height),v=Math.min(d.width,s.width-y),m=Math.min(d.height,s.height-T);s=new g(y+s.left,T+s.top,v,m)}}return{image:o,rect:s||new g(0,0,0,0)}}function Nt(n){let e=n,t=e.shape.toLowerCase(),o=e.coords.split(",");if(t==="rect"&&o.length===4){let s=Number(o[0]),i=Number(o[1]);return new g(s,i,Number(o[2])-s,Number(o[3])-i)}else if(t==="circle"&&o.length===3){let s=Number(o[0]),i=Number(o[1]),a=Number(o[2]);return new g(s-a,i-a,2*a,2*a)}else if(t==="poly"&&o.length>2){let s=Number(o[0]),i=Number(o[1]),a=s,E=i;for(let d=2;d+1<o.length;d+=2)s=Math.min(s,Number(o[d])),a=Math.max(a,Number(o[d])),i=Math.min(i,Number(o[d+1])),E=Math.max(E,Number(o[d+1]));return new g(s,i,a-s,E-i)}return new g(0,0,0,0)}function q(n,e){let t=C(n).toBox();if(e){let o=e instanceof g?e:new g(e.x,e.y,1,1);t.left=W(t.left+o.left,t.left,t.right),t.top=W(t.top+o.top,t.top,t.bottom),t.right=W(t.left+o.width,t.left,t.right),t.bottom=W(t.top+o.height,t.top,t.bottom)}return t}function Le(n){let e=1,t=N(n,"opacity");t&&(e=Number(t));let o=I(n);return o&&(e*=Le(o)),e}function gt(n){let e=n.parentNode;if(e&&e.shadowRoot&&n.assignedSlot!==void 0)return n.assignedSlot?n.assignedSlot.parentNode:null;let t=n;if(t.getDestinationInsertionPoints){let o=t.getDestinationInsertionPoints();if(o.length>0)return o[o.length-1]}return e}var O=class{type;bubbles;cancelable;constructor(e,t,o){this.type=e,this.bubbles=t,this.cancelable=o}create(e,t){let s=("ownerDocument"in e?e.ownerDocument:e.document).createEvent("HTMLEvents");return s.initEvent(this.type,this.bubbles,this.cancelable),s}toString(){return this.type}},S=class extends O{create(e,t){let o=t,s="ownerDocument"in e?e.ownerDocument:e.document,i=s.defaultView,a=s.createEvent("MouseEvents"),E=1;return this.type==="mousewheel"&&(a.wheelDelta=o.wheelDelta),a.initMouseEvent(this.type,this.bubbles,this.cancelable,i,E,o.clientX,o.clientY,o.clientX,o.clientY,o.ctrlKey,o.altKey,o.shiftKey,o.metaKey,o.button,o.relatedTarget),a}},X=class extends O{create(e,t){let o=t,i=("ownerDocument"in e?e.ownerDocument:e.document).createEvent("Events");return i.initEvent(this.type,this.bubbles,this.cancelable),i.altKey=o.altKey,i.ctrlKey=o.ctrlKey,i.metaKey=o.metaKey,i.shiftKey=o.shiftKey,i.keyCode=o.charCode||o.keyCode,i.charCode=this.type==="keypress"?i.keyCode:0,i}};function re(n,e){return e.map(t=>new Touch({identifier:t.identifier,target:n,screenX:t.screenX,screenY:t.screenY,clientX:t.clientX,clientY:t.clientY,pageX:t.pageX,pageY:t.pageY}))}var Y=class extends O{create(e,t){let o=t,s="ownerDocument"in e?e.ownerDocument:e.document,i=s.defaultView,a=typeof TouchEvent.prototype.initTouchEvent=="function"?"INIT_TOUCH_EVENT":"TOUCH_EVENT_CTOR",E=re(e,o.changedTouches),d=o.touches===o.changedTouches?E:re(e,o.touches),y=o.targetTouches===o.changedTouches?E:re(e,o.targetTouches);if(a==="INIT_TOUCH_EVENT"){let T=s.createEvent("TouchEvent");return T.initTouchEvent(this.type,this.bubbles,this.cancelable,i,1,0,0,o.clientX,o.clientY,o.ctrlKey,o.altKey,o.shiftKey,o.metaKey,d,y,E,o.scale,o.rotation),T.relatedTarget=o.relatedTarget,T}return new TouchEvent(this.type,{touches:d,targetTouches:y,changedTouches:E,bubbles:this.bubbles,cancelable:this.cancelable,ctrlKey:o.ctrlKey,shiftKey:o.shiftKey,altKey:o.altKey,metaKey:o.metaKey})}},f={BLUR:new O("blur",!1,!1),CHANGE:new O("change",!0,!1),FOCUS:new O("focus",!1,!1),FOCUSIN:new O("focusin",!0,!1),FOCUSOUT:new O("focusout",!0,!1),INPUT:new O("input",!0,!1),ORIENTATIONCHANGE:new O("orientationchange",!1,!1),SELECT:new O("select",!0,!1),SUBMIT:new O("submit",!0,!0),TEXTINPUT:new O("textInput",!0,!0),CLICK:new S("click",!0,!0),CONTEXTMENU:new S("contextmenu",!0,!0),DBLCLICK:new S("dblclick",!0,!0),MOUSEDOWN:new S("mousedown",!0,!0),MOUSEMOVE:new S("mousemove",!0,!1),MOUSEOUT:new S("mouseout",!0,!0),MOUSEOVER:new S("mouseover",!0,!0),MOUSEUP:new S("mouseup",!0,!0),MOUSEWHEEL:new S("mousewheel",!0,!0),KEYDOWN:new X("keydown",!0,!0),KEYPRESS:new X("keypress",!0,!0),KEYUP:new X("keyup",!0,!0),TOUCHEND:new Y("touchend",!0,!0),TOUCHMOVE:new Y("touchmove",!0,!0),TOUCHSTART:new Y("touchstart",!0,!0)};function M(n,e,t){let o=e.create(n,t);return"isTrusted"in o||(o.isTrusted=!1),n.dispatchEvent(o)}var ie=class{pressedModifiers=0;isPressed(e){return(this.pressedModifiers&e)!==0}setPressed(e,t){t?this.pressedModifiers|=e:this.pressedModifiers&=~e}isShiftPressed(){return this.isPressed(1)}isControlPressed(){return this.isPressed(2)}isAltPressed(){return this.isPressed(4)}isMetaPressed(){return this.isPressed(8)}},le=class{fireHtmlEvent(e,t){return M(e,t)}fireKeyboardEvent(e,t,o){return M(e,t,o)}fireMouseEvent(e,t,o){return M(e,t,o)}fireTouchEvent(e,t,o){return M(e,t,o)}};function Ot(n){return c(n,"FORM")}var D=class{el;select=null;modifiersState;eventEmitter;constructor(e,t){this.el=document.documentElement;let o=$(this.el);o&&this.setElement(o),this.modifiersState=e||new ie,this.eventEmitter=t||new le}getElement(){return this.el}setElement(e){this.el=e,c(e,"OPTION")?this.select=ne(e,t=>c(t,"SELECT")):this.select=null}fireHtmlEvent(e){return this.eventEmitter.fireHtmlEvent(this.el,e)}fireKeyboardEvent(e,t){return this.eventEmitter.fireKeyboardEvent(this.el,e,t)}fireMouseEvent(e,t,o,s=null,i=0,a=!1){if(!a&&!R(this.el))return!1;if(s&&!(e===f.MOUSEOVER||e===f.MOUSEOUT))throw new p(12,`Event type does not allow related target: ${e}`);let E={clientX:t.x,clientY:t.y,button:o,altKey:this.modifiersState.isAltPressed(),ctrlKey:this.modifiersState.isControlPressed(),shiftKey:this.modifiersState.isShiftPressed(),metaKey:this.modifiersState.isMetaPressed(),wheelDelta:i,relatedTarget:s},d=this.select?this.getTargetOfOptionMouseEvent(e):this.el;return d?this.eventEmitter.fireMouseEvent(d,e,E):!0}fireTouchEvent(e,t,o,s,i){let a={touches:[],targetTouches:[],changedTouches:[],altKey:this.modifiersState.isAltPressed(),ctrlKey:this.modifiersState.isControlPressed(),shiftKey:this.modifiersState.isShiftPressed(),metaKey:this.modifiersState.isMetaPressed(),relatedTarget:null,scale:0,rotation:0,clientX:o.x,clientY:o.y},E=J(this.el.ownerDocument),d=(y,T)=>{let v={identifier:y,screenX:T.x,screenY:T.y,clientX:T.x,clientY:T.y,pageX:T.x+E.x,pageY:T.y+E.y};a.changedTouches.push(v),(e===f.TOUCHSTART||e===f.TOUCHMOVE)&&(a.touches.push(v),a.targetTouches.push(v))};return d(t,o),s!==void 0&&i!==void 0&&d(s,i),this.eventEmitter.fireTouchEvent(this.el,e,a)}getTargetOfOptionMouseEvent(e){let t=this.select;switch(e){case f.CLICK:case f.MOUSEUP:return t.multiple?this.el:this.select;default:return t.multiple?this.el:null}}clickElement(e,t,o=!1){if(!o&&!R(this.el))return;let s=!this.select&&k(this.el),i=s&&F(this.el);this.fireMouseEvent(f.CLICK,e,t,null,0,o)&&s&&this.toggleRadioButtonOrCheckbox(i)}maybeToggleOption(){if(!this.select||!R(this.el))return;let e=this.select,t=F(this.el);t&&!e.multiple||(this.el.selected=!t,e.multiple||M(e,f.CHANGE))}toggleRadioButtonOrCheckbox(e){}focusOnElement(){let e=ne(this.el,o=>!!o&&c(o)&&Ce(o),!0);e=e||this.el;let t=$(e);return e===t?!1:(t&&typeof t.blur=="function"&&(c(t,"BODY")||t.blur()),typeof e.focus=="function"?(e.focus(),!0):!1)}submitForm(e){if(!Ot(e))throw new p(12,"Element is not a form, so could not submit.");M(e,f.SUBMIT)&&(c(e.submit)?HTMLFormElement.prototype.submit.call(e):e.submit())}};function yt(n){return navigator.userAgent.includes(n)}function ue(){return yt("Macintosh")}var ce=class{code;character;shiftChar;constructor(e,t,o){this.code=e,this.character=t||null,this.shiftChar=o||this.character}},ae=new Map;function r(n,e,t){let o=new ce(n,e,t);return e&&(!ae.has(e)||t)&&(ae.set(e,{key:o,shift:!1}),t&&ae.set(t,{key:o,shift:!0})),o}var l={BACKSPACE:r(8),TAB:r(9),ENTER:r(13),SHIFT:r(16),CONTROL:r(17),ALT:r(18),PAUSE:r(19),CAPS_LOCK:r(20),ESC:r(27),SPACE:r(32," "),PAGE_UP:r(33),PAGE_DOWN:r(34),END:r(35),HOME:r(36),LEFT:r(37),UP:r(38),RIGHT:r(39),DOWN:r(40),PRINT_SCREEN:r(44),INSERT:r(45),DELETE:r(46),ZERO:r(48,"0",")"),ONE:r(49,"1","!"),TWO:r(50,"2","@"),THREE:r(51,"3","#"),FOUR:r(52,"4","$"),FIVE:r(53,"5","%"),SIX:r(54,"6","^"),SEVEN:r(55,"7","&"),EIGHT:r(56,"8","*"),NINE:r(57,"9","("),A:r(65,"a","A"),B:r(66,"b","B"),C:r(67,"c","C"),D:r(68,"d","D"),E:r(69,"e","E"),F:r(70,"f","F"),G:r(71,"g","G"),H:r(72,"h","H"),I:r(73,"i","I"),J:r(74,"j","J"),K:r(75,"k","K"),L:r(76,"l","L"),M:r(77,"m","M"),N:r(78,"n","N"),O:r(79,"o","O"),P:r(80,"p","P"),Q:r(81,"q","Q"),R:r(82,"r","R"),S:r(83,"s","S"),T:r(84,"t","T"),U:r(85,"u","U"),V:r(86,"v","V"),W:r(87,"w","W"),X:r(88,"x","X"),Y:r(89,"y","Y"),Z:r(90,"z","Z"),META:r(91),META_RIGHT:r(ue()?93:92),CONTEXT_MENU:r(ue()?0:null),NUM_ZERO:r(96,"0"),NUM_ONE:r(97,"1"),NUM_TWO:r(98,"2"),NUM_THREE:r(99,"3"),NUM_FOUR:r(100,"4"),NUM_FIVE:r(101,"5"),NUM_SIX:r(102,"6"),NUM_SEVEN:r(103,"7"),NUM_EIGHT:r(104,"8"),NUM_NINE:r(105,"9"),NUM_MULTIPLY:r(106,"*"),NUM_PLUS:r(107,"+"),NUM_MINUS:r(109,"-"),NUM_PERIOD:r(110,"."),NUM_DIVISION:r(111,"/"),NUM_LOCK:r(144),F1:r(112),F2:r(113),F3:r(114),F4:r(115),F5:r(116),F6:r(117),F7:r(118),F8:r(119),F9:r(120),F10:r(121),F11:r(122),F12:r(123),EQUALS:r(187,"=","+"),SEPARATOR:r(108,","),HYPHEN:r(189,"-","_"),COMMA:r(188,",","<"),PERIOD:r(190,".",">"),SLASH:r(191,"/","?"),BACKTICK:r(192,"`","~"),OPEN_BRACKET:r(219,"[","{"),BACKSLASH:r(220,"\\","|"),CLOSE_BRACKET:r(221,"]","}"),SEMICOLON:r(186,";",":"),APOSTROPHE:r(222,"'",'"')};var St=[l.ALT,l.CONTROL,l.META,l.SHIFT],vn=new Map([[l.SHIFT.code,1],[l.CONTROL.code,2],[l.ALT.code,4],[l.META.code,8]]);var Rt=3,P=new Map([[f.CLICK,[0,1,2,null]],[f.CONTEXTMENU,[null,null,2,null]],[f.MOUSEUP,[0,1,2,null]],[f.MOUSEOUT,[0,1,2,0]],[f.MOUSEMOVE,[0,1,2,0]]]);P.set(f.DBLCLICK,P.get(f.CLICK));P.set(f.MOUSEDOWN,P.get(f.MOUSEUP));P.set(f.MOUSEOVER,P.get(f.MOUSEOUT));var G=class extends D{buttonPressed=null;elementPressed=null;clientXY=new w(0,0);nextClickIsDoubleClick=!1;hasEverInteracted=!1;constructor(e,t,o){super(t,o),e&&(typeof e.buttonPressed=="number"&&(this.buttonPressed=e.buttonPressed),e.elementPressed&&c(e.elementPressed)&&(this.elementPressed=e.elementPressed),this.clientXY=new w(e.clientXY.x,e.clientXY.y),this.nextClickIsDoubleClick=!!e.nextClickIsDoubleClick,this.hasEverInteracted=!!e.hasEverInteracted,e.element&&c(e.element)&&this.setElement(e.element))}fireMousedown(){return c(this.getElement(),"OPTION")||c(this.getElement(),"SELECT")?!0:this.fireMouseEventInternal(f.MOUSEDOWN)}pressButton(e){if(this.buttonPressed!==null)throw new p(13,"Cannot press more than one button or an already pressed button.");this.buttonPressed=e,this.elementPressed=this.getElement(),this.fireMousedown()&&this.focusOnElement()}releaseButton(e=!1){if(this.buttonPressed===null)throw new p(13,"Cannot release a button when no button is pressed.");this.maybeToggleOption();let t=R(this.getElement());this.fireMouseEventInternal(f.MOUSEUP,void 0,void 0,e);try{this.buttonPressed===0&&this.getElement()===this.elementPressed?(this.clickElement(this.clientXY,this.getButtonValue(f.CLICK),t),this.maybeDoubleClickElement()):this.buttonPressed===2&&this.fireMouseEventInternal(f.CONTEXTMENU)}catch{}this.buttonPressed=null,this.elementPressed=null}maybeDoubleClickElement(){this.nextClickIsDoubleClick&&this.fireMouseEventInternal(f.DBLCLICK),this.nextClickIsDoubleClick=!this.nextClickIsDoubleClick}move(e,t){let o=R(e),s=C(e);this.clientXY.x=t.x+s.left,this.clientXY.y=t.y+s.top;let i=this.getElement();if(e!==i){try{i&&i.ownerDocument.defaultView?.closed&&(i=null)}catch{i=null}if(i){let a=i===document.documentElement||i===document.body;i=!this.hasEverInteracted&&a?null:i,this.fireMouseEventInternal(f.MOUSEOUT,e)}this.setElement(e),this.fireMouseEventInternal(f.MOUSEOVER,i,void 0,o)}this.fireMouseEventInternal(f.MOUSEMOVE,void 0,void 0,o),this.nextClickIsDoubleClick=!1}scroll(e){if(e===0)throw new p(13,"Must scroll a non-zero number of ticks.");let t=e>0?-120:120;for(let o=0;o<Math.abs(e);o++)this.fireMouseEventInternal(f.MOUSEWHEEL,void 0,t)}fireMouseEventInternal(e,t,o,s){return this.hasEverInteracted=!0,this.fireMouseEvent(e,this.clientXY,this.getButtonValue(e),t??null,o??0,s??!1)}getButtonValue(e){let t=P.get(e);if(!t)return 0;let o=this.buttonPressed===null?Rt:this.buttonPressed,s=t[o];if(s===null)throw new p(13,"Event does not permit the specified mouse button.");return s}getState(){return{buttonPressed:this.buttonPressed,elementPressed:this.elementPressed,clientXY:{x:this.clientXY.x,y:this.clientXY.y},nextClickIsDoubleClick:this.nextClickIsDoubleClick,hasEverInteracted:this.hasEverInteracted,element:this.getElement()}}};function Mt(n){if(!z(n,!0))throw new p(11,"Element is not currently visible and may not be manipulated")}function Ee(n,e,t,o){let s=_t(n,e),i=t||new G;i.move(n,s),i.pressButton(0),i.releaseButton(o)}function _t(n,e){if(Mt(n),de(n,e),e)return H.fromCoordinate(e);let t=fe(n);return new H(t.width/2,t.height/2)}function De(n){let e=n,t=e.offsetWidth,o=e.offsetHeight;if(!t&&!o&&e.getBoundingClientRect){let i=e.getBoundingClientRect();return new B(i.right-i.left,i.bottom-i.top)}return new B(t,o)}function Pe(n){if(N(n,"display")!=="none")return De(n);let e=n.style,t=e.display,o=e.visibility,s=e.position;e.visibility="hidden",e.position="absolute",e.display="inline";let i=De(n);return e.display=t,e.position=s,e.visibility=o,i}function fe(n){let e=Pe(n),t=n.offsetParent;return e.width>0&&e.height>0||!t?e:fe(t)}function de(n,e){let t=L(n,e);if(t!=="scroll")return t==="none";if(n.scrollIntoView&&(n.scrollIntoView(),L(n,e)==="none"))return!0;let o=q(n,e);for(let s=I(n);s;s=I(s))At(o,s);return L(n,e)==="none"}function At(n,e){let t=C(e),o=Lt(e),s=n.left-t.left-o.left,i=n.top-t.top-o.top,a=e.clientWidth+n.left-n.right,E=e.clientHeight+n.top-n.bottom;e.scrollLeft+=Math.min(s,Math.max(s-a,0)),e.scrollTop+=Math.min(i,Math.max(i-E,0))}function Lt(n){let e=n.ownerDocument.defaultView?.getComputedStyle(n),t=parseFloat(e?.borderLeftWidth||"0"),o=parseFloat(e?.borderRightWidth||"0"),s=parseFloat(e?.borderTopWidth||"0"),i=parseFloat(e?.borderBottomWidth||"0");return new V(s,o,i,t)}var he="ELEMENT",pe="element-6066-11e4-a52e-4f735466cecf",Te="WINDOW";function Dt(n){return typeof n=="object"&&n!==null||typeof n=="function"}function Pt(n){return Array.isArray(n)?!0:typeof n=="object"&&n!==null&&typeof n.length=="number"}function xt(n){function e(t,o){if(t==null)return null;switch(typeof t){case"string":case"number":case"boolean":return t;case"function":return t.toString();case"object":break;default:return null}let s=t;if(o.includes(s))throw new p(17,"Recursive object cannot be transferred");if("nodeType"in s&&(s.nodeType===Node.ELEMENT_NODE||s.nodeType===Node.DOCUMENT_NODE)){let a=Ue(s);return{[he]:a,[pe]:a}}if("document"in s)return{[Te]:Ue(s)};if(Array.isArray(t))return t.map(a=>e(a,o));if(o=[...o,s],Pt(t))return Array.from(t).map(a=>e(a,o));let i={};for(let a of Object.keys(s))i[a]=e(s[a],o);return i}return e(n,[])}function Q(n,e){if(Array.isArray(n))return n.map(t=>Q(t,e));if(Dt(n)){if(typeof n=="function")return n;let t=n;if(he in t)return j(t[he],e);if(pe in t)return j(t[pe],e);if(Te in t)return j(t[Te],e);let o={};for(let s of Object.keys(t))o[s]=Q(t[s],e);return o}return n}function Ke(n){return{status:0,value:xt(n)}}function ke(n){let e="code"in n&&typeof n.code=="number"?n.code:13;return{status:e,value:{message:n.message,error:te(e)}}}var xe="$wdc_",Ut=":wdc:";function be(n){let e=n||document,t=e[xe];return t||(t=e[xe]={},t.nextId=Date.now()),t.nextId||(t.nextId=Date.now()),t}var Kt=200;function Fe(n,e){if("setInterval"in e)return!!e.closed;let t=e;for(;t;){if(t===n.documentElement)return!1;t.host&&t.nodeType===Node.DOCUMENT_FRAGMENT_NODE&&(t=t.host),t=t.parentNode}return!0}function kt(n){let e=be(n);for(let t of Object.keys(e)){let o=e[t];!o||typeof o!="object"||Fe(n,o)&&delete e[t]}}function Ue(n){let e=n.ownerDocument,t=be(e);Object.keys(t).length-1>=Kt&&kt(e||document);let o;for(let s of Object.keys(t))if(t[s]===n){o=s;break}return o||(o=Ut+t.nextId++,t[o]=n),o}function j(n,e){n=decodeURIComponent(n);let t=e||document,o=be(t);if(!(n in o))throw new p(10,"Element does not exist in cache");let s=o[n];if(!Fe(t,s))return s;throw delete o[n],"setInterval"in s?new p(23,"Window has been closed."):new p(10,"Element is no longer attached to the DOM")}var u={NULL:"\uE000",CANCEL:"\uE001",HELP:"\uE002",BACK_SPACE:"\uE003",TAB:"\uE004",CLEAR:"\uE005",RETURN:"\uE006",ENTER:"\uE007",SHIFT:"\uE008",CONTROL:"\uE009",ALT:"\uE00A",PAUSE:"\uE00B",ESCAPE:"\uE00C",SPACE:"\uE00D",PAGE_UP:"\uE00E",PAGE_DOWN:"\uE00F",END:"\uE010",HOME:"\uE011",ARROW_LEFT:"\uE012",LEFT:"\uE012",ARROW_UP:"\uE013",UP:"\uE013",ARROW_RIGHT:"\uE014",RIGHT:"\uE014",ARROW_DOWN:"\uE015",DOWN:"\uE015",INSERT:"\uE016",DELETE:"\uE017",SEMICOLON:"\uE018",EQUALS:"\uE019",NUMPAD0:"\uE01A",NUMPAD1:"\uE01B",NUMPAD2:"\uE01C",NUMPAD3:"\uE01D",NUMPAD4:"\uE01E",NUMPAD5:"\uE01F",NUMPAD6:"\uE020",NUMPAD7:"\uE021",NUMPAD8:"\uE022",NUMPAD9:"\uE023",MULTIPLY:"\uE024",ADD:"\uE025",SEPARATOR:"\uE026",SUBTRACT:"\uE027",DECIMAL:"\uE028",DIVIDE:"\uE029",F1:"\uE031",F2:"\uE032",F3:"\uE033",F4:"\uE034",F5:"\uE035",F6:"\uE036",F7:"\uE037",F8:"\uE038",F9:"\uE039",F10:"\uE03A",F11:"\uE03B",F12:"\uE03C",COMMAND:"\uE03D",META:"\uE03D"};var io=new Map([[u.NULL,null],[u.BACK_SPACE,l.BACKSPACE],[u.TAB,l.TAB],[u.RETURN,l.ENTER],[u.ENTER,l.ENTER],[u.SHIFT,l.SHIFT],[u.CONTROL,l.CONTROL],[u.ALT,l.ALT],[u.PAUSE,l.PAUSE],[u.ESCAPE,l.ESC],[u.SPACE,l.SPACE],[u.PAGE_UP,l.PAGE_UP],[u.PAGE_DOWN,l.PAGE_DOWN],[u.END,l.END],[u.HOME,l.HOME],[u.LEFT,l.LEFT],[u.UP,l.UP],[u.RIGHT,l.RIGHT],[u.DOWN,l.DOWN],[u.INSERT,l.INSERT],[u.DELETE,l.DELETE],[u.SEMICOLON,l.SEMICOLON],[u.EQUALS,l.EQUALS],[u.NUMPAD0,l.NUM_ZERO],[u.NUMPAD1,l.NUM_ONE],[u.NUMPAD2,l.NUM_TWO],[u.NUMPAD3,l.NUM_THREE],[u.NUMPAD4,l.NUM_FOUR],[u.NUMPAD5,l.NUM_FIVE],[u.NUMPAD6,l.NUM_SIX],[u.NUMPAD7,l.NUM_SEVEN],[u.NUMPAD8,l.NUM_EIGHT],[u.NUMPAD9,l.NUM_NINE],[u.MULTIPLY,l.NUM_MULTIPLY],[u.ADD,l.NUM_PLUS],[u.SUBTRACT,l.NUM_MINUS],[u.DECIMAL,l.NUM_PERIOD],[u.DIVIDE,l.NUM_DIVISION],[u.SEPARATOR,l.SEPARATOR],[u.F1,l.F1],[u.F2,l.F2],[u.F3,l.F3],[u.F4,l.F4],[u.F5,l.F5],[u.F6,l.F6],[u.F7,l.F7],[u.F8,l.F8],[u.F9,l.F9],[u.F10,l.F10],[u.F11,l.F11],[u.F12,l.F12],[u.META,l.META]]);function Ve(n){return n?j(n.WINDOW):window}function Bt(n,e,t){let o;try{let s=Ve(t),i=Q(e,s.document),a=n.apply(null,i);o=Ke(a)}catch(s){o=ke(s)}return JSON.stringify(o)}function We(n,e){return Bt(Ee,[n],e)}return qe(Vt);})();
 
 return AtomExport.default.apply(null,arguments);}
