+++ atoms/src/entrypoints/clear_session_storage.ts
@@ -1,1 +1,1 @@
-export {clear as default} from '../webdriver/inject/sessionStorage.js';
+export {clear as default} from '../webdriver/inject/session-storage.js';
