+++ atoms/src/entrypoints/find_element_fragment.ts
@@ -1,1 +1,1 @@
-export {findElement as default} from '../webdriver/inject/findElement.js';
+export {findElement as default} from '../webdriver/inject/find-element.js';
