+++ atoms/src/entrypoints/find_elements.ts
@@ -1,1 +1,1 @@
-export {findElements as default} from '../webdriver/inject/findElement.js';
+export {findElements as default} from '../webdriver/inject/find-element.js';
