+++ atoms/src/entrypoints/get_session_storage_keys.ts
@@ -1,1 +1,1 @@
-export {keySet as default} from '../webdriver/inject/sessionStorage.js';
+export {keySet as default} from '../webdriver/inject/session-storage.js';
