+++ atoms/src/entrypoints/get_session_storage_size.ts
@@ -1,1 +1,1 @@
-export {size as default} from '../webdriver/inject/sessionStorage.js';
+export {size as default} from '../webdriver/inject/session-storage.js';
