+++ atoms/src/entrypoints/remove_local_storage_item.ts
@@ -1,1 +1,1 @@
-export {removeItem as default} from '../webdriver/inject/localStorage.js';
+export {removeItem as default} from '../webdriver/inject/local-storage.js';
