+++ atoms/src/entrypoints/remove_session_storage_item.ts
@@ -1,1 +1,1 @@
-export {removeItem as default} from '../webdriver/inject/sessionStorage.js';
+export {removeItem as default} from '../webdriver/inject/session-storage.js';
