+++ atoms/submit.js
@@ -1,3 +1,3 @@
-function(){"use strict";var AtomExport=(()=>{var J=Object.defineProperty;var Ve=Object.getOwnPropertyDescriptor;var We=Object.getOwnPropertyNames;var Xe=Object.prototype.hasOwnProperty;var Ye=(n,e)=>{for(var t in e)J(n,t,{get:e[t],enumerable:!0})},Ge=(n,e,t,o)=>{if(e&&typeof e=="object"||typeof e=="function")for(let s of We(e))!Xe.call(n,s)&&s!==t&&J(n,s,{get:()=>e[s],enumerable:!(o=Ve(e,s))||o.enumerable});return n};var je=n=>Ge(J({},"__esModule",{value:!0}),n);var Ut={};Ye(Ut,{default:()=>ke});var Ne={aliceblue:"#f0f8ff",antiquewhite:"#faebd7",aqua:"#00ffff",aquamarine:"#7fffd4",azure:"#f0ffff",beige:"#f5f5dc",bisque:"#ffe4c4",black:"#000000",blanchedalmond:"#ffebcd",blue:"#0000ff",blueviolet:"#8a2be2",brown:"#a52a2a",burlywood:"#deb887",cadetblue:"#5f9ea0",chartreuse:"#7fff00",chocolate:"#d2691e",coral:"#ff7f50",cornflowerblue:"#6495ed",cornsilk:"#fff8dc",crimson:"#dc143c",cyan:"#00ffff",darkblue:"#00008b",darkcyan:"#008b8b",darkgoldenrod:"#b8860b",darkgray:"#a9a9a9",darkgreen:"#006400",darkgrey:"#a9a9a9",darkkhaki:"#bdb76b",darkmagenta:"#8b008b",darkolivegreen:"#556b2f",darkorange:"#ff8c00",darkorchid:"#9932cc",darkred:"#8b0000",darksalmon:"#e9967a",darkseagreen:"#8fbc8f",darkslateblue:"#483d8b",darkslategray:"#2f4f4f",darkslategrey:"#2f4f4f",darkturquoise:"#00ced1",darkviolet:"#9400d3",deeppink:"#ff1493",deepskyblue:"#00bfff",dimgray:"#696969",dimgrey:"#696969",dodgerblue:"#1e90ff",firebrick:"#b22222",floralwhite:"#fffaf0",forestgreen:"#228b22",fuchsia:"#ff00ff",gainsboro:"#dcdcdc",ghostwhite:"#f8f8ff",gold:"#ffd700",goldenrod:"#daa520",gray:"#808080",green:"#008000",greenyellow:"#adff2f",grey:"#808080",honeydew:"#f0fff0",hotpink:"#ff69b4",indianred:"#cd5c5c",indigo:"#4b0082",ivory:"#fffff0",khaki:"#f0e68c",lavender:"#e6e6fa",lavenderblush:"#fff0f5",lawngreen:"#7cfc00",lemonchiffon:"#fffacd",lightblue:"#add8e6",lightcoral:"#f08080",lightcyan:"#e0ffff",lightgoldenrodyellow:"#fafad2",lightgray:"#d3d3d3",lightgreen:"#90ee90",lightgrey:"#d3d3d3",lightpink:"#ffb6c1",lightsalmon:"#ffa07a",lightseagreen:"#20b2aa",lightskyblue:"#87cefa",lightslategray:"#778899",lightslategrey:"#778899",lightsteelblue:"#b0c4de",lightyellow:"#ffffe0",lime:"#00ff00",limegreen:"#32cd32",linen:"#faf0e6",magenta:"#ff00ff",maroon:"#800000",mediumaquamarine:"#66cdaa",mediumblue:"#0000cd",mediumorchid:"#ba55d3",mediumpurple:"#9370db",mediumseagreen:"#3cb371",mediumslateblue:"#7b68ee",mediumspringgreen:"#00fa9a",mediumturquoise:"#48d1cc",mediumvioletred:"#c71585",midnightblue:"#191970",mintcream:"#f5fffa",mistyrose:"#ffe4e1",moccasin:"#ffe4b5",navajowhite:"#ffdead",navy:"#000080",oldlace:"#fdf5e6",olive:"#808000",olivedrab:"#6b8e23",orange:"#ffa500",orangered:"#ff4500",orchid:"#da70d6",palegoldenrod:"#eee8aa",palegreen:"#98fb98",paleturquoise:"#afeeee",palevioletred:"#db7093",papayawhip:"#ffefd5",peachpuff:"#ffdab9",peru:"#cd853f",pink:"#ffc0cb",plum:"#dda0dd",powderblue:"#b0e0e6",purple:"#800080",red:"#ff0000",rosybrown:"#bc8f8f",royalblue:"#4169e1",saddlebrown:"#8b4513",salmon:"#fa8072",sandybrown:"#f4a460",seagreen:"#2e8b57",seashell:"#fff5ee",sienna:"#a0522d",silver:"#c0c0c0",skyblue:"#87ceeb",slateblue:"#6a5acd",slategray:"#708090",slategrey:"#708090",snow:"#fffafa",springgreen:"#00ff7f",steelblue:"#4682b4",tan:"#d2b48c",teal:"#008080",thistle:"#d8bfd8",tomato:"#ff6347",turquoise:"#40e0d0",violet:"#ee82ee",wheat:"#f5deb3",white:"#ffffff",whitesmoke:"#f5f5f5",yellow:"#ffff00",yellowgreen:"#9acd32"};var $e=new Set(["backgroundColor","borderTopColor","borderRightColor","borderBottomColor","borderLeftColor","color","outlineColor"]),ze=/#([0-9a-fA-F])([0-9a-fA-F])([0-9a-fA-F])/,Je=/^#(?:[0-9a-f]{3}){1,2}$/i,qe=/^(?:rgba)?\((\d{1,3}),\s?(\d{1,3}),\s?(\d{1,3}),\s?(0|1|0\.\d*)\)$/i,Ze=/^(?:rgb)?\((0|[1-9]\d{0,2}),\s?(0|[1-9]\d{0,2}),\s?(0|[1-9]\d{0,2})\)$/i;function Qe(n){n=n.toLowerCase();let e=Ne[n];if(!e&&(e=n.charAt(0)==="#"?n:`#${n}`,e.length===4&&(e=e.replace(ze,"#$1$1$2$2$3$3")),!Je.test(e)))return null;let t=parseInt(e.substring(1,3),16),o=parseInt(e.substring(3,5),16),s=parseInt(e.substring(5,7),16);return[t,o,s,1]}function et(n){let e=n.match(qe);if(e){let t=Number(e[1]),o=Number(e[2]),s=Number(e[3]),i=Number(e[4]);if(t>=0&&t<=255&&o>=0&&o<=255&&s>=0&&s<=255&&i>=0&&i<=1)return[t,o,s,i]}return null}function tt(n){let e=n.match(Ze);if(e){let t=Number(e[1]),o=Number(e[2]),s=Number(e[3]);if(t>=0&&t<=255&&o>=0&&o<=255&&s>=0&&s<=255)return[t,o,s,1]}return null}function ge(n,e){if(!$e.has(n))return e;let t=et(e)||tt(e)||Qe(e);return t?`rgba(${t.join(", ")})`:e}var nt={15:"element not selectable",11:"element not visible",31:"unknown error",30:"unknown error",24:"invalid cookie domain",29:"invalid element coordinates",12:"invalid element state",32:"invalid selector",51:"invalid selector",52:"invalid selector",17:"javascript error",405:"unsupported operation",34:"move target out of bounds",27:"no such alert",7:"no such element",8:"no such frame",23:"no such window",28:"script timeout",33:"session not created",10:"stale element reference",21:"timeout",25:"unable to set cookie",26:"unexpected alert open",13:"unknown error",9:"unknown command"};function q(n){return nt[n]??"unknown error"}var b=class extends Error{code;state;isAutomationError=!0;constructor(e,t=""){super(t),this.code=e,this.state=q(e);let o=this.state.replace(/(^|\s+)[a-z]/g,i=>i.toUpperCase().trimStart()).replace(/\s+/g,"");this.name=/Error$/.test(o)?o:`${o}Error`;let s=new Error(this.message);s.name=this.name,this.stack=s.stack||""}};var ot=/[;]+(?=(?:(?:[^"]*"){2})*[^"]*$)(?=(?:(?:[^']*'){2})*[^']*$)(?=(?:[^()]*\([^()]*\))*[^()]*$)/;function st(n){let e=n.split(ot),t=[];for(let s of e){let i=s.indexOf(":");if(i>0){let a=[s.slice(0,i),s.slice(i+1)];t.push(a[0].toLowerCase(),":",a[1],";")}}let o=t.join("");return o.charAt(o.length-1)===";"?o:`${o};`}function Oe(n,e){if(e=e.toLowerCase(),e==="style")return st(n.style.cssText);let t=n.getAttributeNode(e);return t&&t.specified?t.value:null}function L(n,e){return n[e]}function c(n,e){return n instanceof HTMLFormElement?n.nodeType===Node.ELEMENT_NODE&&(!e||e==="FORM"):!!n&&n.nodeType===Node.ELEMENT_NODE&&(!e||n.tagName.toUpperCase()===e)}function U(n){return n.nodeType===Node.DOCUMENT_NODE?n:n.ownerDocument}function W(n,e,t=!1){let o=t?n:n.parentNode;for(;o;){if(e(o))return o;o=o.parentNode}return null}function K(n){if(c(n,"OPTION"))return!0;if(c(n,"INPUT")){let e=n.type.toLowerCase();return e==="checkbox"||e==="radio"}return!1}function F(n){if(!K(n))throw new b(15,"Element is not selectable");let e=n.type?.toLowerCase();return!!L(n,e==="checkbox"||e==="radio"?"checked":"selected")}function ye(n,e){if(!n)throw new b(32,"No selector specified");n=n.trim();let t;try{t=e.querySelector(n)}catch{throw new b(32,"An invalid or illegal selector was specified")}return t&&t.nodeType===Node.ELEMENT_NODE?t:null}var S=class{x;y;constructor(e=0,t=0){this.x=e,this.y=t}};var X=class{top;right;bottom;left;constructor(e,t,o,s){this.top=e,this.right=t,this.bottom=o,this.left=s}},N=class n{left;top;width;height;constructor(e,t,o,s){this.left=e,this.top=t,this.width=o,this.height=s}toBox(){let e=this.left+this.width,t=this.top+this.height;return new X(this.top,e,t,this.left)}intersects(e){return n.intersects(this,e)}static intersects(e,t){return e.left<=t.left+t.width&&t.left<=e.left+e.width&&e.top<=t.top+t.height&&t.top<=e.top+e.height}};function k(n,e,t){return Math.min(Math.max(n,e),t)}var it=typeof ShadowRoot=="function";function lt(n){return n.replace(/-([a-z])/g,(e,t)=>t.toUpperCase())}function G(n){let e="document"in n?n.document:U(n);try{let t=e?.activeElement;return t&&t.nodeName?t:null}catch{return null}}function M(n){return ee(n,!0)&&ve(n)&&!ut(n)}function ut(n){return g(n,"pointer-events")==="none"}var at=["A","AREA","BUTTON","INPUT","LABEL","SELECT","TEXTAREA"];function we(n){return at.some(e=>c(n,e))||Oe(n,"tabindex")!=null&&Number(L(n,"tabIndex"))>=0||Q(n)}var ct=["BUTTON","INPUT","OPTGROUP","OPTION","SELECT","TEXTAREA"];function ve(n){if(!ct.some(o=>c(n,o)))return!0;if(L(n,"disabled"))return!1;if(n.parentNode&&n.parentNode.nodeType===Node.ELEMENT_NODE&&c(n,"OPTGROUP")||c(n,"OPTION"))return ve(n.parentNode);let t=n;for(;t;){let o=t.parentNode;if(o&&c(o,"FIELDSET")&&L(o,"disabled")){if(!c(t,"LEGEND"))return!1;let s=t.previousElementSibling;for(;s;){if(c(s,"LEGEND"))return!1;s=s.previousElementSibling}}t=o}return!0}var Et=new Set(["text","search","tel","url","email","password","number"]);function ft(n){if(c(n,"TEXTAREA"))return!0;if(c(n,"INPUT")){let e=n.type.toLowerCase();return Et.has(e)}return Se(n)}function dt(n){return c(n,"INPUT")&&n.type.toLowerCase()==="file"}function I(n,e){return c(n,"INPUT")&&n.type.toLowerCase()===e}function Se(n){let e=n;if(e.contentEditable===void 0)return!1;if(e.isContentEditable!==void 0)return e.isContentEditable;function t(o){if(o.contentEditable==="inherit"){let s=D(o);return s?t(s):!1}return o.contentEditable==="true"}return t(e)}function Q(n){return(ft(n)||dt(n)||I(n,"range")||I(n,"date")||I(n,"month")||I(n,"week")||I(n,"time")||I(n,"datetime-local")||I(n,"color"))&&!L(n,"readOnly")}function D(n){let e=n.parentNode;for(;e&&e.nodeType!==Node.ELEMENT_NODE&&e.nodeType!==Node.DOCUMENT_NODE&&e.nodeType!==Node.DOCUMENT_FRAGMENT_NODE;)e=e.parentNode;return e&&c(e)?e:null}function mt(n,e){let t=n.ownerDocument.defaultView;if(t?.getComputedStyle){let o=t.getComputedStyle(n,null);if(o)return o[e]||o.getPropertyValue(e)||""}return""}function Re(n,e){let t=n.style,o=t[e];if(o===void 0&&typeof t.getPropertyValue=="function"&&(o=t.getPropertyValue(e)),o!=="inherit")return o!==void 0?o:null;let s=D(n);return s?Re(s,e):null}function g(n,e){let t=lt(e);(t==="float"||t==="cssFloat"||t==="styleFloat")&&(t="cssFloat");let o=mt(n,t)||Re(n,t);return o===null?null:ge(t,o)}function Z(n,e,t){if(!c(n))throw new Error("Argument to isShown must be of type Element");if(c(n,"BODY"))return!0;if(c(n,"OPTION")||c(n,"OPTGROUP")){let E=n.parentNode;for(;E&&!c(E,"SELECT");)E=E.parentNode;return!!E&&Z(E,!0,t)}let o=Ce(n);if(o)return!!o.image&&o.rect.width>0&&o.rect.height>0&&Z(o.image,e,t);if(c(n,"INPUT")&&n.type.toLowerCase()==="hidden"||c(n,"NOSCRIPT"))return!1;let s=g(n,"visibility");if(s==="collapse"||s==="hidden"||!t(n)||!e&&Ie(n)===0)return!1;function i(E){let f=_(E);if(f.height>0&&f.width>0)return!0;if(c(E,"PATH")&&(f.height>0||f.width>0)){let p=g(E,"stroke-width");return!!p&&parseInt(p,10)>0}let y=g(E,"visibility");return y==="collapse"||y==="hidden"||!t(E)?!1:g(E,"overflow")!=="hidden"&&[...E.childNodes].some(p=>{if(p.nodeType===Node.TEXT_NODE){let w=p.nodeValue??"";return!(/^[\s]*$/.test(w)&&/[\n\r\t]/.test(w))}return c(p)&&i(p)})}if(!i(n))return!1;function a(E){return Y(E)==="hidden"&&[...E.childNodes].every(f=>!c(f)||a(f)||!i(f))}return!a(n)}function ee(n,e=!1){function t(o){if(c(o)){let i=o;if(g(i,"display")==="none"||g(i,"content-visibility")==="hidden")return!1}let s=pt(o);if(it&&s instanceof ShadowRoot){if(s.host.shadowRoot&&s.host.shadowRoot!==s)return!1;s=s.host}return s&&(s.nodeType===Node.DOCUMENT_NODE||s.nodeType===Node.DOCUMENT_FRAGMENT_NODE)?!0:s&&c(s,"DETAILS")&&!s.open&&!c(o,"SUMMARY")?!1:!!s&&t(s)}return Z(n,e,t)}function j(n){let e=n.scrollingElement||n.documentElement,t=n.defaultView;return new S(t.pageXOffset||e.scrollLeft,t.pageYOffset||e.scrollTop)}function Y(n,e){let t=te(n,e),o=U(n),s=o.documentElement,i=o.body,a=g(s,"overflow"),E=!1;function f(d,T){if(d===s)return!0;let m=g(d,"display");return!(m?.startsWith("inline")||m==="contents"||T==="absolute"&&g(d,"position")==="static")}function y(d){let T=g(d,"position");if(T==="fixed")return E=!0,d===s?null:s;let m=D(d);for(;m&&!f(m,T);)m=D(m);return m}function p(d){let T=d;if(a==="visible"){if(d===s&&i)T=i;else if(d===i)return{x:"visible",y:"visible"}}let m={x:g(T,"overflow-x")??"",y:g(T,"overflow-y")??""};return d===s&&(m.x=m.x==="visible"?"auto":m.x,m.y=m.y==="visible"?"auto":m.y),m}function w(d){return d===s?j(o):new S(d.scrollLeft,d.scrollTop)}for(let d=y(n);d;d=y(d)){let T=p(d);if(T.x==="visible"&&T.y==="visible")continue;let m=_(d);if(m.width===0||m.height===0)return"hidden";let he=t.right<m.left,pe=t.bottom<m.top;if(he&&T.x==="hidden"||pe&&T.y==="hidden")return"hidden";if(he&&T.x!=="visible"||pe&&T.y!=="visible"){let x=w(d),He=t.right<m.left-x.x,Be=t.bottom<m.top-x.y;return He&&T.x!=="visible"||Be&&T.x!=="visible"||Y(d)==="hidden"?"hidden":"scroll"}let Te=t.left>=m.left+m.width,be=t.top>=m.top+m.height;if(Te&&T.x==="hidden"||be&&T.y==="hidden")return"hidden";if(Te&&T.x!=="visible"||be&&T.y!=="visible"){if(E){let x=w(d);if(t.left>=s.scrollWidth-x.x||t.right>=s.scrollHeight-x.y)return"hidden"}return Y(d)==="hidden"?"hidden":"scroll"}}return"none"}function _(n){let e=Ce(n);if(e)return e.rect;if(c(n,"HTML")){let o=U(n),s=o.defaultView,i=o.compatMode==="CSS1Compat"?o.documentElement:o.body;return new N(0,0,i.clientWidth||s.innerWidth,i.clientHeight||s.innerHeight)}let t;try{t=n.getBoundingClientRect()}catch{return new N(0,0,0,0)}return new N(t.left,t.top,t.right-t.left,t.bottom-t.top)}function Ce(n){let e=c(n,"MAP");if(!e&&!c(n,"AREA"))return null;let t=e?n:n.parentNode&&c(n.parentNode,"MAP")?n.parentNode:null,o=null,s=null,i=t&&t.name;if(t&&i){let a=U(t),E=`*[usemap="#${i}"]`;if(o=ye(E,a),o&&(s=_(o),!e&&n.shape.toLowerCase()!=="default")){let f=ht(n),y=Math.min(Math.max(f.left,0),s.width),p=Math.min(Math.max(f.top,0),s.height),w=Math.min(f.width,s.width-y),d=Math.min(f.height,s.height-p);s=new N(y+s.left,p+s.top,w,d)}}return{image:o,rect:s||new N(0,0,0,0)}}function ht(n){let e=n,t=e.shape.toLowerCase(),o=e.coords.split(",");if(t==="rect"&&o.length===4){let s=Number(o[0]),i=Number(o[1]);return new N(s,i,Number(o[2])-s,Number(o[3])-i)}else if(t==="circle"&&o.length===3){let s=Number(o[0]),i=Number(o[1]),a=Number(o[2]);return new N(s-a,i-a,2*a,2*a)}else if(t==="poly"&&o.length>2){let s=Number(o[0]),i=Number(o[1]),a=s,E=i;for(let f=2;f+1<o.length;f+=2)s=Math.min(s,Number(o[f])),a=Math.max(a,Number(o[f])),i=Math.min(i,Number(o[f+1])),E=Math.max(E,Number(o[f+1]));return new N(s,i,a-s,E-i)}return new N(0,0,0,0)}function te(n,e){let t=_(n).toBox();if(e){let o=e instanceof N?e:new N(e.x,e.y,1,1);t.left=k(t.left+o.left,t.left,t.right),t.top=k(t.top+o.top,t.top,t.bottom),t.right=k(t.left+o.width,t.left,t.right),t.bottom=k(t.top+o.height,t.top,t.bottom)}return t}function Ie(n){let e=1,t=g(n,"opacity");t&&(e=Number(t));let o=D(n);return o&&(e*=Ie(o)),e}function pt(n){let e=n.parentNode;if(e&&e.shadowRoot&&n.assignedSlot!==void 0)return n.assignedSlot?n.assignedSlot.parentNode:null;let t=n;if(t.getDestinationInsertionPoints){let o=t.getDestinationInsertionPoints();if(o.length>0)return o[o.length-1]}return e}var O=class{type;bubbles;cancelable;constructor(e,t,o){this.type=e,this.bubbles=t,this.cancelable=o}create(e,t){let s=("ownerDocument"in e?e.ownerDocument:e.document).createEvent("HTMLEvents");return s.initEvent(this.type,this.bubbles,this.cancelable),s}toString(){return this.type}},v=class extends O{create(e,t){let o=t,s="ownerDocument"in e?e.ownerDocument:e.document,i=s.defaultView,a=s.createEvent("MouseEvents"),E=1;return this.type==="mousewheel"&&(a.wheelDelta=o.wheelDelta),a.initMouseEvent(this.type,this.bubbles,this.cancelable,i,E,o.clientX,o.clientY,o.clientX,o.clientY,o.ctrlKey,o.altKey,o.shiftKey,o.metaKey,o.button,o.relatedTarget),a}},H=class extends O{create(e,t){let o=t,i=("ownerDocument"in e?e.ownerDocument:e.document).createEvent("Events");return i.initEvent(this.type,this.bubbles,this.cancelable),i.altKey=o.altKey,i.ctrlKey=o.ctrlKey,i.metaKey=o.metaKey,i.shiftKey=o.shiftKey,i.keyCode=o.charCode||o.keyCode,i.charCode=this.type==="keypress"?i.keyCode:0,i}};function ne(n,e){return e.map(t=>new Touch({identifier:t.identifier,target:n,screenX:t.screenX,screenY:t.screenY,clientX:t.clientX,clientY:t.clientY,pageX:t.pageX,pageY:t.pageY}))}var B=class extends O{create(e,t){let o=t,s="ownerDocument"in e?e.ownerDocument:e.document,i=s.defaultView,a=typeof TouchEvent.prototype.initTouchEvent=="function"?"INIT_TOUCH_EVENT":"TOUCH_EVENT_CTOR",E=ne(e,o.changedTouches),f=o.touches===o.changedTouches?E:ne(e,o.touches),y=o.targetTouches===o.changedTouches?E:ne(e,o.targetTouches);if(a==="INIT_TOUCH_EVENT"){let p=s.createEvent("TouchEvent");return p.initTouchEvent(this.type,this.bubbles,this.cancelable,i,1,0,0,o.clientX,o.clientY,o.ctrlKey,o.altKey,o.shiftKey,o.metaKey,f,y,E,o.scale,o.rotation),p.relatedTarget=o.relatedTarget,p}return new TouchEvent(this.type,{touches:f,targetTouches:y,changedTouches:E,bubbles:this.bubbles,cancelable:this.cancelable,ctrlKey:o.ctrlKey,shiftKey:o.shiftKey,altKey:o.altKey,metaKey:o.metaKey})}},h={BLUR:new O("blur",!1,!1),CHANGE:new O("change",!0,!1),FOCUS:new O("focus",!1,!1),FOCUSIN:new O("focusin",!0,!1),FOCUSOUT:new O("focusout",!0,!1),INPUT:new O("input",!0,!1),ORIENTATIONCHANGE:new O("orientationchange",!1,!1),SELECT:new O("select",!0,!1),SUBMIT:new O("submit",!0,!0),TEXTINPUT:new O("textInput",!0,!0),CLICK:new v("click",!0,!0),CONTEXTMENU:new v("contextmenu",!0,!0),DBLCLICK:new v("dblclick",!0,!0),MOUSEDOWN:new v("mousedown",!0,!0),MOUSEMOVE:new v("mousemove",!0,!1),MOUSEOUT:new v("mouseout",!0,!0),MOUSEOVER:new v("mouseover",!0,!0),MOUSEUP:new v("mouseup",!0,!0),MOUSEWHEEL:new v("mousewheel",!0,!0),KEYDOWN:new H("keydown",!0,!0),KEYPRESS:new H("keypress",!0,!0),KEYUP:new H("keyup",!0,!0),TOUCHEND:new B("touchend",!0,!0),TOUCHMOVE:new B("touchmove",!0,!0),TOUCHSTART:new B("touchstart",!0,!0)};function R(n,e,t){let o=e.create(n,t);return"isTrusted"in o||(o.isTrusted=!1),n.dispatchEvent(o)}var oe=class{pressedModifiers=0;isPressed(e){return(this.pressedModifiers&e)!==0}setPressed(e,t){t?this.pressedModifiers|=e:this.pressedModifiers&=~e}isShiftPressed(){return this.isPressed(1)}isControlPressed(){return this.isPressed(2)}isAltPressed(){return this.isPressed(4)}isMetaPressed(){return this.isPressed(8)}},se=class{fireHtmlEvent(e,t){return R(e,t)}fireKeyboardEvent(e,t,o){return R(e,t,o)}fireMouseEvent(e,t,o){return R(e,t,o)}fireTouchEvent(e,t,o){return R(e,t,o)}};function Me(n){return c(n,"FORM")}function re(n){return W(n,Me,!0)}var A=class{el;select=null;modifiersState;eventEmitter;constructor(e,t){this.el=document.documentElement;let o=G(this.el);o&&this.setElement(o),this.modifiersState=e||new oe,this.eventEmitter=t||new se}getElement(){return this.el}setElement(e){this.el=e,c(e,"OPTION")?this.select=W(e,t=>c(t,"SELECT")):this.select=null}fireHtmlEvent(e){return this.eventEmitter.fireHtmlEvent(this.el,e)}fireKeyboardEvent(e,t){return this.eventEmitter.fireKeyboardEvent(this.el,e,t)}fireMouseEvent(e,t,o,s=null,i=0,a=!1){if(!a&&!M(this.el))return!1;if(s&&!(e===h.MOUSEOVER||e===h.MOUSEOUT))throw new b(12,`Event type does not allow related target: ${e}`);let E={clientX:t.x,clientY:t.y,button:o,altKey:this.modifiersState.isAltPressed(),ctrlKey:this.modifiersState.isControlPressed(),shiftKey:this.modifiersState.isShiftPressed(),metaKey:this.modifiersState.isMetaPressed(),wheelDelta:i,relatedTarget:s},f=this.select?this.getTargetOfOptionMouseEvent(e):this.el;return f?this.eventEmitter.fireMouseEvent(f,e,E):!0}fireTouchEvent(e,t,o,s,i){let a={touches:[],targetTouches:[],changedTouches:[],altKey:this.modifiersState.isAltPressed(),ctrlKey:this.modifiersState.isControlPressed(),shiftKey:this.modifiersState.isShiftPressed(),metaKey:this.modifiersState.isMetaPressed(),relatedTarget:null,scale:0,rotation:0,clientX:o.x,clientY:o.y},E=j(this.el.ownerDocument),f=(y,p)=>{let w={identifier:y,screenX:p.x,screenY:p.y,clientX:p.x,clientY:p.y,pageX:p.x+E.x,pageY:p.y+E.y};a.changedTouches.push(w),(e===h.TOUCHSTART||e===h.TOUCHMOVE)&&(a.touches.push(w),a.targetTouches.push(w))};return f(t,o),s!==void 0&&i!==void 0&&f(s,i),this.eventEmitter.fireTouchEvent(this.el,e,a)}getTargetOfOptionMouseEvent(e){let t=this.select;switch(e){case h.CLICK:case h.MOUSEUP:return t.multiple?this.el:this.select;default:return t.multiple?this.el:null}}clickElement(e,t,o=!1){if(!o&&!M(this.el))return;let s=!this.select&&K(this.el),i=s&&F(this.el);this.fireMouseEvent(h.CLICK,e,t,null,0,o)&&s&&this.toggleRadioButtonOrCheckbox(i)}maybeToggleOption(){if(!this.select||!M(this.el))return;let e=this.select,t=F(this.el);t&&!e.multiple||(this.el.selected=!t,e.multiple||R(e,h.CHANGE))}toggleRadioButtonOrCheckbox(e){}focusOnElement(){let e=W(this.el,o=>!!o&&c(o)&&we(o),!0);e=e||this.el;let t=G(e);return e===t?!1:(t&&typeof t.blur=="function"&&(c(t,"BODY")||t.blur()),typeof e.focus=="function"?(e.focus(),!0):!1)}submitForm(e){if(!Me(e))throw new b(12,"Element is not a form, so could not submit.");R(e,h.SUBMIT)&&(c(e.submit)?HTMLFormElement.prototype.submit.call(e):e.submit())}};function Tt(n){return navigator.userAgent.includes(n)}function ie(){return Tt("Macintosh")}var ue=class{code;character;shiftChar;constructor(e,t,o){this.code=e,this.character=t||null,this.shiftChar=o||this.character}},le=new Map;function r(n,e,t){let o=new ue(n,e,t);return e&&(!le.has(e)||t)&&(le.set(e,{key:o,shift:!1}),t&&le.set(t,{key:o,shift:!0})),o}var l={BACKSPACE:r(8),TAB:r(9),ENTER:r(13),SHIFT:r(16),CONTROL:r(17),ALT:r(18),PAUSE:r(19),CAPS_LOCK:r(20),ESC:r(27),SPACE:r(32," "),PAGE_UP:r(33),PAGE_DOWN:r(34),END:r(35),HOME:r(36),LEFT:r(37),UP:r(38),RIGHT:r(39),DOWN:r(40),PRINT_SCREEN:r(44),INSERT:r(45),DELETE:r(46),ZERO:r(48,"0",")"),ONE:r(49,"1","!"),TWO:r(50,"2","@"),THREE:r(51,"3","#"),FOUR:r(52,"4","$"),FIVE:r(53,"5","%"),SIX:r(54,"6","^"),SEVEN:r(55,"7","&"),EIGHT:r(56,"8","*"),NINE:r(57,"9","("),A:r(65,"a","A"),B:r(66,"b","B"),C:r(67,"c","C"),D:r(68,"d","D"),E:r(69,"e","E"),F:r(70,"f","F"),G:r(71,"g","G"),H:r(72,"h","H"),I:r(73,"i","I"),J:r(74,"j","J"),K:r(75,"k","K"),L:r(76,"l","L"),M:r(77,"m","M"),N:r(78,"n","N"),O:r(79,"o","O"),P:r(80,"p","P"),Q:r(81,"q","Q"),R:r(82,"r","R"),S:r(83,"s","S"),T:r(84,"t","T"),U:r(85,"u","U"),V:r(86,"v","V"),W:r(87,"w","W"),X:r(88,"x","X"),Y:r(89,"y","Y"),Z:r(90,"z","Z"),META:r(91),META_RIGHT:r(ie()?93:92),CONTEXT_MENU:r(ie()?0:null),NUM_ZERO:r(96,"0"),NUM_ONE:r(97,"1"),NUM_TWO:r(98,"2"),NUM_THREE:r(99,"3"),NUM_FOUR:r(100,"4"),NUM_FIVE:r(101,"5"),NUM_SIX:r(102,"6"),NUM_SEVEN:r(103,"7"),NUM_EIGHT:r(104,"8"),NUM_NINE:r(105,"9"),NUM_MULTIPLY:r(106,"*"),NUM_PLUS:r(107,"+"),NUM_MINUS:r(109,"-"),NUM_PERIOD:r(110,"."),NUM_DIVISION:r(111,"/"),NUM_LOCK:r(144),F1:r(112),F2:r(113),F3:r(114),F4:r(115),F5:r(116),F6:r(117),F7:r(118),F8:r(119),F9:r(120),F10:r(121),F11:r(122),F12:r(123),EQUALS:r(187,"=","+"),SEPARATOR:r(108,","),HYPHEN:r(189,"-","_"),COMMA:r(188,",","<"),PERIOD:r(190,".",">"),SLASH:r(191,"/","?"),BACKTICK:r(192,"`","~"),OPEN_BRACKET:r(219,"[","{"),BACKSLASH:r(220,"\\","|"),CLOSE_BRACKET:r(221,"]","}"),SEMICOLON:r(186,";",":"),APOSTROPHE:r(222,"'",'"')};var Nt=[l.ALT,l.CONTROL,l.META,l.SHIFT],bn=new Map([[l.SHIFT.code,1],[l.CONTROL.code,2],[l.ALT.code,4],[l.META.code,8]]);var P=new Map([[h.CLICK,[0,1,2,null]],[h.CONTEXTMENU,[null,null,2,null]],[h.MOUSEUP,[0,1,2,null]],[h.MOUSEOUT,[0,1,2,0]],[h.MOUSEMOVE,[0,1,2,0]]]);P.set(h.DBLCLICK,P.get(h.CLICK));P.set(h.MOUSEDOWN,P.get(h.MOUSEUP));P.set(h.MOUSEOVER,P.get(h.MOUSEOUT));var ae;function wt(){return ae||(ae=new A),ae}function vt(n,e){let t=wt();t.setElement(n),t.submitForm(e)}function _e(n){let e=re(n);if(!e)throw new b(7,"Element was not in a form, so could not submit.");vt(n,e)}var Ee="ELEMENT",fe="element-6066-11e4-a52e-4f735466cecf",de="WINDOW";function St(n){return typeof n=="object"&&n!==null||typeof n=="function"}function Rt(n){return Array.isArray(n)?!0:typeof n=="object"&&n!==null&&typeof n.length=="number"}function Ct(n){function e(t,o){if(t==null)return null;switch(typeof t){case"string":case"number":case"boolean":return t;case"function":return t.toString();case"object":break;default:return null}let s=t;if(o.includes(s))throw new b(17,"Recursive object cannot be transferred");if("nodeType"in s&&(s.nodeType===Node.ELEMENT_NODE||s.nodeType===Node.DOCUMENT_NODE)){let a=Le(s);return{[Ee]:a,[fe]:a}}if("document"in s)return{[de]:Le(s)};if(Array.isArray(t))return t.map(a=>e(a,o));if(o=[...o,s],Rt(t))return Array.from(t).map(a=>e(a,o));let i={};for(let a of Object.keys(s))i[a]=e(s[a],o);return i}return e(n,[])}function z(n,e){if(Array.isArray(n))return n.map(t=>z(t,e));if(St(n)){if(typeof n=="function")return n;let t=n;if(Ee in t)return V(t[Ee],e);if(fe in t)return V(t[fe],e);if(de in t)return V(t[de],e);let o={};for(let s of Object.keys(t))o[s]=z(t[s],e);return o}return n}function De(n){return{status:0,value:Ct(n)}}function Pe(n){let e="code"in n&&typeof n.code=="number"?n.code:13;return{status:e,value:{message:n.message,error:q(e)}}}var Ae="$wdc_",It=":wdc:";function me(n){let e=n||document,t=e[Ae];return t||(t=e[Ae]={},t.nextId=Date.now()),t.nextId||(t.nextId=Date.now()),t}var Mt=200;function xe(n,e){if("setInterval"in e)return!!e.closed;let t=e;for(;t;){if(t===n.documentElement)return!1;t.host&&t.nodeType===Node.DOCUMENT_FRAGMENT_NODE&&(t=t.host),t=t.parentNode}return!0}function _t(n){let e=me(n);for(let t of Object.keys(e)){let o=e[t];!o||typeof o!="object"||xe(n,o)&&delete e[t]}}function Le(n){let e=n.ownerDocument,t=me(e);Object.keys(t).length-1>=Mt&&_t(e||document);let o;for(let s of Object.keys(t))if(t[s]===n){o=s;break}return o||(o=It+t.nextId++,t[o]=n),o}function V(n,e){n=decodeURIComponent(n);let t=e||document,o=me(t);if(!(n in o))throw new b(10,"Element does not exist in cache");let s=o[n];if(!xe(t,s))return s;throw delete o[n],"setInterval"in s?new b(23,"Window has been closed."):new b(10,"Element is no longer attached to the DOM")}var u={NULL:"\uE000",CANCEL:"\uE001",HELP:"\uE002",BACK_SPACE:"\uE003",TAB:"\uE004",CLEAR:"\uE005",RETURN:"\uE006",ENTER:"\uE007",SHIFT:"\uE008",CONTROL:"\uE009",ALT:"\uE00A",PAUSE:"\uE00B",ESCAPE:"\uE00C",SPACE:"\uE00D",PAGE_UP:"\uE00E",PAGE_DOWN:"\uE00F",END:"\uE010",HOME:"\uE011",ARROW_LEFT:"\uE012",LEFT:"\uE012",ARROW_UP:"\uE013",UP:"\uE013",ARROW_RIGHT:"\uE014",RIGHT:"\uE014",ARROW_DOWN:"\uE015",DOWN:"\uE015",INSERT:"\uE016",DELETE:"\uE017",SEMICOLON:"\uE018",EQUALS:"\uE019",NUMPAD0:"\uE01A",NUMPAD1:"\uE01B",NUMPAD2:"\uE01C",NUMPAD3:"\uE01D",NUMPAD4:"\uE01E",NUMPAD5:"\uE01F",NUMPAD6:"\uE020",NUMPAD7:"\uE021",NUMPAD8:"\uE022",NUMPAD9:"\uE023",MULTIPLY:"\uE024",ADD:"\uE025",SEPARATOR:"\uE026",SUBTRACT:"\uE027",DECIMAL:"\uE028",DIVIDE:"\uE029",F1:"\uE031",F2:"\uE032",F3:"\uE033",F4:"\uE034",F5:"\uE035",F6:"\uE036",F7:"\uE037",F8:"\uE038",F9:"\uE039",F10:"\uE03A",F11:"\uE03B",F12:"\uE03C",COMMAND:"\uE03D",META:"\uE03D"};var oo=new Map([[u.NULL,null],[u.BACK_SPACE,l.BACKSPACE],[u.TAB,l.TAB],[u.RETURN,l.ENTER],[u.ENTER,l.ENTER],[u.SHIFT,l.SHIFT],[u.CONTROL,l.CONTROL],[u.ALT,l.ALT],[u.PAUSE,l.PAUSE],[u.ESCAPE,l.ESC],[u.SPACE,l.SPACE],[u.PAGE_UP,l.PAGE_UP],[u.PAGE_DOWN,l.PAGE_DOWN],[u.END,l.END],[u.HOME,l.HOME],[u.LEFT,l.LEFT],[u.UP,l.UP],[u.RIGHT,l.RIGHT],[u.DOWN,l.DOWN],[u.INSERT,l.INSERT],[u.DELETE,l.DELETE],[u.SEMICOLON,l.SEMICOLON],[u.EQUALS,l.EQUALS],[u.NUMPAD0,l.NUM_ZERO],[u.NUMPAD1,l.NUM_ONE],[u.NUMPAD2,l.NUM_TWO],[u.NUMPAD3,l.NUM_THREE],[u.NUMPAD4,l.NUM_FOUR],[u.NUMPAD5,l.NUM_FIVE],[u.NUMPAD6,l.NUM_SIX],[u.NUMPAD7,l.NUM_SEVEN],[u.NUMPAD8,l.NUM_EIGHT],[u.NUMPAD9,l.NUM_NINE],[u.MULTIPLY,l.NUM_MULTIPLY],[u.ADD,l.NUM_PLUS],[u.SUBTRACT,l.NUM_MINUS],[u.DECIMAL,l.NUM_PERIOD],[u.DIVIDE,l.NUM_DIVISION],[u.SEPARATOR,l.SEPARATOR],[u.F1,l.F1],[u.F2,l.F2],[u.F3,l.F3],[u.F4,l.F4],[u.F5,l.F5],[u.F6,l.F6],[u.F7,l.F7],[u.F8,l.F8],[u.F9,l.F9],[u.F10,l.F10],[u.F11,l.F11],[u.F12,l.F12],[u.META,l.META]]);function Fe(n){return n?V(n.WINDOW):window}function xt(n,e,t){let o;try{let s=Fe(t),i=z(e,s.document),a=n.apply(null,i);o=De(a)}catch(s){o=Pe(s)}return JSON.stringify(o)}function ke(n,e){return xt(_e,[n],e)}return je(Ut);})();
+function(){"use strict";var AtomExport=(()=>{var J=Object.defineProperty;var Ve=Object.getOwnPropertyDescriptor;var We=Object.getOwnPropertyNames;var Xe=Object.prototype.hasOwnProperty;var Ye=(n,e)=>{for(var t in e)J(n,t,{get:e[t],enumerable:!0})},Ge=(n,e,t,o)=>{if(e&&typeof e=="object"||typeof e=="function")for(let s of We(e))!Xe.call(n,s)&&s!==t&&J(n,s,{get:()=>e[s],enumerable:!(o=Ve(e,s))||o.enumerable});return n};var je=n=>Ge(J({},"__esModule",{value:!0}),n);var Ut={};Ye(Ut,{default:()=>ke});var $e={15:"element not selectable",11:"element not visible",31:"unknown error",30:"unknown error",24:"invalid cookie domain",29:"invalid element coordinates",12:"invalid element state",32:"invalid selector",51:"invalid selector",52:"invalid selector",17:"javascript error",405:"unsupported operation",34:"move target out of bounds",27:"no such alert",7:"no such element",8:"no such frame",23:"no such window",28:"script timeout",33:"session not created",10:"stale element reference",21:"timeout",25:"unable to set cookie",26:"unexpected alert open",13:"unknown error",9:"unknown command"};function q(n){return $e[n]??"unknown error"}var b=class extends Error{code;state;isAutomationError=!0;constructor(e,t=""){super(t),this.code=e,this.state=q(e);let o=this.state.replace(/(^|\s+)[a-z]/g,i=>i.toUpperCase().trimStart()).replace(/\s+/g,"");this.name=/Error$/.test(o)?o:`${o}Error`;let s=new Error(this.message);s.name=this.name,this.stack=s.stack||""}};var ze=/[;]+(?=(?:(?:[^"]*"){2})*[^"]*$)(?=(?:(?:[^']*'){2})*[^']*$)(?=(?:[^()]*\([^()]*\))*[^()]*$)/;function Je(n){let e=n.split(ze),t=[];for(let s of e){let i=s.indexOf(":");if(i>0){let a=[s.slice(0,i),s.slice(i+1)];t.push(a[0].toLowerCase(),":",a[1],";")}}let o=t.join("");return o.charAt(o.length-1)===";"?o:`${o};`}function Ne(n,e){if(e=e.toLowerCase(),e==="style")return Je(n.style.cssText);let t=n.getAttributeNode(e);return t&&t.specified?t.value:null}function L(n,e){return n[e]}function c(n,e){return n instanceof HTMLFormElement?n.nodeType===Node.ELEMENT_NODE&&(!e||e==="FORM"):!!n&&n.nodeType===Node.ELEMENT_NODE&&(!e||n.tagName.toUpperCase()===e)}function U(n){return n.nodeType===Node.DOCUMENT_NODE?n:n.ownerDocument}function W(n,e,t=!1){let o=t?n:n.parentNode;for(;o;){if(e(o))return o;o=o.parentNode}return null}function K(n){if(c(n,"OPTION"))return!0;if(c(n,"INPUT")){let e=n.type.toLowerCase();return e==="checkbox"||e==="radio"}return!1}function F(n){if(!K(n))throw new b(15,"Element is not selectable");let e=n.type?.toLowerCase();return!!L(n,e==="checkbox"||e==="radio"?"checked":"selected")}var ge={aliceblue:"#f0f8ff",antiquewhite:"#faebd7",aqua:"#00ffff",aquamarine:"#7fffd4",azure:"#f0ffff",beige:"#f5f5dc",bisque:"#ffe4c4",black:"#000000",blanchedalmond:"#ffebcd",blue:"#0000ff",blueviolet:"#8a2be2",brown:"#a52a2a",burlywood:"#deb887",cadetblue:"#5f9ea0",chartreuse:"#7fff00",chocolate:"#d2691e",coral:"#ff7f50",cornflowerblue:"#6495ed",cornsilk:"#fff8dc",crimson:"#dc143c",cyan:"#00ffff",darkblue:"#00008b",darkcyan:"#008b8b",darkgoldenrod:"#b8860b",darkgray:"#a9a9a9",darkgreen:"#006400",darkgrey:"#a9a9a9",darkkhaki:"#bdb76b",darkmagenta:"#8b008b",darkolivegreen:"#556b2f",darkorange:"#ff8c00",darkorchid:"#9932cc",darkred:"#8b0000",darksalmon:"#e9967a",darkseagreen:"#8fbc8f",darkslateblue:"#483d8b",darkslategray:"#2f4f4f",darkslategrey:"#2f4f4f",darkturquoise:"#00ced1",darkviolet:"#9400d3",deeppink:"#ff1493",deepskyblue:"#00bfff",dimgray:"#696969",dimgrey:"#696969",dodgerblue:"#1e90ff",firebrick:"#b22222",floralwhite:"#fffaf0",forestgreen:"#228b22",fuchsia:"#ff00ff",gainsboro:"#dcdcdc",ghostwhite:"#f8f8ff",gold:"#ffd700",goldenrod:"#daa520",gray:"#808080",green:"#008000",greenyellow:"#adff2f",grey:"#808080",honeydew:"#f0fff0",hotpink:"#ff69b4",indianred:"#cd5c5c",indigo:"#4b0082",ivory:"#fffff0",khaki:"#f0e68c",lavender:"#e6e6fa",lavenderblush:"#fff0f5",lawngreen:"#7cfc00",lemonchiffon:"#fffacd",lightblue:"#add8e6",lightcoral:"#f08080",lightcyan:"#e0ffff",lightgoldenrodyellow:"#fafad2",lightgray:"#d3d3d3",lightgreen:"#90ee90",lightgrey:"#d3d3d3",lightpink:"#ffb6c1",lightsalmon:"#ffa07a",lightseagreen:"#20b2aa",lightskyblue:"#87cefa",lightslategray:"#778899",lightslategrey:"#778899",lightsteelblue:"#b0c4de",lightyellow:"#ffffe0",lime:"#00ff00",limegreen:"#32cd32",linen:"#faf0e6",magenta:"#ff00ff",maroon:"#800000",mediumaquamarine:"#66cdaa",mediumblue:"#0000cd",mediumorchid:"#ba55d3",mediumpurple:"#9370db",mediumseagreen:"#3cb371",mediumslateblue:"#7b68ee",mediumspringgreen:"#00fa9a",mediumturquoise:"#48d1cc",mediumvioletred:"#c71585",midnightblue:"#191970",mintcream:"#f5fffa",mistyrose:"#ffe4e1",moccasin:"#ffe4b5",navajowhite:"#ffdead",navy:"#000080",oldlace:"#fdf5e6",olive:"#808000",olivedrab:"#6b8e23",orange:"#ffa500",orangered:"#ff4500",orchid:"#da70d6",palegoldenrod:"#eee8aa",palegreen:"#98fb98",paleturquoise:"#afeeee",palevioletred:"#db7093",papayawhip:"#ffefd5",peachpuff:"#ffdab9",peru:"#cd853f",pink:"#ffc0cb",plum:"#dda0dd",powderblue:"#b0e0e6",purple:"#800080",red:"#ff0000",rosybrown:"#bc8f8f",royalblue:"#4169e1",saddlebrown:"#8b4513",salmon:"#fa8072",sandybrown:"#f4a460",seagreen:"#2e8b57",seashell:"#fff5ee",sienna:"#a0522d",silver:"#c0c0c0",skyblue:"#87ceeb",slateblue:"#6a5acd",slategray:"#708090",slategrey:"#708090",snow:"#fffafa",springgreen:"#00ff7f",steelblue:"#4682b4",tan:"#d2b48c",teal:"#008080",thistle:"#d8bfd8",tomato:"#ff6347",turquoise:"#40e0d0",violet:"#ee82ee",wheat:"#f5deb3",white:"#ffffff",whitesmoke:"#f5f5f5",yellow:"#ffff00",yellowgreen:"#9acd32"};var qe=new Set(["backgroundColor","borderTopColor","borderRightColor","borderBottomColor","borderLeftColor","color","outlineColor"]),Ze=/#([0-9a-fA-F])([0-9a-fA-F])([0-9a-fA-F])/,Qe=/^#(?:[0-9a-f]{3}){1,2}$/i,et=/^(?:rgba)?\((\d{1,3}),\s?(\d{1,3}),\s?(\d{1,3}),\s?(0|1|0\.\d*)\)$/i,tt=/^(?:rgb)?\((0|[1-9]\d{0,2}),\s?(0|[1-9]\d{0,2}),\s?(0|[1-9]\d{0,2})\)$/i;function nt(n){n=n.toLowerCase();let e=ge[n];if(!e&&(e=n.charAt(0)==="#"?n:`#${n}`,e.length===4&&(e=e.replace(Ze,"#$1$1$2$2$3$3")),!Qe.test(e)))return null;let t=parseInt(e.substring(1,3),16),o=parseInt(e.substring(3,5),16),s=parseInt(e.substring(5,7),16);return[t,o,s,1]}function ot(n){let e=n.match(et);if(e){let t=Number(e[1]),o=Number(e[2]),s=Number(e[3]),i=Number(e[4]);if(t>=0&&t<=255&&o>=0&&o<=255&&s>=0&&s<=255&&i>=0&&i<=1)return[t,o,s,i]}return null}function st(n){let e=n.match(tt);if(e){let t=Number(e[1]),o=Number(e[2]),s=Number(e[3]);if(t>=0&&t<=255&&o>=0&&o<=255&&s>=0&&s<=255)return[t,o,s,1]}return null}function Oe(n,e){if(!qe.has(n))return e;let t=ot(e)||st(e)||nt(e);return t?`rgba(${t.join(", ")})`:e}function ye(n,e){if(!n)throw new b(32,"No selector specified");n=n.trim();let t;try{t=e.querySelector(n)}catch{throw new b(32,"An invalid or illegal selector was specified")}return t&&t.nodeType===Node.ELEMENT_NODE?t:null}var S=class{x;y;constructor(e=0,t=0){this.x=e,this.y=t}};var X=class{top;right;bottom;left;constructor(e,t,o,s){this.top=e,this.right=t,this.bottom=o,this.left=s}},N=class n{left;top;width;height;constructor(e,t,o,s){this.left=e,this.top=t,this.width=o,this.height=s}toBox(){let e=this.left+this.width,t=this.top+this.height;return new X(this.top,e,t,this.left)}intersects(e){return n.intersects(this,e)}static intersects(e,t){return e.left<=t.left+t.width&&t.left<=e.left+e.width&&e.top<=t.top+t.height&&t.top<=e.top+e.height}};function k(n,e,t){return Math.min(Math.max(n,e),t)}var it=typeof ShadowRoot=="function";function lt(n){return n.replace(/-([a-z])/g,(e,t)=>t.toUpperCase())}function G(n){let e="document"in n?n.document:U(n);try{let t=e?.activeElement;return t&&t.nodeName?t:null}catch{return null}}function M(n){return ee(n,!0)&&ve(n)&&!ut(n)}function ut(n){return g(n,"pointer-events")==="none"}var at=["A","AREA","BUTTON","INPUT","LABEL","SELECT","TEXTAREA"];function we(n){return at.some(e=>c(n,e))||Ne(n,"tabindex")!=null&&Number(L(n,"tabIndex"))>=0||Q(n)}var ct=["BUTTON","INPUT","OPTGROUP","OPTION","SELECT","TEXTAREA"];function ve(n){if(!ct.some(o=>c(n,o)))return!0;if(L(n,"disabled"))return!1;if(n.parentNode&&n.parentNode.nodeType===Node.ELEMENT_NODE&&c(n,"OPTGROUP")||c(n,"OPTION"))return ve(n.parentNode);let t=n;for(;t;){let o=t.parentNode;if(o&&c(o,"FIELDSET")&&L(o,"disabled")){if(!c(t,"LEGEND"))return!1;let s=t.previousElementSibling;for(;s;){if(c(s,"LEGEND"))return!1;s=s.previousElementSibling}}t=o}return!0}var Et=new Set(["text","search","tel","url","email","password","number"]);function ft(n){if(c(n,"TEXTAREA"))return!0;if(c(n,"INPUT")){let e=n.type.toLowerCase();return Et.has(e)}return Se(n)}function dt(n){return c(n,"INPUT")&&n.type.toLowerCase()==="file"}function I(n,e){return c(n,"INPUT")&&n.type.toLowerCase()===e}function Se(n){let e=n;if(e.contentEditable===void 0)return!1;if(e.isContentEditable!==void 0)return e.isContentEditable;function t(o){if(o.contentEditable==="inherit"){let s=D(o);return s?t(s):!1}return o.contentEditable==="true"}return t(e)}function Q(n){return(ft(n)||dt(n)||I(n,"range")||I(n,"date")||I(n,"month")||I(n,"week")||I(n,"time")||I(n,"datetime-local")||I(n,"color"))&&!L(n,"readOnly")}function D(n){let e=n.parentNode;for(;e&&e.nodeType!==Node.ELEMENT_NODE&&e.nodeType!==Node.DOCUMENT_NODE&&e.nodeType!==Node.DOCUMENT_FRAGMENT_NODE;)e=e.parentNode;return e&&c(e)?e:null}function mt(n,e){let t=n.ownerDocument.defaultView;if(t?.getComputedStyle){let o=t.getComputedStyle(n,null);if(o)return o[e]||o.getPropertyValue(e)||""}return""}function Re(n,e){let t=n.style,o=t[e];if(o===void 0&&typeof t.getPropertyValue=="function"&&(o=t.getPropertyValue(e)),o!=="inherit")return o!==void 0?o:null;let s=D(n);return s?Re(s,e):null}function g(n,e){let t=lt(e);(t==="float"||t==="cssFloat"||t==="styleFloat")&&(t="cssFloat");let o=mt(n,t)||Re(n,t);return o===null?null:Oe(t,o)}function Z(n,e,t){if(!c(n))throw new Error("Argument to isShown must be of type Element");if(c(n,"BODY"))return!0;if(c(n,"OPTION")||c(n,"OPTGROUP")){let E=n.parentNode;for(;E&&!c(E,"SELECT");)E=E.parentNode;return!!E&&Z(E,!0,t)}let o=Ce(n);if(o)return!!o.image&&o.rect.width>0&&o.rect.height>0&&Z(o.image,e,t);if(c(n,"INPUT")&&n.type.toLowerCase()==="hidden"||c(n,"NOSCRIPT"))return!1;let s=g(n,"visibility");if(s==="collapse"||s==="hidden"||!t(n)||!e&&Ie(n)===0)return!1;function i(E){let f=_(E);if(f.height>0&&f.width>0)return!0;if(c(E,"PATH")&&(f.height>0||f.width>0)){let p=g(E,"stroke-width");return!!p&&parseInt(p,10)>0}let y=g(E,"visibility");return y==="collapse"||y==="hidden"||!t(E)?!1:g(E,"overflow")!=="hidden"&&[...E.childNodes].some(p=>{if(p.nodeType===Node.TEXT_NODE){let w=p.nodeValue??"";return!(/^[\s]*$/.test(w)&&/[\n\r\t]/.test(w))}return c(p)&&i(p)})}if(!i(n))return!1;function a(E){return Y(E)==="hidden"&&[...E.childNodes].every(f=>!c(f)||a(f)||!i(f))}return!a(n)}function ee(n,e=!1){function t(o){if(c(o)){let i=o;if(g(i,"display")==="none"||g(i,"content-visibility")==="hidden")return!1}let s=pt(o);if(it&&s instanceof ShadowRoot){if(s.host.shadowRoot&&s.host.shadowRoot!==s)return!1;s=s.host}return s&&(s.nodeType===Node.DOCUMENT_NODE||s.nodeType===Node.DOCUMENT_FRAGMENT_NODE)?!0:s&&c(s,"DETAILS")&&!s.open&&!c(o,"SUMMARY")?!1:!!s&&t(s)}return Z(n,e,t)}function j(n){let e=n.scrollingElement||n.documentElement,t=n.defaultView;return new S(t.pageXOffset||e.scrollLeft,t.pageYOffset||e.scrollTop)}function Y(n,e){let t=te(n,e),o=U(n),s=o.documentElement,i=o.body,a=g(s,"overflow"),E=!1;function f(d,T){if(d===s)return!0;let m=g(d,"display");return!(m?.startsWith("inline")||m==="contents"||T==="absolute"&&g(d,"position")==="static")}function y(d){let T=g(d,"position");if(T==="fixed")return E=!0,d===s?null:s;let m=D(d);for(;m&&!f(m,T);)m=D(m);return m}function p(d){let T=d;if(a==="visible"){if(d===s&&i)T=i;else if(d===i)return{x:"visible",y:"visible"}}let m={x:g(T,"overflow-x")??"",y:g(T,"overflow-y")??""};return d===s&&(m.x=m.x==="visible"?"auto":m.x,m.y=m.y==="visible"?"auto":m.y),m}function w(d){return d===s?j(o):new S(d.scrollLeft,d.scrollTop)}for(let d=y(n);d;d=y(d)){let T=p(d);if(T.x==="visible"&&T.y==="visible")continue;let m=_(d);if(m.width===0||m.height===0)return"hidden";let he=t.right<m.left,pe=t.bottom<m.top;if(he&&T.x==="hidden"||pe&&T.y==="hidden")return"hidden";if(he&&T.x!=="visible"||pe&&T.y!=="visible"){let x=w(d),He=t.right<m.left-x.x,Be=t.bottom<m.top-x.y;return He&&T.x!=="visible"||Be&&T.x!=="visible"||Y(d)==="hidden"?"hidden":"scroll"}let Te=t.left>=m.left+m.width,be=t.top>=m.top+m.height;if(Te&&T.x==="hidden"||be&&T.y==="hidden")return"hidden";if(Te&&T.x!=="visible"||be&&T.y!=="visible"){if(E){let x=w(d);if(t.left>=s.scrollWidth-x.x||t.right>=s.scrollHeight-x.y)return"hidden"}return Y(d)==="hidden"?"hidden":"scroll"}}return"none"}function _(n){let e=Ce(n);if(e)return e.rect;if(c(n,"HTML")){let o=U(n),s=o.defaultView,i=o.compatMode==="CSS1Compat"?o.documentElement:o.body;return new N(0,0,i.clientWidth||s.innerWidth,i.clientHeight||s.innerHeight)}let t;try{t=n.getBoundingClientRect()}catch{return new N(0,0,0,0)}return new N(t.left,t.top,t.right-t.left,t.bottom-t.top)}function Ce(n){let e=c(n,"MAP");if(!e&&!c(n,"AREA"))return null;let t=e?n:n.parentNode&&c(n.parentNode,"MAP")?n.parentNode:null,o=null,s=null,i=t&&t.name;if(t&&i){let a=U(t),E=`*[usemap="#${i}"]`;if(o=ye(E,a),o&&(s=_(o),!e&&n.shape.toLowerCase()!=="default")){let f=ht(n),y=Math.min(Math.max(f.left,0),s.width),p=Math.min(Math.max(f.top,0),s.height),w=Math.min(f.width,s.width-y),d=Math.min(f.height,s.height-p);s=new N(y+s.left,p+s.top,w,d)}}return{image:o,rect:s||new N(0,0,0,0)}}function ht(n){let e=n,t=e.shape.toLowerCase(),o=e.coords.split(",");if(t==="rect"&&o.length===4){let s=Number(o[0]),i=Number(o[1]);return new N(s,i,Number(o[2])-s,Number(o[3])-i)}else if(t==="circle"&&o.length===3){let s=Number(o[0]),i=Number(o[1]),a=Number(o[2]);return new N(s-a,i-a,2*a,2*a)}else if(t==="poly"&&o.length>2){let s=Number(o[0]),i=Number(o[1]),a=s,E=i;for(let f=2;f+1<o.length;f+=2)s=Math.min(s,Number(o[f])),a=Math.max(a,Number(o[f])),i=Math.min(i,Number(o[f+1])),E=Math.max(E,Number(o[f+1]));return new N(s,i,a-s,E-i)}return new N(0,0,0,0)}function te(n,e){let t=_(n).toBox();if(e){let o=e instanceof N?e:new N(e.x,e.y,1,1);t.left=k(t.left+o.left,t.left,t.right),t.top=k(t.top+o.top,t.top,t.bottom),t.right=k(t.left+o.width,t.left,t.right),t.bottom=k(t.top+o.height,t.top,t.bottom)}return t}function Ie(n){let e=1,t=g(n,"opacity");t&&(e=Number(t));let o=D(n);return o&&(e*=Ie(o)),e}function pt(n){let e=n.parentNode;if(e&&e.shadowRoot&&n.assignedSlot!==void 0)return n.assignedSlot?n.assignedSlot.parentNode:null;let t=n;if(t.getDestinationInsertionPoints){let o=t.getDestinationInsertionPoints();if(o.length>0)return o[o.length-1]}return e}var O=class{type;bubbles;cancelable;constructor(e,t,o){this.type=e,this.bubbles=t,this.cancelable=o}create(e,t){let s=("ownerDocument"in e?e.ownerDocument:e.document).createEvent("HTMLEvents");return s.initEvent(this.type,this.bubbles,this.cancelable),s}toString(){return this.type}},v=class extends O{create(e,t){let o=t,s="ownerDocument"in e?e.ownerDocument:e.document,i=s.defaultView,a=s.createEvent("MouseEvents"),E=1;return this.type==="mousewheel"&&(a.wheelDelta=o.wheelDelta),a.initMouseEvent(this.type,this.bubbles,this.cancelable,i,E,o.clientX,o.clientY,o.clientX,o.clientY,o.ctrlKey,o.altKey,o.shiftKey,o.metaKey,o.button,o.relatedTarget),a}},H=class extends O{create(e,t){let o=t,i=("ownerDocument"in e?e.ownerDocument:e.document).createEvent("Events");return i.initEvent(this.type,this.bubbles,this.cancelable),i.altKey=o.altKey,i.ctrlKey=o.ctrlKey,i.metaKey=o.metaKey,i.shiftKey=o.shiftKey,i.keyCode=o.charCode||o.keyCode,i.charCode=this.type==="keypress"?i.keyCode:0,i}};function ne(n,e){return e.map(t=>new Touch({identifier:t.identifier,target:n,screenX:t.screenX,screenY:t.screenY,clientX:t.clientX,clientY:t.clientY,pageX:t.pageX,pageY:t.pageY}))}var B=class extends O{create(e,t){let o=t,s="ownerDocument"in e?e.ownerDocument:e.document,i=s.defaultView,a=typeof TouchEvent.prototype.initTouchEvent=="function"?"INIT_TOUCH_EVENT":"TOUCH_EVENT_CTOR",E=ne(e,o.changedTouches),f=o.touches===o.changedTouches?E:ne(e,o.touches),y=o.targetTouches===o.changedTouches?E:ne(e,o.targetTouches);if(a==="INIT_TOUCH_EVENT"){let p=s.createEvent("TouchEvent");return p.initTouchEvent(this.type,this.bubbles,this.cancelable,i,1,0,0,o.clientX,o.clientY,o.ctrlKey,o.altKey,o.shiftKey,o.metaKey,f,y,E,o.scale,o.rotation),p.relatedTarget=o.relatedTarget,p}return new TouchEvent(this.type,{touches:f,targetTouches:y,changedTouches:E,bubbles:this.bubbles,cancelable:this.cancelable,ctrlKey:o.ctrlKey,shiftKey:o.shiftKey,altKey:o.altKey,metaKey:o.metaKey})}},h={BLUR:new O("blur",!1,!1),CHANGE:new O("change",!0,!1),FOCUS:new O("focus",!1,!1),FOCUSIN:new O("focusin",!0,!1),FOCUSOUT:new O("focusout",!0,!1),INPUT:new O("input",!0,!1),ORIENTATIONCHANGE:new O("orientationchange",!1,!1),SELECT:new O("select",!0,!1),SUBMIT:new O("submit",!0,!0),TEXTINPUT:new O("textInput",!0,!0),CLICK:new v("click",!0,!0),CONTEXTMENU:new v("contextmenu",!0,!0),DBLCLICK:new v("dblclick",!0,!0),MOUSEDOWN:new v("mousedown",!0,!0),MOUSEMOVE:new v("mousemove",!0,!1),MOUSEOUT:new v("mouseout",!0,!0),MOUSEOVER:new v("mouseover",!0,!0),MOUSEUP:new v("mouseup",!0,!0),MOUSEWHEEL:new v("mousewheel",!0,!0),KEYDOWN:new H("keydown",!0,!0),KEYPRESS:new H("keypress",!0,!0),KEYUP:new H("keyup",!0,!0),TOUCHEND:new B("touchend",!0,!0),TOUCHMOVE:new B("touchmove",!0,!0),TOUCHSTART:new B("touchstart",!0,!0)};function R(n,e,t){let o=e.create(n,t);return"isTrusted"in o||(o.isTrusted=!1),n.dispatchEvent(o)}var oe=class{pressedModifiers=0;isPressed(e){return(this.pressedModifiers&e)!==0}setPressed(e,t){t?this.pressedModifiers|=e:this.pressedModifiers&=~e}isShiftPressed(){return this.isPressed(1)}isControlPressed(){return this.isPressed(2)}isAltPressed(){return this.isPressed(4)}isMetaPressed(){return this.isPressed(8)}},se=class{fireHtmlEvent(e,t){return R(e,t)}fireKeyboardEvent(e,t,o){return R(e,t,o)}fireMouseEvent(e,t,o){return R(e,t,o)}fireTouchEvent(e,t,o){return R(e,t,o)}};function Me(n){return c(n,"FORM")}function re(n){return W(n,Me,!0)}var A=class{el;select=null;modifiersState;eventEmitter;constructor(e,t){this.el=document.documentElement;let o=G(this.el);o&&this.setElement(o),this.modifiersState=e||new oe,this.eventEmitter=t||new se}getElement(){return this.el}setElement(e){this.el=e,c(e,"OPTION")?this.select=W(e,t=>c(t,"SELECT")):this.select=null}fireHtmlEvent(e){return this.eventEmitter.fireHtmlEvent(this.el,e)}fireKeyboardEvent(e,t){return this.eventEmitter.fireKeyboardEvent(this.el,e,t)}fireMouseEvent(e,t,o,s=null,i=0,a=!1){if(!a&&!M(this.el))return!1;if(s&&!(e===h.MOUSEOVER||e===h.MOUSEOUT))throw new b(12,`Event type does not allow related target: ${e}`);let E={clientX:t.x,clientY:t.y,button:o,altKey:this.modifiersState.isAltPressed(),ctrlKey:this.modifiersState.isControlPressed(),shiftKey:this.modifiersState.isShiftPressed(),metaKey:this.modifiersState.isMetaPressed(),wheelDelta:i,relatedTarget:s},f=this.select?this.getTargetOfOptionMouseEvent(e):this.el;return f?this.eventEmitter.fireMouseEvent(f,e,E):!0}fireTouchEvent(e,t,o,s,i){let a={touches:[],targetTouches:[],changedTouches:[],altKey:this.modifiersState.isAltPressed(),ctrlKey:this.modifiersState.isControlPressed(),shiftKey:this.modifiersState.isShiftPressed(),metaKey:this.modifiersState.isMetaPressed(),relatedTarget:null,scale:0,rotation:0,clientX:o.x,clientY:o.y},E=j(this.el.ownerDocument),f=(y,p)=>{let w={identifier:y,screenX:p.x,screenY:p.y,clientX:p.x,clientY:p.y,pageX:p.x+E.x,pageY:p.y+E.y};a.changedTouches.push(w),(e===h.TOUCHSTART||e===h.TOUCHMOVE)&&(a.touches.push(w),a.targetTouches.push(w))};return f(t,o),s!==void 0&&i!==void 0&&f(s,i),this.eventEmitter.fireTouchEvent(this.el,e,a)}getTargetOfOptionMouseEvent(e){let t=this.select;switch(e){case h.CLICK:case h.MOUSEUP:return t.multiple?this.el:this.select;default:return t.multiple?this.el:null}}clickElement(e,t,o=!1){if(!o&&!M(this.el))return;let s=!this.select&&K(this.el),i=s&&F(this.el);this.fireMouseEvent(h.CLICK,e,t,null,0,o)&&s&&this.toggleRadioButtonOrCheckbox(i)}maybeToggleOption(){if(!this.select||!M(this.el))return;let e=this.select,t=F(this.el);t&&!e.multiple||(this.el.selected=!t,e.multiple||R(e,h.CHANGE))}toggleRadioButtonOrCheckbox(e){}focusOnElement(){let e=W(this.el,o=>!!o&&c(o)&&we(o),!0);e=e||this.el;let t=G(e);return e===t?!1:(t&&typeof t.blur=="function"&&(c(t,"BODY")||t.blur()),typeof e.focus=="function"?(e.focus(),!0):!1)}submitForm(e){if(!Me(e))throw new b(12,"Element is not a form, so could not submit.");R(e,h.SUBMIT)&&(c(e.submit)?HTMLFormElement.prototype.submit.call(e):e.submit())}};function Tt(n){return navigator.userAgent.includes(n)}function ie(){return Tt("Macintosh")}var ue=class{code;character;shiftChar;constructor(e,t,o){this.code=e,this.character=t||null,this.shiftChar=o||this.character}},le=new Map;function r(n,e,t){let o=new ue(n,e,t);return e&&(!le.has(e)||t)&&(le.set(e,{key:o,shift:!1}),t&&le.set(t,{key:o,shift:!0})),o}var l={BACKSPACE:r(8),TAB:r(9),ENTER:r(13),SHIFT:r(16),CONTROL:r(17),ALT:r(18),PAUSE:r(19),CAPS_LOCK:r(20),ESC:r(27),SPACE:r(32," "),PAGE_UP:r(33),PAGE_DOWN:r(34),END:r(35),HOME:r(36),LEFT:r(37),UP:r(38),RIGHT:r(39),DOWN:r(40),PRINT_SCREEN:r(44),INSERT:r(45),DELETE:r(46),ZERO:r(48,"0",")"),ONE:r(49,"1","!"),TWO:r(50,"2","@"),THREE:r(51,"3","#"),FOUR:r(52,"4","$"),FIVE:r(53,"5","%"),SIX:r(54,"6","^"),SEVEN:r(55,"7","&"),EIGHT:r(56,"8","*"),NINE:r(57,"9","("),A:r(65,"a","A"),B:r(66,"b","B"),C:r(67,"c","C"),D:r(68,"d","D"),E:r(69,"e","E"),F:r(70,"f","F"),G:r(71,"g","G"),H:r(72,"h","H"),I:r(73,"i","I"),J:r(74,"j","J"),K:r(75,"k","K"),L:r(76,"l","L"),M:r(77,"m","M"),N:r(78,"n","N"),O:r(79,"o","O"),P:r(80,"p","P"),Q:r(81,"q","Q"),R:r(82,"r","R"),S:r(83,"s","S"),T:r(84,"t","T"),U:r(85,"u","U"),V:r(86,"v","V"),W:r(87,"w","W"),X:r(88,"x","X"),Y:r(89,"y","Y"),Z:r(90,"z","Z"),META:r(91),META_RIGHT:r(ie()?93:92),CONTEXT_MENU:r(ie()?0:null),NUM_ZERO:r(96,"0"),NUM_ONE:r(97,"1"),NUM_TWO:r(98,"2"),NUM_THREE:r(99,"3"),NUM_FOUR:r(100,"4"),NUM_FIVE:r(101,"5"),NUM_SIX:r(102,"6"),NUM_SEVEN:r(103,"7"),NUM_EIGHT:r(104,"8"),NUM_NINE:r(105,"9"),NUM_MULTIPLY:r(106,"*"),NUM_PLUS:r(107,"+"),NUM_MINUS:r(109,"-"),NUM_PERIOD:r(110,"."),NUM_DIVISION:r(111,"/"),NUM_LOCK:r(144),F1:r(112),F2:r(113),F3:r(114),F4:r(115),F5:r(116),F6:r(117),F7:r(118),F8:r(119),F9:r(120),F10:r(121),F11:r(122),F12:r(123),EQUALS:r(187,"=","+"),SEPARATOR:r(108,","),HYPHEN:r(189,"-","_"),COMMA:r(188,",","<"),PERIOD:r(190,".",">"),SLASH:r(191,"/","?"),BACKTICK:r(192,"`","~"),OPEN_BRACKET:r(219,"[","{"),BACKSLASH:r(220,"\\","|"),CLOSE_BRACKET:r(221,"]","}"),SEMICOLON:r(186,";",":"),APOSTROPHE:r(222,"'",'"')};var Nt=[l.ALT,l.CONTROL,l.META,l.SHIFT],bn=new Map([[l.SHIFT.code,1],[l.CONTROL.code,2],[l.ALT.code,4],[l.META.code,8]]);var P=new Map([[h.CLICK,[0,1,2,null]],[h.CONTEXTMENU,[null,null,2,null]],[h.MOUSEUP,[0,1,2,null]],[h.MOUSEOUT,[0,1,2,0]],[h.MOUSEMOVE,[0,1,2,0]]]);P.set(h.DBLCLICK,P.get(h.CLICK));P.set(h.MOUSEDOWN,P.get(h.MOUSEUP));P.set(h.MOUSEOVER,P.get(h.MOUSEOUT));var ae;function wt(){return ae||(ae=new A),ae}function vt(n,e){let t=wt();t.setElement(n),t.submitForm(e)}function _e(n){let e=re(n);if(!e)throw new b(7,"Element was not in a form, so could not submit.");vt(n,e)}var Ee="ELEMENT",fe="element-6066-11e4-a52e-4f735466cecf",de="WINDOW";function St(n){return typeof n=="object"&&n!==null||typeof n=="function"}function Rt(n){return Array.isArray(n)?!0:typeof n=="object"&&n!==null&&typeof n.length=="number"}function Ct(n){function e(t,o){if(t==null)return null;switch(typeof t){case"string":case"number":case"boolean":return t;case"function":return t.toString();case"object":break;default:return null}let s=t;if(o.includes(s))throw new b(17,"Recursive object cannot be transferred");if("nodeType"in s&&(s.nodeType===Node.ELEMENT_NODE||s.nodeType===Node.DOCUMENT_NODE)){let a=Le(s);return{[Ee]:a,[fe]:a}}if("document"in s)return{[de]:Le(s)};if(Array.isArray(t))return t.map(a=>e(a,o));if(o=[...o,s],Rt(t))return Array.from(t).map(a=>e(a,o));let i={};for(let a of Object.keys(s))i[a]=e(s[a],o);return i}return e(n,[])}function z(n,e){if(Array.isArray(n))return n.map(t=>z(t,e));if(St(n)){if(typeof n=="function")return n;let t=n;if(Ee in t)return V(t[Ee],e);if(fe in t)return V(t[fe],e);if(de in t)return V(t[de],e);let o={};for(let s of Object.keys(t))o[s]=z(t[s],e);return o}return n}function De(n){return{status:0,value:Ct(n)}}function Pe(n){let e="code"in n&&typeof n.code=="number"?n.code:13;return{status:e,value:{message:n.message,error:q(e)}}}var Ae="$wdc_",It=":wdc:";function me(n){let e=n||document,t=e[Ae];return t||(t=e[Ae]={},t.nextId=Date.now()),t.nextId||(t.nextId=Date.now()),t}var Mt=200;function xe(n,e){if("setInterval"in e)return!!e.closed;let t=e;for(;t;){if(t===n.documentElement)return!1;t.host&&t.nodeType===Node.DOCUMENT_FRAGMENT_NODE&&(t=t.host),t=t.parentNode}return!0}function _t(n){let e=me(n);for(let t of Object.keys(e)){let o=e[t];!o||typeof o!="object"||xe(n,o)&&delete e[t]}}function Le(n){let e=n.ownerDocument,t=me(e);Object.keys(t).length-1>=Mt&&_t(e||document);let o;for(let s of Object.keys(t))if(t[s]===n){o=s;break}return o||(o=It+t.nextId++,t[o]=n),o}function V(n,e){n=decodeURIComponent(n);let t=e||document,o=me(t);if(!(n in o))throw new b(10,"Element does not exist in cache");let s=o[n];if(!xe(t,s))return s;throw delete o[n],"setInterval"in s?new b(23,"Window has been closed."):new b(10,"Element is no longer attached to the DOM")}var u={NULL:"\uE000",CANCEL:"\uE001",HELP:"\uE002",BACK_SPACE:"\uE003",TAB:"\uE004",CLEAR:"\uE005",RETURN:"\uE006",ENTER:"\uE007",SHIFT:"\uE008",CONTROL:"\uE009",ALT:"\uE00A",PAUSE:"\uE00B",ESCAPE:"\uE00C",SPACE:"\uE00D",PAGE_UP:"\uE00E",PAGE_DOWN:"\uE00F",END:"\uE010",HOME:"\uE011",ARROW_LEFT:"\uE012",LEFT:"\uE012",ARROW_UP:"\uE013",UP:"\uE013",ARROW_RIGHT:"\uE014",RIGHT:"\uE014",ARROW_DOWN:"\uE015",DOWN:"\uE015",INSERT:"\uE016",DELETE:"\uE017",SEMICOLON:"\uE018",EQUALS:"\uE019",NUMPAD0:"\uE01A",NUMPAD1:"\uE01B",NUMPAD2:"\uE01C",NUMPAD3:"\uE01D",NUMPAD4:"\uE01E",NUMPAD5:"\uE01F",NUMPAD6:"\uE020",NUMPAD7:"\uE021",NUMPAD8:"\uE022",NUMPAD9:"\uE023",MULTIPLY:"\uE024",ADD:"\uE025",SEPARATOR:"\uE026",SUBTRACT:"\uE027",DECIMAL:"\uE028",DIVIDE:"\uE029",F1:"\uE031",F2:"\uE032",F3:"\uE033",F4:"\uE034",F5:"\uE035",F6:"\uE036",F7:"\uE037",F8:"\uE038",F9:"\uE039",F10:"\uE03A",F11:"\uE03B",F12:"\uE03C",COMMAND:"\uE03D",META:"\uE03D"};var oo=new Map([[u.NULL,null],[u.BACK_SPACE,l.BACKSPACE],[u.TAB,l.TAB],[u.RETURN,l.ENTER],[u.ENTER,l.ENTER],[u.SHIFT,l.SHIFT],[u.CONTROL,l.CONTROL],[u.ALT,l.ALT],[u.PAUSE,l.PAUSE],[u.ESCAPE,l.ESC],[u.SPACE,l.SPACE],[u.PAGE_UP,l.PAGE_UP],[u.PAGE_DOWN,l.PAGE_DOWN],[u.END,l.END],[u.HOME,l.HOME],[u.LEFT,l.LEFT],[u.UP,l.UP],[u.RIGHT,l.RIGHT],[u.DOWN,l.DOWN],[u.INSERT,l.INSERT],[u.DELETE,l.DELETE],[u.SEMICOLON,l.SEMICOLON],[u.EQUALS,l.EQUALS],[u.NUMPAD0,l.NUM_ZERO],[u.NUMPAD1,l.NUM_ONE],[u.NUMPAD2,l.NUM_TWO],[u.NUMPAD3,l.NUM_THREE],[u.NUMPAD4,l.NUM_FOUR],[u.NUMPAD5,l.NUM_FIVE],[u.NUMPAD6,l.NUM_SIX],[u.NUMPAD7,l.NUM_SEVEN],[u.NUMPAD8,l.NUM_EIGHT],[u.NUMPAD9,l.NUM_NINE],[u.MULTIPLY,l.NUM_MULTIPLY],[u.ADD,l.NUM_PLUS],[u.SUBTRACT,l.NUM_MINUS],[u.DECIMAL,l.NUM_PERIOD],[u.DIVIDE,l.NUM_DIVISION],[u.SEPARATOR,l.SEPARATOR],[u.F1,l.F1],[u.F2,l.F2],[u.F3,l.F3],[u.F4,l.F4],[u.F5,l.F5],[u.F6,l.F6],[u.F7,l.F7],[u.F8,l.F8],[u.F9,l.F9],[u.F10,l.F10],[u.F11,l.F11],[u.F12,l.F12],[u.META,l.META]]);function Fe(n){return n?V(n.WINDOW):window}function xt(n,e,t){let o;try{let s=Fe(t),i=z(e,s.document),a=n.apply(null,i);o=De(a)}catch(s){o=Pe(s)}return JSON.stringify(o)}function ke(n,e){return xt(_e,[n],e)}return je(Ut);})();
 
 return AtomExport.default.apply(null,arguments);}
