+++ atoms/type.js
@@ -1,4 +1,4 @@
-function(){"use strict";var AtomExport=(()=>{var ae=Object.defineProperty;var st=Object.getOwnPropertyDescriptor;var rt=Object.getOwnPropertyNames;var it=Object.prototype.hasOwnProperty;var lt=(n,e)=>{for(var t in e)ae(n,t,{get:e[t],enumerable:!0})},ut=(n,e,t,o)=>{if(e&&typeof e=="object"||typeof e=="function")for(let s of rt(e))!it.call(n,s)&&s!==t&&ae(n,s,{get:()=>e[s],enumerable:!(o=st(e,s))||o.enumerable});return n};var at=n=>ut(ae({},"__esModule",{value:!0}),n);var Zt={};lt(Zt,{default:()=>tt});var De={aliceblue:"#f0f8ff",antiquewhite:"#faebd7",aqua:"#00ffff",aquamarine:"#7fffd4",azure:"#f0ffff",beige:"#f5f5dc",bisque:"#ffe4c4",black:"#000000",blanchedalmond:"#ffebcd",blue:"#0000ff",blueviolet:"#8a2be2",brown:"#a52a2a",burlywood:"#deb887",cadetblue:"#5f9ea0",chartreuse:"#7fff00",chocolate:"#d2691e",coral:"#ff7f50",cornflowerblue:"#6495ed",cornsilk:"#fff8dc",crimson:"#dc143c",cyan:"#00ffff",darkblue:"#00008b",darkcyan:"#008b8b",darkgoldenrod:"#b8860b",darkgray:"#a9a9a9",darkgreen:"#006400",darkgrey:"#a9a9a9",darkkhaki:"#bdb76b",darkmagenta:"#8b008b",darkolivegreen:"#556b2f",darkorange:"#ff8c00",darkorchid:"#9932cc",darkred:"#8b0000",darksalmon:"#e9967a",darkseagreen:"#8fbc8f",darkslateblue:"#483d8b",darkslategray:"#2f4f4f",darkslategrey:"#2f4f4f",darkturquoise:"#00ced1",darkviolet:"#9400d3",deeppink:"#ff1493",deepskyblue:"#00bfff",dimgray:"#696969",dimgrey:"#696969",dodgerblue:"#1e90ff",firebrick:"#b22222",floralwhite:"#fffaf0",forestgreen:"#228b22",fuchsia:"#ff00ff",gainsboro:"#dcdcdc",ghostwhite:"#f8f8ff",gold:"#ffd700",goldenrod:"#daa520",gray:"#808080",green:"#008000",greenyellow:"#adff2f",grey:"#808080",honeydew:"#f0fff0",hotpink:"#ff69b4",indianred:"#cd5c5c",indigo:"#4b0082",ivory:"#fffff0",khaki:"#f0e68c",lavender:"#e6e6fa",lavenderblush:"#fff0f5",lawngreen:"#7cfc00",lemonchiffon:"#fffacd",lightblue:"#add8e6",lightcoral:"#f08080",lightcyan:"#e0ffff",lightgoldenrodyellow:"#fafad2",lightgray:"#d3d3d3",lightgreen:"#90ee90",lightgrey:"#d3d3d3",lightpink:"#ffb6c1",lightsalmon:"#ffa07a",lightseagreen:"#20b2aa",lightskyblue:"#87cefa",lightslategray:"#778899",lightslategrey:"#778899",lightsteelblue:"#b0c4de",lightyellow:"#ffffe0",lime:"#00ff00",limegreen:"#32cd32",linen:"#faf0e6",magenta:"#ff00ff",maroon:"#800000",mediumaquamarine:"#66cdaa",mediumblue:"#0000cd",mediumorchid:"#ba55d3",mediumpurple:"#9370db",mediumseagreen:"#3cb371",mediumslateblue:"#7b68ee",mediumspringgreen:"#00fa9a",mediumturquoise:"#48d1cc",mediumvioletred:"#c71585",midnightblue:"#191970",mintcream:"#f5fffa",mistyrose:"#ffe4e1",moccasin:"#ffe4b5",navajowhite:"#ffdead",navy:"#000080",oldlace:"#fdf5e6",olive:"#808000",olivedrab:"#6b8e23",orange:"#ffa500",orangered:"#ff4500",orchid:"#da70d6",palegoldenrod:"#eee8aa",palegreen:"#98fb98",paleturquoise:"#afeeee",palevioletred:"#db7093",papayawhip:"#ffefd5",peachpuff:"#ffdab9",peru:"#cd853f",pink:"#ffc0cb",plum:"#dda0dd",powderblue:"#b0e0e6",purple:"#800080",red:"#ff0000",rosybrown:"#bc8f8f",royalblue:"#4169e1",saddlebrown:"#8b4513",salmon:"#fa8072",sandybrown:"#f4a460",seagreen:"#2e8b57",seashell:"#fff5ee",sienna:"#a0522d",silver:"#c0c0c0",skyblue:"#87ceeb",slateblue:"#6a5acd",slategray:"#708090",slategrey:"#708090",snow:"#fffafa",springgreen:"#00ff7f",steelblue:"#4682b4",tan:"#d2b48c",teal:"#008080",thistle:"#d8bfd8",tomato:"#ff6347",turquoise:"#40e0d0",violet:"#ee82ee",wheat:"#f5deb3",white:"#ffffff",whitesmoke:"#f5f5f5",yellow:"#ffff00",yellowgreen:"#9acd32"};var ct=new Set(["backgroundColor","borderTopColor","borderRightColor","borderBottomColor","borderLeftColor","color","outlineColor"]),Et=/#([0-9a-fA-F])([0-9a-fA-F])([0-9a-fA-F])/,ft=/^#(?:[0-9a-f]{3}){1,2}$/i,dt=/^(?:rgba)?\((\d{1,3}),\s?(\d{1,3}),\s?(\d{1,3}),\s?(0|1|0\.\d*)\)$/i,mt=/^(?:rgb)?\((0|[1-9]\d{0,2}),\s?(0|[1-9]\d{0,2}),\s?(0|[1-9]\d{0,2})\)$/i;function ht(n){n=n.toLowerCase();let e=De[n];if(!e&&(e=n.charAt(0)==="#"?n:`#${n}`,e.length===4&&(e=e.replace(Et,"#$1$1$2$2$3$3")),!ft.test(e)))return null;let t=parseInt(e.substring(1,3),16),o=parseInt(e.substring(3,5),16),s=parseInt(e.substring(5,7),16);return[t,o,s,1]}function pt(n){let e=n.match(dt);if(e){let t=Number(e[1]),o=Number(e[2]),s=Number(e[3]),i=Number(e[4]);if(t>=0&&t<=255&&o>=0&&o<=255&&s>=0&&s<=255&&i>=0&&i<=1)return[t,o,s,i]}return null}function Tt(n){let e=n.match(mt);if(e){let t=Number(e[1]),o=Number(e[2]),s=Number(e[3]);if(t>=0&&t<=255&&o>=0&&o<=255&&s>=0&&s<=255)return[t,o,s,1]}return null}function Pe(n,e){if(!ct.has(n))return e;let t=pt(e)||Tt(e)||ht(e);return t?`rgba(${t.join(", ")})`:e}var bt={15:"element not selectable",11:"element not visible",31:"unknown error",30:"unknown error",24:"invalid cookie domain",29:"invalid element coordinates",12:"invalid element state",32:"invalid selector",51:"invalid selector",52:"invalid selector",17:"javascript error",405:"unsupported operation",34:"move target out of bounds",27:"no such alert",7:"no such element",8:"no such frame",23:"no such window",28:"script timeout",33:"session not created",10:"stale element reference",21:"timeout",25:"unable to set cookie",26:"unexpected alert open",13:"unknown error",9:"unknown command"};function ce(n){return bt[n]??"unknown error"}var T=class extends Error{code;state;isAutomationError=!0;constructor(e,t=""){super(t),this.code=e,this.state=ce(e);let o=this.state.replace(/(^|\s+)[a-z]/g,i=>i.toUpperCase().trimStart()).replace(/\s+/g,"");this.name=/Error$/.test(o)?o:`${o}Error`;let s=new Error(this.message);s.name=this.name,this.stack=s.stack||""}};var Nt=/[;]+(?=(?:(?:[^"]*"){2})*[^"]*$)(?=(?:(?:[^']*'){2})*[^']*$)(?=(?:[^()]*\([^()]*\))*[^()]*$)/;function gt(n){let e=n.split(Nt),t=[];for(let s of e){let i=s.indexOf(":");if(i>0){let u=[s.slice(0,i),s.slice(i+1)];t.push(u[0].toLowerCase(),":",u[1],";")}}let o=t.join("");return o.charAt(o.length-1)===";"?o:`${o};`}function xe(n,e){if(e=e.toLowerCase(),e==="style")return gt(n.style.cssText);let t=n.getAttributeNode(e);return t&&t.specified?t.value:null}function U(n,e){return n[e]}function E(n,e){return n instanceof HTMLFormElement?n.nodeType===Node.ELEMENT_NODE&&(!e||e==="FORM"):!!n&&n.nodeType===Node.ELEMENT_NODE&&(!e||n.tagName.toUpperCase()===e)}function B(n){return n.nodeType===Node.DOCUMENT_NODE?n:n.ownerDocument}function ee(n,e,t=!1){let o=t?n:n.parentNode;for(;o;){if(e(o))return o;o=o.parentNode}return null}function V(n){if(E(n,"OPTION"))return!0;if(E(n,"INPUT")){let e=n.type.toLowerCase();return e==="checkbox"||e==="radio"}return!1}function W(n){if(!V(n))throw new T(15,"Element is not selectable");let e=n.type?.toLowerCase();return!!U(n,e==="checkbox"||e==="radio"?"checked":"selected")}function Ue(n,e){if(!n)throw new T(32,"No selector specified");n=n.trim();let t;try{t=e.querySelector(n)}catch{throw new T(32,"An invalid or illegal selector was specified")}return t&&t.nodeType===Node.ELEMENT_NODE?t:null}var R=class{x;y;constructor(e=0,t=0){this.x=e,this.y=t}};var X=class{top;right;bottom;left;constructor(e,t,o,s){this.top=e,this.right=t,this.bottom=o,this.left=s}},g=class n{left;top;width;height;constructor(e,t,o,s){this.left=e,this.top=t,this.width=o,this.height=s}toBox(){let e=this.left+this.width,t=this.top+this.height;return new X(this.top,e,t,this.left)}intersects(e){return n.intersects(this,e)}static intersects(e,t){return e.left<=t.left+t.width&&t.left<=e.left+e.width&&e.top<=t.top+t.height&&t.top<=e.top+e.height}};function Y(n,e,t){return Math.min(Math.max(n,e),t)}var yt=typeof ShadowRoot=="function";function wt(n){return n.replace(/-([a-z])/g,(e,t)=>t.toUpperCase())}function te(n){let e="document"in n?n.document:B(n);try{let t=e?.activeElement;return t&&t.nodeName?t:null}catch{return null}}function I(n){return fe(n,!0)&&Fe(n)&&!vt(n)}function vt(n){return O(n,"pointer-events")==="none"}var St=["A","AREA","BUTTON","INPUT","LABEL","SELECT","TEXTAREA"];function Ke(n){return St.some(e=>E(n,e))||xe(n,"tabindex")!=null&&Number(U(n,"tabIndex"))>=0||G(n)}var Rt=["BUTTON","INPUT","OPTGROUP","OPTION","SELECT","TEXTAREA"];function Fe(n){if(!Rt.some(o=>E(n,o)))return!0;if(U(n,"disabled"))return!1;if(n.parentNode&&n.parentNode.nodeType===Node.ELEMENT_NODE&&E(n,"OPTGROUP")||E(n,"OPTION"))return Fe(n.parentNode);let t=n;for(;t;){let o=t.parentNode;if(o&&E(o,"FIELDSET")&&U(o,"disabled")){if(!E(t,"LEGEND"))return!1;let s=t.previousElementSibling;for(;s;){if(E(s,"LEGEND"))return!1;s=s.previousElementSibling}}t=o}return!0}var Ct=new Set(["text","search","tel","url","email","password","number"]);function It(n){if(E(n,"TEXTAREA"))return!0;if(E(n,"INPUT")){let e=n.type.toLowerCase();return Ct.has(e)}return ke(n)}function Mt(n){return E(n,"INPUT")&&n.type.toLowerCase()==="file"}function A(n,e){return E(n,"INPUT")&&n.type.toLowerCase()===e}function ke(n){let e=n;if(e.contentEditable===void 0)return!1;if(e.isContentEditable!==void 0)return e.isContentEditable;function t(o){if(o.contentEditable==="inherit"){let s=C(o);return s?t(s):!1}return o.contentEditable==="true"}return t(e)}function G(n){return(It(n)||Mt(n)||A(n,"range")||A(n,"date")||A(n,"month")||A(n,"week")||A(n,"time")||A(n,"datetime-local")||A(n,"color"))&&!U(n,"readOnly")}function C(n){let e=n.parentNode;for(;e&&e.nodeType!==Node.ELEMENT_NODE&&e.nodeType!==Node.DOCUMENT_NODE&&e.nodeType!==Node.DOCUMENT_FRAGMENT_NODE;)e=e.parentNode;return e&&E(e)?e:null}function _t(n,e){let t=n.ownerDocument.defaultView;if(t?.getComputedStyle){let o=t.getComputedStyle(n,null);if(o)return o[e]||o.getPropertyValue(e)||""}return""}function He(n,e){let t=n.style,o=t[e];if(o===void 0&&typeof t.getPropertyValue=="function"&&(o=t.getPropertyValue(e)),o!=="inherit")return o!==void 0?o:null;let s=C(n);return s?He(s,e):null}function O(n,e){let t=wt(e);(t==="float"||t==="cssFloat"||t==="styleFloat")&&(t="cssFloat");let o=_t(n,t)||He(n,t);return o===null?null:Pe(t,o)}function Ee(n,e,t){if(!E(n))throw new Error("Argument to isShown must be of type Element");if(E(n,"BODY"))return!0;if(E(n,"OPTION")||E(n,"OPTGROUP")){let a=n.parentNode;for(;a&&!E(a,"SELECT");)a=a.parentNode;return!!a&&Ee(a,!0,t)}let o=Be(n);if(o)return!!o.image&&o.rect.width>0&&o.rect.height>0&&Ee(o.image,e,t);if(E(n,"INPUT")&&n.type.toLowerCase()==="hidden"||E(n,"NOSCRIPT"))return!1;let s=O(n,"visibility");if(s==="collapse"||s==="hidden"||!t(n)||!e&&Ve(n)===0)return!1;function i(a){let f=M(a);if(f.height>0&&f.width>0)return!0;if(E(a,"PATH")&&(f.height>0||f.width>0)){let h=O(a,"stroke-width");return!!h&&parseInt(h,10)>0}let b=O(a,"visibility");return b==="collapse"||b==="hidden"||!t(a)?!1:O(a,"overflow")!=="hidden"&&[...a.childNodes].some(h=>{if(h.nodeType===Node.TEXT_NODE){let v=h.nodeValue??"";return!(/^[\s]*$/.test(v)&&/[\n\r\t]/.test(v))}return E(h)&&i(h)})}if(!i(n))return!1;function u(a){return L(a)==="hidden"&&[...a.childNodes].every(f=>!E(f)||u(f)||!i(f))}return!u(n)}function fe(n,e=!1){function t(o){if(E(o)){let i=o;if(O(i,"display")==="none"||O(i,"content-visibility")==="hidden")return!1}let s=Lt(o);if(yt&&s instanceof ShadowRoot){if(s.host.shadowRoot&&s.host.shadowRoot!==s)return!1;s=s.host}return s&&(s.nodeType===Node.DOCUMENT_NODE||s.nodeType===Node.DOCUMENT_FRAGMENT_NODE)?!0:s&&E(s,"DETAILS")&&!s.open&&!E(o,"SUMMARY")?!1:!!s&&t(s)}return Ee(n,e,t)}function ne(n){let e=n.scrollingElement||n.documentElement,t=n.defaultView;return new R(t.pageXOffset||e.scrollLeft,t.pageYOffset||e.scrollTop)}function L(n,e){let t=oe(n,e),o=B(n),s=o.documentElement,i=o.body,u=O(s,"overflow"),a=!1;function f(m,N){if(m===s)return!0;let p=O(m,"display");return!(p?.startsWith("inline")||p==="contents"||N==="absolute"&&O(m,"position")==="static")}function b(m){let N=O(m,"position");if(N==="fixed")return a=!0,m===s?null:s;let p=C(m);for(;p&&!f(p,N);)p=C(p);return p}function h(m){let N=m;if(u==="visible"){if(m===s&&i)N=i;else if(m===i)return{x:"visible",y:"visible"}}let p={x:O(N,"overflow-x")??"",y:O(N,"overflow-y")??""};return m===s&&(p.x=p.x==="visible"?"auto":p.x,p.y=p.y==="visible"?"auto":p.y),p}function v(m){return m===s?ne(o):new R(m.scrollLeft,m.scrollTop)}for(let m=b(n);m;m=b(m)){let N=h(m);if(N.x==="visible"&&N.y==="visible")continue;let p=M(m);if(p.width===0||p.height===0)return"hidden";let Me=t.right<p.left,_e=t.bottom<p.top;if(Me&&N.x==="hidden"||_e&&N.y==="hidden")return"hidden";if(Me&&N.x!=="visible"||_e&&N.y!=="visible"){let H=v(m),nt=t.right<p.left-H.x,ot=t.bottom<p.top-H.y;return nt&&N.x!=="visible"||ot&&N.x!=="visible"||L(m)==="hidden"?"hidden":"scroll"}let Ae=t.left>=p.left+p.width,Le=t.top>=p.top+p.height;if(Ae&&N.x==="hidden"||Le&&N.y==="hidden")return"hidden";if(Ae&&N.x!=="visible"||Le&&N.y!=="visible"){if(a){let H=v(m);if(t.left>=s.scrollWidth-H.x||t.right>=s.scrollHeight-H.y)return"hidden"}return L(m)==="hidden"?"hidden":"scroll"}}return"none"}function M(n){let e=Be(n);if(e)return e.rect;if(E(n,"HTML")){let o=B(n),s=o.defaultView,i=o.compatMode==="CSS1Compat"?o.documentElement:o.body;return new g(0,0,i.clientWidth||s.innerWidth,i.clientHeight||s.innerHeight)}let t;try{t=n.getBoundingClientRect()}catch{return new g(0,0,0,0)}return new g(t.left,t.top,t.right-t.left,t.bottom-t.top)}function Be(n){let e=E(n,"MAP");if(!e&&!E(n,"AREA"))return null;let t=e?n:n.parentNode&&E(n.parentNode,"MAP")?n.parentNode:null,o=null,s=null,i=t&&t.name;if(t&&i){let u=B(t),a=`*[usemap="#${i}"]`;if(o=Ue(a,u),o&&(s=M(o),!e&&n.shape.toLowerCase()!=="default")){let f=At(n),b=Math.min(Math.max(f.left,0),s.width),h=Math.min(Math.max(f.top,0),s.height),v=Math.min(f.width,s.width-b),m=Math.min(f.height,s.height-h);s=new g(b+s.left,h+s.top,v,m)}}return{image:o,rect:s||new g(0,0,0,0)}}function At(n){let e=n,t=e.shape.toLowerCase(),o=e.coords.split(",");if(t==="rect"&&o.length===4){let s=Number(o[0]),i=Number(o[1]);return new g(s,i,Number(o[2])-s,Number(o[3])-i)}else if(t==="circle"&&o.length===3){let s=Number(o[0]),i=Number(o[1]),u=Number(o[2]);return new g(s-u,i-u,2*u,2*u)}else if(t==="poly"&&o.length>2){let s=Number(o[0]),i=Number(o[1]),u=s,a=i;for(let f=2;f+1<o.length;f+=2)s=Math.min(s,Number(o[f])),u=Math.max(u,Number(o[f])),i=Math.min(i,Number(o[f+1])),a=Math.max(a,Number(o[f+1]));return new g(s,i,u-s,a-i)}return new g(0,0,0,0)}function oe(n,e){let t=M(n).toBox();if(e){let o=e instanceof g?e:new g(e.x,e.y,1,1);t.left=Y(t.left+o.left,t.left,t.right),t.top=Y(t.top+o.top,t.top,t.bottom),t.right=Y(t.left+o.width,t.left,t.right),t.bottom=Y(t.top+o.height,t.top,t.bottom)}return t}function Ve(n){let e=1,t=O(n,"opacity");t&&(e=Number(t));let o=C(n);return o&&(e*=Ve(o)),e}function Lt(n){let e=n.parentNode;if(e&&e.shadowRoot&&n.assignedSlot!==void 0)return n.assignedSlot?n.assignedSlot.parentNode:null;let t=n;if(t.getDestinationInsertionPoints){let o=t.getDestinationInsertionPoints();if(o.length>0)return o[o.length-1]}return e}var y=class{type;bubbles;cancelable;constructor(e,t,o){this.type=e,this.bubbles=t,this.cancelable=o}create(e,t){let s=("ownerDocument"in e?e.ownerDocument:e.document).createEvent("HTMLEvents");return s.initEvent(this.type,this.bubbles,this.cancelable),s}toString(){return this.type}},S=class extends y{create(e,t){let o=t,s="ownerDocument"in e?e.ownerDocument:e.document,i=s.defaultView,u=s.createEvent("MouseEvents"),a=1;return this.type==="mousewheel"&&(u.wheelDelta=o.wheelDelta),u.initMouseEvent(this.type,this.bubbles,this.cancelable,i,a,o.clientX,o.clientY,o.clientX,o.clientY,o.ctrlKey,o.altKey,o.shiftKey,o.metaKey,o.button,o.relatedTarget),u}},j=class extends y{create(e,t){let o=t,i=("ownerDocument"in e?e.ownerDocument:e.document).createEvent("Events");return i.initEvent(this.type,this.bubbles,this.cancelable),i.altKey=o.altKey,i.ctrlKey=o.ctrlKey,i.metaKey=o.metaKey,i.shiftKey=o.shiftKey,i.keyCode=o.charCode||o.keyCode,i.charCode=this.type==="keypress"?i.keyCode:0,i}};function de(n,e){return e.map(t=>new Touch({identifier:t.identifier,target:n,screenX:t.screenX,screenY:t.screenY,clientX:t.clientX,clientY:t.clientY,pageX:t.pageX,pageY:t.pageY}))}var $=class extends y{create(e,t){let o=t,s="ownerDocument"in e?e.ownerDocument:e.document,i=s.defaultView,u=typeof TouchEvent.prototype.initTouchEvent=="function"?"INIT_TOUCH_EVENT":"TOUCH_EVENT_CTOR",a=de(e,o.changedTouches),f=o.touches===o.changedTouches?a:de(e,o.touches),b=o.targetTouches===o.changedTouches?a:de(e,o.targetTouches);if(u==="INIT_TOUCH_EVENT"){let h=s.createEvent("TouchEvent");return h.initTouchEvent(this.type,this.bubbles,this.cancelable,i,1,0,0,o.clientX,o.clientY,o.ctrlKey,o.altKey,o.shiftKey,o.metaKey,f,b,a,o.scale,o.rotation),h.relatedTarget=o.relatedTarget,h}return new TouchEvent(this.type,{touches:f,targetTouches:b,changedTouches:a,bubbles:this.bubbles,cancelable:this.cancelable,ctrlKey:o.ctrlKey,shiftKey:o.shiftKey,altKey:o.altKey,metaKey:o.metaKey})}},d={BLUR:new y("blur",!1,!1),CHANGE:new y("change",!0,!1),FOCUS:new y("focus",!1,!1),FOCUSIN:new y("focusin",!0,!1),FOCUSOUT:new y("focusout",!0,!1),INPUT:new y("input",!0,!1),ORIENTATIONCHANGE:new y("orientationchange",!1,!1),SELECT:new y("select",!0,!1),SUBMIT:new y("submit",!0,!0),TEXTINPUT:new y("textInput",!0,!0),CLICK:new S("click",!0,!0),CONTEXTMENU:new S("contextmenu",!0,!0),DBLCLICK:new S("dblclick",!0,!0),MOUSEDOWN:new S("mousedown",!0,!0),MOUSEMOVE:new S("mousemove",!0,!1),MOUSEOUT:new S("mouseout",!0,!0),MOUSEOVER:new S("mouseover",!0,!0),MOUSEUP:new S("mouseup",!0,!0),MOUSEWHEEL:new S("mousewheel",!0,!0),KEYDOWN:new j("keydown",!0,!0),KEYPRESS:new j("keypress",!0,!0),KEYUP:new j("keyup",!0,!0),TOUCHEND:new $("touchend",!0,!0),TOUCHMOVE:new $("touchmove",!0,!0),TOUCHSTART:new $("touchstart",!0,!0)};function w(n,e,t){let o=e.create(n,t);return"isTrusted"in o||(o.isTrusted=!1),n.dispatchEvent(o)}var me=class{pressedModifiers=0;isPressed(e){return(this.pressedModifiers&e)!==0}setPressed(e,t){t?this.pressedModifiers|=e:this.pressedModifiers&=~e}isShiftPressed(){return this.isPressed(1)}isControlPressed(){return this.isPressed(2)}isAltPressed(){return this.isPressed(4)}isMetaPressed(){return this.isPressed(8)}},he=class{fireHtmlEvent(e,t){return w(e,t)}fireKeyboardEvent(e,t,o){return w(e,t,o)}fireMouseEvent(e,t,o){return w(e,t,o)}fireTouchEvent(e,t,o){return w(e,t,o)}};function We(n){return E(n,"FORM")}function pe(n){return ee(n,We,!0)}function Xe(n){if(E(n,"INPUT")){let e=n.type.toLowerCase();if(e==="submit"||e==="image")return!0}return!!(E(n,"BUTTON")&&n.type.toLowerCase()==="submit")}var D=class{el;select=null;modifiersState;eventEmitter;constructor(e,t){this.el=document.documentElement;let o=te(this.el);o&&this.setElement(o),this.modifiersState=e||new me,this.eventEmitter=t||new he}getElement(){return this.el}setElement(e){this.el=e,E(e,"OPTION")?this.select=ee(e,t=>E(t,"SELECT")):this.select=null}fireHtmlEvent(e){return this.eventEmitter.fireHtmlEvent(this.el,e)}fireKeyboardEvent(e,t){return this.eventEmitter.fireKeyboardEvent(this.el,e,t)}fireMouseEvent(e,t,o,s=null,i=0,u=!1){if(!u&&!I(this.el))return!1;if(s&&!(e===d.MOUSEOVER||e===d.MOUSEOUT))throw new T(12,`Event type does not allow related target: ${e}`);let a={clientX:t.x,clientY:t.y,button:o,altKey:this.modifiersState.isAltPressed(),ctrlKey:this.modifiersState.isControlPressed(),shiftKey:this.modifiersState.isShiftPressed(),metaKey:this.modifiersState.isMetaPressed(),wheelDelta:i,relatedTarget:s},f=this.select?this.getTargetOfOptionMouseEvent(e):this.el;return f?this.eventEmitter.fireMouseEvent(f,e,a):!0}fireTouchEvent(e,t,o,s,i){let u={touches:[],targetTouches:[],changedTouches:[],altKey:this.modifiersState.isAltPressed(),ctrlKey:this.modifiersState.isControlPressed(),shiftKey:this.modifiersState.isShiftPressed(),metaKey:this.modifiersState.isMetaPressed(),relatedTarget:null,scale:0,rotation:0,clientX:o.x,clientY:o.y},a=ne(this.el.ownerDocument),f=(b,h)=>{let v={identifier:b,screenX:h.x,screenY:h.y,clientX:h.x,clientY:h.y,pageX:h.x+a.x,pageY:h.y+a.y};u.changedTouches.push(v),(e===d.TOUCHSTART||e===d.TOUCHMOVE)&&(u.touches.push(v),u.targetTouches.push(v))};return f(t,o),s!==void 0&&i!==void 0&&f(s,i),this.eventEmitter.fireTouchEvent(this.el,e,u)}getTargetOfOptionMouseEvent(e){let t=this.select;switch(e){case d.CLICK:case d.MOUSEUP:return t.multiple?this.el:this.select;default:return t.multiple?this.el:null}}clickElement(e,t,o=!1){if(!o&&!I(this.el))return;let s=!this.select&&V(this.el),i=s&&W(this.el);this.fireMouseEvent(d.CLICK,e,t,null,0,o)&&s&&this.toggleRadioButtonOrCheckbox(i)}maybeToggleOption(){if(!this.select||!I(this.el))return;let e=this.select,t=W(this.el);t&&!e.multiple||(this.el.selected=!t,e.multiple||w(e,d.CHANGE))}toggleRadioButtonOrCheckbox(e){}focusOnElement(){let e=ee(this.el,o=>!!o&&E(o)&&Ke(o),!0);e=e||this.el;let t=te(e);return e===t?!1:(t&&typeof t.blur=="function"&&(E(t,"BODY")||t.blur()),typeof e.focus=="function"?(e.focus(),!0):!1)}submitForm(e){if(!We(e))throw new T(12,"Element is not a form, so could not submit.");w(e,d.SUBMIT)&&(E(e.submit)?HTMLFormElement.prototype.submit.call(e):e.submit())}};function Dt(n){return navigator.userAgent.includes(n)}function Te(){return Dt("Macintosh")}function P(n){try{return typeof n.selectionStart=="number"}catch{return!1}}function x(n,e){P(n)&&(n.selectionStart=e)}function z(n){return F(n)[0]}function K(n,e){P(n)&&(n.selectionEnd=e)}function be(n){return F(n)[1]}function F(n){let e=0,t=0;if(P(n)){let o=n;e=o.selectionStart,t=o.selectionEnd}return[e,t]}function J(n,e){if(P(n)){let t=n;t.selectionStart=e,t.selectionEnd=e}}function re(n,e){if(!P(n))throw new Error("Cannot set the selection end");let t=n,o=t.value,s=t.selectionStart,i=o.slice(0,s),u=o.slice(t.selectionEnd);t.value=i+e+u,t.selectionStart=s,t.selectionEnd=s+e.length}function ie(n){if(!P(n))throw new Error("Element does not support selection")}function Ne(n){return P(n)}var Oe=class{code;character;shiftChar;constructor(e,t,o){this.code=e,this.character=t||null,this.shiftChar=o||this.character}},le=new Map;function r(n,e,t){let o=new Oe(n,e,t);return e&&(!le.has(e)||t)&&(le.set(e,{key:o,shift:!1}),t&&le.set(t,{key:o,shift:!0})),o}var l={BACKSPACE:r(8),TAB:r(9),ENTER:r(13),SHIFT:r(16),CONTROL:r(17),ALT:r(18),PAUSE:r(19),CAPS_LOCK:r(20),ESC:r(27),SPACE:r(32," "),PAGE_UP:r(33),PAGE_DOWN:r(34),END:r(35),HOME:r(36),LEFT:r(37),UP:r(38),RIGHT:r(39),DOWN:r(40),PRINT_SCREEN:r(44),INSERT:r(45),DELETE:r(46),ZERO:r(48,"0",")"),ONE:r(49,"1","!"),TWO:r(50,"2","@"),THREE:r(51,"3","#"),FOUR:r(52,"4","$"),FIVE:r(53,"5","%"),SIX:r(54,"6","^"),SEVEN:r(55,"7","&"),EIGHT:r(56,"8","*"),NINE:r(57,"9","("),A:r(65,"a","A"),B:r(66,"b","B"),C:r(67,"c","C"),D:r(68,"d","D"),E:r(69,"e","E"),F:r(70,"f","F"),G:r(71,"g","G"),H:r(72,"h","H"),I:r(73,"i","I"),J:r(74,"j","J"),K:r(75,"k","K"),L:r(76,"l","L"),M:r(77,"m","M"),N:r(78,"n","N"),O:r(79,"o","O"),P:r(80,"p","P"),Q:r(81,"q","Q"),R:r(82,"r","R"),S:r(83,"s","S"),T:r(84,"t","T"),U:r(85,"u","U"),V:r(86,"v","V"),W:r(87,"w","W"),X:r(88,"x","X"),Y:r(89,"y","Y"),Z:r(90,"z","Z"),META:r(91),META_RIGHT:r(Te()?93:92),CONTEXT_MENU:r(Te()?0:null),NUM_ZERO:r(96,"0"),NUM_ONE:r(97,"1"),NUM_TWO:r(98,"2"),NUM_THREE:r(99,"3"),NUM_FOUR:r(100,"4"),NUM_FIVE:r(101,"5"),NUM_SIX:r(102,"6"),NUM_SEVEN:r(103,"7"),NUM_EIGHT:r(104,"8"),NUM_NINE:r(105,"9"),NUM_MULTIPLY:r(106,"*"),NUM_PLUS:r(107,"+"),NUM_MINUS:r(109,"-"),NUM_PERIOD:r(110,"."),NUM_DIVISION:r(111,"/"),NUM_LOCK:r(144),F1:r(112),F2:r(113),F3:r(114),F4:r(115),F5:r(116),F6:r(117),F7:r(118),F8:r(119),F9:r(120),F10:r(121),F11:r(122),F12:r(123),EQUALS:r(187,"=","+"),SEPARATOR:r(108,","),HYPHEN:r(189,"-","_"),COMMA:r(188,",","<"),PERIOD:r(190,".",">"),SLASH:r(191,"/","?"),BACKTICK:r(192,"`","~"),OPEN_BRACKET:r(219,"[","{"),BACKSLASH:r(220,"\\","|"),CLOSE_BRACKET:r(221,"]","}"),SEMICOLON:r(186,";",":"),APOSTROPHE:r(222,"'",'"')};function Ye(n){if(n.length!==1)throw new T(13,`Argument not a single character: ${n}`);let e=le.get(n);if(!e){let t=n.toUpperCase(),o=t.charCodeAt(0),s=r(o,n.toLowerCase(),t);e={key:s,shift:n!==s.character}}return e}var q=[l.ALT,l.CONTROL,l.META,l.SHIFT],xt=new Map([[l.SHIFT.code,1],[l.CONTROL.code,2],[l.ALT.code,4],[l.META.code,8]]),ge=`
+function(){"use strict";var AtomExport=(()=>{var ae=Object.defineProperty;var st=Object.getOwnPropertyDescriptor;var rt=Object.getOwnPropertyNames;var it=Object.prototype.hasOwnProperty;var lt=(n,e)=>{for(var t in e)ae(n,t,{get:e[t],enumerable:!0})},ut=(n,e,t,o)=>{if(e&&typeof e=="object"||typeof e=="function")for(let s of rt(e))!it.call(n,s)&&s!==t&&ae(n,s,{get:()=>e[s],enumerable:!(o=st(e,s))||o.enumerable});return n};var at=n=>ut(ae({},"__esModule",{value:!0}),n);var Zt={};lt(Zt,{default:()=>tt});var ct={15:"element not selectable",11:"element not visible",31:"unknown error",30:"unknown error",24:"invalid cookie domain",29:"invalid element coordinates",12:"invalid element state",32:"invalid selector",51:"invalid selector",52:"invalid selector",17:"javascript error",405:"unsupported operation",34:"move target out of bounds",27:"no such alert",7:"no such element",8:"no such frame",23:"no such window",28:"script timeout",33:"session not created",10:"stale element reference",21:"timeout",25:"unable to set cookie",26:"unexpected alert open",13:"unknown error",9:"unknown command"};function ce(n){return ct[n]??"unknown error"}var T=class extends Error{code;state;isAutomationError=!0;constructor(e,t=""){super(t),this.code=e,this.state=ce(e);let o=this.state.replace(/(^|\s+)[a-z]/g,i=>i.toUpperCase().trimStart()).replace(/\s+/g,"");this.name=/Error$/.test(o)?o:`${o}Error`;let s=new Error(this.message);s.name=this.name,this.stack=s.stack||""}};var Et=/[;]+(?=(?:(?:[^"]*"){2})*[^"]*$)(?=(?:(?:[^']*'){2})*[^']*$)(?=(?:[^()]*\([^()]*\))*[^()]*$)/;function ft(n){let e=n.split(Et),t=[];for(let s of e){let i=s.indexOf(":");if(i>0){let u=[s.slice(0,i),s.slice(i+1)];t.push(u[0].toLowerCase(),":",u[1],";")}}let o=t.join("");return o.charAt(o.length-1)===";"?o:`${o};`}function De(n,e){if(e=e.toLowerCase(),e==="style")return ft(n.style.cssText);let t=n.getAttributeNode(e);return t&&t.specified?t.value:null}function U(n,e){return n[e]}function E(n,e){return n instanceof HTMLFormElement?n.nodeType===Node.ELEMENT_NODE&&(!e||e==="FORM"):!!n&&n.nodeType===Node.ELEMENT_NODE&&(!e||n.tagName.toUpperCase()===e)}function B(n){return n.nodeType===Node.DOCUMENT_NODE?n:n.ownerDocument}function ee(n,e,t=!1){let o=t?n:n.parentNode;for(;o;){if(e(o))return o;o=o.parentNode}return null}function V(n){if(E(n,"OPTION"))return!0;if(E(n,"INPUT")){let e=n.type.toLowerCase();return e==="checkbox"||e==="radio"}return!1}function W(n){if(!V(n))throw new T(15,"Element is not selectable");let e=n.type?.toLowerCase();return!!U(n,e==="checkbox"||e==="radio"?"checked":"selected")}var Pe={aliceblue:"#f0f8ff",antiquewhite:"#faebd7",aqua:"#00ffff",aquamarine:"#7fffd4",azure:"#f0ffff",beige:"#f5f5dc",bisque:"#ffe4c4",black:"#000000",blanchedalmond:"#ffebcd",blue:"#0000ff",blueviolet:"#8a2be2",brown:"#a52a2a",burlywood:"#deb887",cadetblue:"#5f9ea0",chartreuse:"#7fff00",chocolate:"#d2691e",coral:"#ff7f50",cornflowerblue:"#6495ed",cornsilk:"#fff8dc",crimson:"#dc143c",cyan:"#00ffff",darkblue:"#00008b",darkcyan:"#008b8b",darkgoldenrod:"#b8860b",darkgray:"#a9a9a9",darkgreen:"#006400",darkgrey:"#a9a9a9",darkkhaki:"#bdb76b",darkmagenta:"#8b008b",darkolivegreen:"#556b2f",darkorange:"#ff8c00",darkorchid:"#9932cc",darkred:"#8b0000",darksalmon:"#e9967a",darkseagreen:"#8fbc8f",darkslateblue:"#483d8b",darkslategray:"#2f4f4f",darkslategrey:"#2f4f4f",darkturquoise:"#00ced1",darkviolet:"#9400d3",deeppink:"#ff1493",deepskyblue:"#00bfff",dimgray:"#696969",dimgrey:"#696969",dodgerblue:"#1e90ff",firebrick:"#b22222",floralwhite:"#fffaf0",forestgreen:"#228b22",fuchsia:"#ff00ff",gainsboro:"#dcdcdc",ghostwhite:"#f8f8ff",gold:"#ffd700",goldenrod:"#daa520",gray:"#808080",green:"#008000",greenyellow:"#adff2f",grey:"#808080",honeydew:"#f0fff0",hotpink:"#ff69b4",indianred:"#cd5c5c",indigo:"#4b0082",ivory:"#fffff0",khaki:"#f0e68c",lavender:"#e6e6fa",lavenderblush:"#fff0f5",lawngreen:"#7cfc00",lemonchiffon:"#fffacd",lightblue:"#add8e6",lightcoral:"#f08080",lightcyan:"#e0ffff",lightgoldenrodyellow:"#fafad2",lightgray:"#d3d3d3",lightgreen:"#90ee90",lightgrey:"#d3d3d3",lightpink:"#ffb6c1",lightsalmon:"#ffa07a",lightseagreen:"#20b2aa",lightskyblue:"#87cefa",lightslategray:"#778899",lightslategrey:"#778899",lightsteelblue:"#b0c4de",lightyellow:"#ffffe0",lime:"#00ff00",limegreen:"#32cd32",linen:"#faf0e6",magenta:"#ff00ff",maroon:"#800000",mediumaquamarine:"#66cdaa",mediumblue:"#0000cd",mediumorchid:"#ba55d3",mediumpurple:"#9370db",mediumseagreen:"#3cb371",mediumslateblue:"#7b68ee",mediumspringgreen:"#00fa9a",mediumturquoise:"#48d1cc",mediumvioletred:"#c71585",midnightblue:"#191970",mintcream:"#f5fffa",mistyrose:"#ffe4e1",moccasin:"#ffe4b5",navajowhite:"#ffdead",navy:"#000080",oldlace:"#fdf5e6",olive:"#808000",olivedrab:"#6b8e23",orange:"#ffa500",orangered:"#ff4500",orchid:"#da70d6",palegoldenrod:"#eee8aa",palegreen:"#98fb98",paleturquoise:"#afeeee",palevioletred:"#db7093",papayawhip:"#ffefd5",peachpuff:"#ffdab9",peru:"#cd853f",pink:"#ffc0cb",plum:"#dda0dd",powderblue:"#b0e0e6",purple:"#800080",red:"#ff0000",rosybrown:"#bc8f8f",royalblue:"#4169e1",saddlebrown:"#8b4513",salmon:"#fa8072",sandybrown:"#f4a460",seagreen:"#2e8b57",seashell:"#fff5ee",sienna:"#a0522d",silver:"#c0c0c0",skyblue:"#87ceeb",slateblue:"#6a5acd",slategray:"#708090",slategrey:"#708090",snow:"#fffafa",springgreen:"#00ff7f",steelblue:"#4682b4",tan:"#d2b48c",teal:"#008080",thistle:"#d8bfd8",tomato:"#ff6347",turquoise:"#40e0d0",violet:"#ee82ee",wheat:"#f5deb3",white:"#ffffff",whitesmoke:"#f5f5f5",yellow:"#ffff00",yellowgreen:"#9acd32"};var dt=new Set(["backgroundColor","borderTopColor","borderRightColor","borderBottomColor","borderLeftColor","color","outlineColor"]),mt=/#([0-9a-fA-F])([0-9a-fA-F])([0-9a-fA-F])/,ht=/^#(?:[0-9a-f]{3}){1,2}$/i,pt=/^(?:rgba)?\((\d{1,3}),\s?(\d{1,3}),\s?(\d{1,3}),\s?(0|1|0\.\d*)\)$/i,Tt=/^(?:rgb)?\((0|[1-9]\d{0,2}),\s?(0|[1-9]\d{0,2}),\s?(0|[1-9]\d{0,2})\)$/i;function bt(n){n=n.toLowerCase();let e=Pe[n];if(!e&&(e=n.charAt(0)==="#"?n:`#${n}`,e.length===4&&(e=e.replace(mt,"#$1$1$2$2$3$3")),!ht.test(e)))return null;let t=parseInt(e.substring(1,3),16),o=parseInt(e.substring(3,5),16),s=parseInt(e.substring(5,7),16);return[t,o,s,1]}function Nt(n){let e=n.match(pt);if(e){let t=Number(e[1]),o=Number(e[2]),s=Number(e[3]),i=Number(e[4]);if(t>=0&&t<=255&&o>=0&&o<=255&&s>=0&&s<=255&&i>=0&&i<=1)return[t,o,s,i]}return null}function gt(n){let e=n.match(Tt);if(e){let t=Number(e[1]),o=Number(e[2]),s=Number(e[3]);if(t>=0&&t<=255&&o>=0&&o<=255&&s>=0&&s<=255)return[t,o,s,1]}return null}function xe(n,e){if(!dt.has(n))return e;let t=Nt(e)||gt(e)||bt(e);return t?`rgba(${t.join(", ")})`:e}function Ue(n,e){if(!n)throw new T(32,"No selector specified");n=n.trim();let t;try{t=e.querySelector(n)}catch{throw new T(32,"An invalid or illegal selector was specified")}return t&&t.nodeType===Node.ELEMENT_NODE?t:null}var R=class{x;y;constructor(e=0,t=0){this.x=e,this.y=t}};var X=class{top;right;bottom;left;constructor(e,t,o,s){this.top=e,this.right=t,this.bottom=o,this.left=s}},g=class n{left;top;width;height;constructor(e,t,o,s){this.left=e,this.top=t,this.width=o,this.height=s}toBox(){let e=this.left+this.width,t=this.top+this.height;return new X(this.top,e,t,this.left)}intersects(e){return n.intersects(this,e)}static intersects(e,t){return e.left<=t.left+t.width&&t.left<=e.left+e.width&&e.top<=t.top+t.height&&t.top<=e.top+e.height}};function Y(n,e,t){return Math.min(Math.max(n,e),t)}var yt=typeof ShadowRoot=="function";function wt(n){return n.replace(/-([a-z])/g,(e,t)=>t.toUpperCase())}function te(n){let e="document"in n?n.document:B(n);try{let t=e?.activeElement;return t&&t.nodeName?t:null}catch{return null}}function I(n){return fe(n,!0)&&Fe(n)&&!vt(n)}function vt(n){return O(n,"pointer-events")==="none"}var St=["A","AREA","BUTTON","INPUT","LABEL","SELECT","TEXTAREA"];function Ke(n){return St.some(e=>E(n,e))||De(n,"tabindex")!=null&&Number(U(n,"tabIndex"))>=0||G(n)}var Rt=["BUTTON","INPUT","OPTGROUP","OPTION","SELECT","TEXTAREA"];function Fe(n){if(!Rt.some(o=>E(n,o)))return!0;if(U(n,"disabled"))return!1;if(n.parentNode&&n.parentNode.nodeType===Node.ELEMENT_NODE&&E(n,"OPTGROUP")||E(n,"OPTION"))return Fe(n.parentNode);let t=n;for(;t;){let o=t.parentNode;if(o&&E(o,"FIELDSET")&&U(o,"disabled")){if(!E(t,"LEGEND"))return!1;let s=t.previousElementSibling;for(;s;){if(E(s,"LEGEND"))return!1;s=s.previousElementSibling}}t=o}return!0}var Ct=new Set(["text","search","tel","url","email","password","number"]);function It(n){if(E(n,"TEXTAREA"))return!0;if(E(n,"INPUT")){let e=n.type.toLowerCase();return Ct.has(e)}return ke(n)}function Mt(n){return E(n,"INPUT")&&n.type.toLowerCase()==="file"}function A(n,e){return E(n,"INPUT")&&n.type.toLowerCase()===e}function ke(n){let e=n;if(e.contentEditable===void 0)return!1;if(e.isContentEditable!==void 0)return e.isContentEditable;function t(o){if(o.contentEditable==="inherit"){let s=C(o);return s?t(s):!1}return o.contentEditable==="true"}return t(e)}function G(n){return(It(n)||Mt(n)||A(n,"range")||A(n,"date")||A(n,"month")||A(n,"week")||A(n,"time")||A(n,"datetime-local")||A(n,"color"))&&!U(n,"readOnly")}function C(n){let e=n.parentNode;for(;e&&e.nodeType!==Node.ELEMENT_NODE&&e.nodeType!==Node.DOCUMENT_NODE&&e.nodeType!==Node.DOCUMENT_FRAGMENT_NODE;)e=e.parentNode;return e&&E(e)?e:null}function _t(n,e){let t=n.ownerDocument.defaultView;if(t?.getComputedStyle){let o=t.getComputedStyle(n,null);if(o)return o[e]||o.getPropertyValue(e)||""}return""}function He(n,e){let t=n.style,o=t[e];if(o===void 0&&typeof t.getPropertyValue=="function"&&(o=t.getPropertyValue(e)),o!=="inherit")return o!==void 0?o:null;let s=C(n);return s?He(s,e):null}function O(n,e){let t=wt(e);(t==="float"||t==="cssFloat"||t==="styleFloat")&&(t="cssFloat");let o=_t(n,t)||He(n,t);return o===null?null:xe(t,o)}function Ee(n,e,t){if(!E(n))throw new Error("Argument to isShown must be of type Element");if(E(n,"BODY"))return!0;if(E(n,"OPTION")||E(n,"OPTGROUP")){let a=n.parentNode;for(;a&&!E(a,"SELECT");)a=a.parentNode;return!!a&&Ee(a,!0,t)}let o=Be(n);if(o)return!!o.image&&o.rect.width>0&&o.rect.height>0&&Ee(o.image,e,t);if(E(n,"INPUT")&&n.type.toLowerCase()==="hidden"||E(n,"NOSCRIPT"))return!1;let s=O(n,"visibility");if(s==="collapse"||s==="hidden"||!t(n)||!e&&Ve(n)===0)return!1;function i(a){let f=M(a);if(f.height>0&&f.width>0)return!0;if(E(a,"PATH")&&(f.height>0||f.width>0)){let h=O(a,"stroke-width");return!!h&&parseInt(h,10)>0}let b=O(a,"visibility");return b==="collapse"||b==="hidden"||!t(a)?!1:O(a,"overflow")!=="hidden"&&[...a.childNodes].some(h=>{if(h.nodeType===Node.TEXT_NODE){let v=h.nodeValue??"";return!(/^[\s]*$/.test(v)&&/[\n\r\t]/.test(v))}return E(h)&&i(h)})}if(!i(n))return!1;function u(a){return L(a)==="hidden"&&[...a.childNodes].every(f=>!E(f)||u(f)||!i(f))}return!u(n)}function fe(n,e=!1){function t(o){if(E(o)){let i=o;if(O(i,"display")==="none"||O(i,"content-visibility")==="hidden")return!1}let s=Lt(o);if(yt&&s instanceof ShadowRoot){if(s.host.shadowRoot&&s.host.shadowRoot!==s)return!1;s=s.host}return s&&(s.nodeType===Node.DOCUMENT_NODE||s.nodeType===Node.DOCUMENT_FRAGMENT_NODE)?!0:s&&E(s,"DETAILS")&&!s.open&&!E(o,"SUMMARY")?!1:!!s&&t(s)}return Ee(n,e,t)}function ne(n){let e=n.scrollingElement||n.documentElement,t=n.defaultView;return new R(t.pageXOffset||e.scrollLeft,t.pageYOffset||e.scrollTop)}function L(n,e){let t=oe(n,e),o=B(n),s=o.documentElement,i=o.body,u=O(s,"overflow"),a=!1;function f(m,N){if(m===s)return!0;let p=O(m,"display");return!(p?.startsWith("inline")||p==="contents"||N==="absolute"&&O(m,"position")==="static")}function b(m){let N=O(m,"position");if(N==="fixed")return a=!0,m===s?null:s;let p=C(m);for(;p&&!f(p,N);)p=C(p);return p}function h(m){let N=m;if(u==="visible"){if(m===s&&i)N=i;else if(m===i)return{x:"visible",y:"visible"}}let p={x:O(N,"overflow-x")??"",y:O(N,"overflow-y")??""};return m===s&&(p.x=p.x==="visible"?"auto":p.x,p.y=p.y==="visible"?"auto":p.y),p}function v(m){return m===s?ne(o):new R(m.scrollLeft,m.scrollTop)}for(let m=b(n);m;m=b(m)){let N=h(m);if(N.x==="visible"&&N.y==="visible")continue;let p=M(m);if(p.width===0||p.height===0)return"hidden";let Me=t.right<p.left,_e=t.bottom<p.top;if(Me&&N.x==="hidden"||_e&&N.y==="hidden")return"hidden";if(Me&&N.x!=="visible"||_e&&N.y!=="visible"){let H=v(m),nt=t.right<p.left-H.x,ot=t.bottom<p.top-H.y;return nt&&N.x!=="visible"||ot&&N.x!=="visible"||L(m)==="hidden"?"hidden":"scroll"}let Ae=t.left>=p.left+p.width,Le=t.top>=p.top+p.height;if(Ae&&N.x==="hidden"||Le&&N.y==="hidden")return"hidden";if(Ae&&N.x!=="visible"||Le&&N.y!=="visible"){if(a){let H=v(m);if(t.left>=s.scrollWidth-H.x||t.right>=s.scrollHeight-H.y)return"hidden"}return L(m)==="hidden"?"hidden":"scroll"}}return"none"}function M(n){let e=Be(n);if(e)return e.rect;if(E(n,"HTML")){let o=B(n),s=o.defaultView,i=o.compatMode==="CSS1Compat"?o.documentElement:o.body;return new g(0,0,i.clientWidth||s.innerWidth,i.clientHeight||s.innerHeight)}let t;try{t=n.getBoundingClientRect()}catch{return new g(0,0,0,0)}return new g(t.left,t.top,t.right-t.left,t.bottom-t.top)}function Be(n){let e=E(n,"MAP");if(!e&&!E(n,"AREA"))return null;let t=e?n:n.parentNode&&E(n.parentNode,"MAP")?n.parentNode:null,o=null,s=null,i=t&&t.name;if(t&&i){let u=B(t),a=`*[usemap="#${i}"]`;if(o=Ue(a,u),o&&(s=M(o),!e&&n.shape.toLowerCase()!=="default")){let f=At(n),b=Math.min(Math.max(f.left,0),s.width),h=Math.min(Math.max(f.top,0),s.height),v=Math.min(f.width,s.width-b),m=Math.min(f.height,s.height-h);s=new g(b+s.left,h+s.top,v,m)}}return{image:o,rect:s||new g(0,0,0,0)}}function At(n){let e=n,t=e.shape.toLowerCase(),o=e.coords.split(",");if(t==="rect"&&o.length===4){let s=Number(o[0]),i=Number(o[1]);return new g(s,i,Number(o[2])-s,Number(o[3])-i)}else if(t==="circle"&&o.length===3){let s=Number(o[0]),i=Number(o[1]),u=Number(o[2]);return new g(s-u,i-u,2*u,2*u)}else if(t==="poly"&&o.length>2){let s=Number(o[0]),i=Number(o[1]),u=s,a=i;for(let f=2;f+1<o.length;f+=2)s=Math.min(s,Number(o[f])),u=Math.max(u,Number(o[f])),i=Math.min(i,Number(o[f+1])),a=Math.max(a,Number(o[f+1]));return new g(s,i,u-s,a-i)}return new g(0,0,0,0)}function oe(n,e){let t=M(n).toBox();if(e){let o=e instanceof g?e:new g(e.x,e.y,1,1);t.left=Y(t.left+o.left,t.left,t.right),t.top=Y(t.top+o.top,t.top,t.bottom),t.right=Y(t.left+o.width,t.left,t.right),t.bottom=Y(t.top+o.height,t.top,t.bottom)}return t}function Ve(n){let e=1,t=O(n,"opacity");t&&(e=Number(t));let o=C(n);return o&&(e*=Ve(o)),e}function Lt(n){let e=n.parentNode;if(e&&e.shadowRoot&&n.assignedSlot!==void 0)return n.assignedSlot?n.assignedSlot.parentNode:null;let t=n;if(t.getDestinationInsertionPoints){let o=t.getDestinationInsertionPoints();if(o.length>0)return o[o.length-1]}return e}var y=class{type;bubbles;cancelable;constructor(e,t,o){this.type=e,this.bubbles=t,this.cancelable=o}create(e,t){let s=("ownerDocument"in e?e.ownerDocument:e.document).createEvent("HTMLEvents");return s.initEvent(this.type,this.bubbles,this.cancelable),s}toString(){return this.type}},S=class extends y{create(e,t){let o=t,s="ownerDocument"in e?e.ownerDocument:e.document,i=s.defaultView,u=s.createEvent("MouseEvents"),a=1;return this.type==="mousewheel"&&(u.wheelDelta=o.wheelDelta),u.initMouseEvent(this.type,this.bubbles,this.cancelable,i,a,o.clientX,o.clientY,o.clientX,o.clientY,o.ctrlKey,o.altKey,o.shiftKey,o.metaKey,o.button,o.relatedTarget),u}},j=class extends y{create(e,t){let o=t,i=("ownerDocument"in e?e.ownerDocument:e.document).createEvent("Events");return i.initEvent(this.type,this.bubbles,this.cancelable),i.altKey=o.altKey,i.ctrlKey=o.ctrlKey,i.metaKey=o.metaKey,i.shiftKey=o.shiftKey,i.keyCode=o.charCode||o.keyCode,i.charCode=this.type==="keypress"?i.keyCode:0,i}};function de(n,e){return e.map(t=>new Touch({identifier:t.identifier,target:n,screenX:t.screenX,screenY:t.screenY,clientX:t.clientX,clientY:t.clientY,pageX:t.pageX,pageY:t.pageY}))}var $=class extends y{create(e,t){let o=t,s="ownerDocument"in e?e.ownerDocument:e.document,i=s.defaultView,u=typeof TouchEvent.prototype.initTouchEvent=="function"?"INIT_TOUCH_EVENT":"TOUCH_EVENT_CTOR",a=de(e,o.changedTouches),f=o.touches===o.changedTouches?a:de(e,o.touches),b=o.targetTouches===o.changedTouches?a:de(e,o.targetTouches);if(u==="INIT_TOUCH_EVENT"){let h=s.createEvent("TouchEvent");return h.initTouchEvent(this.type,this.bubbles,this.cancelable,i,1,0,0,o.clientX,o.clientY,o.ctrlKey,o.altKey,o.shiftKey,o.metaKey,f,b,a,o.scale,o.rotation),h.relatedTarget=o.relatedTarget,h}return new TouchEvent(this.type,{touches:f,targetTouches:b,changedTouches:a,bubbles:this.bubbles,cancelable:this.cancelable,ctrlKey:o.ctrlKey,shiftKey:o.shiftKey,altKey:o.altKey,metaKey:o.metaKey})}},d={BLUR:new y("blur",!1,!1),CHANGE:new y("change",!0,!1),FOCUS:new y("focus",!1,!1),FOCUSIN:new y("focusin",!0,!1),FOCUSOUT:new y("focusout",!0,!1),INPUT:new y("input",!0,!1),ORIENTATIONCHANGE:new y("orientationchange",!1,!1),SELECT:new y("select",!0,!1),SUBMIT:new y("submit",!0,!0),TEXTINPUT:new y("textInput",!0,!0),CLICK:new S("click",!0,!0),CONTEXTMENU:new S("contextmenu",!0,!0),DBLCLICK:new S("dblclick",!0,!0),MOUSEDOWN:new S("mousedown",!0,!0),MOUSEMOVE:new S("mousemove",!0,!1),MOUSEOUT:new S("mouseout",!0,!0),MOUSEOVER:new S("mouseover",!0,!0),MOUSEUP:new S("mouseup",!0,!0),MOUSEWHEEL:new S("mousewheel",!0,!0),KEYDOWN:new j("keydown",!0,!0),KEYPRESS:new j("keypress",!0,!0),KEYUP:new j("keyup",!0,!0),TOUCHEND:new $("touchend",!0,!0),TOUCHMOVE:new $("touchmove",!0,!0),TOUCHSTART:new $("touchstart",!0,!0)};function w(n,e,t){let o=e.create(n,t);return"isTrusted"in o||(o.isTrusted=!1),n.dispatchEvent(o)}var me=class{pressedModifiers=0;isPressed(e){return(this.pressedModifiers&e)!==0}setPressed(e,t){t?this.pressedModifiers|=e:this.pressedModifiers&=~e}isShiftPressed(){return this.isPressed(1)}isControlPressed(){return this.isPressed(2)}isAltPressed(){return this.isPressed(4)}isMetaPressed(){return this.isPressed(8)}},he=class{fireHtmlEvent(e,t){return w(e,t)}fireKeyboardEvent(e,t,o){return w(e,t,o)}fireMouseEvent(e,t,o){return w(e,t,o)}fireTouchEvent(e,t,o){return w(e,t,o)}};function We(n){return E(n,"FORM")}function pe(n){return ee(n,We,!0)}function Xe(n){if(E(n,"INPUT")){let e=n.type.toLowerCase();if(e==="submit"||e==="image")return!0}return!!(E(n,"BUTTON")&&n.type.toLowerCase()==="submit")}var D=class{el;select=null;modifiersState;eventEmitter;constructor(e,t){this.el=document.documentElement;let o=te(this.el);o&&this.setElement(o),this.modifiersState=e||new me,this.eventEmitter=t||new he}getElement(){return this.el}setElement(e){this.el=e,E(e,"OPTION")?this.select=ee(e,t=>E(t,"SELECT")):this.select=null}fireHtmlEvent(e){return this.eventEmitter.fireHtmlEvent(this.el,e)}fireKeyboardEvent(e,t){return this.eventEmitter.fireKeyboardEvent(this.el,e,t)}fireMouseEvent(e,t,o,s=null,i=0,u=!1){if(!u&&!I(this.el))return!1;if(s&&!(e===d.MOUSEOVER||e===d.MOUSEOUT))throw new T(12,`Event type does not allow related target: ${e}`);let a={clientX:t.x,clientY:t.y,button:o,altKey:this.modifiersState.isAltPressed(),ctrlKey:this.modifiersState.isControlPressed(),shiftKey:this.modifiersState.isShiftPressed(),metaKey:this.modifiersState.isMetaPressed(),wheelDelta:i,relatedTarget:s},f=this.select?this.getTargetOfOptionMouseEvent(e):this.el;return f?this.eventEmitter.fireMouseEvent(f,e,a):!0}fireTouchEvent(e,t,o,s,i){let u={touches:[],targetTouches:[],changedTouches:[],altKey:this.modifiersState.isAltPressed(),ctrlKey:this.modifiersState.isControlPressed(),shiftKey:this.modifiersState.isShiftPressed(),metaKey:this.modifiersState.isMetaPressed(),relatedTarget:null,scale:0,rotation:0,clientX:o.x,clientY:o.y},a=ne(this.el.ownerDocument),f=(b,h)=>{let v={identifier:b,screenX:h.x,screenY:h.y,clientX:h.x,clientY:h.y,pageX:h.x+a.x,pageY:h.y+a.y};u.changedTouches.push(v),(e===d.TOUCHSTART||e===d.TOUCHMOVE)&&(u.touches.push(v),u.targetTouches.push(v))};return f(t,o),s!==void 0&&i!==void 0&&f(s,i),this.eventEmitter.fireTouchEvent(this.el,e,u)}getTargetOfOptionMouseEvent(e){let t=this.select;switch(e){case d.CLICK:case d.MOUSEUP:return t.multiple?this.el:this.select;default:return t.multiple?this.el:null}}clickElement(e,t,o=!1){if(!o&&!I(this.el))return;let s=!this.select&&V(this.el),i=s&&W(this.el);this.fireMouseEvent(d.CLICK,e,t,null,0,o)&&s&&this.toggleRadioButtonOrCheckbox(i)}maybeToggleOption(){if(!this.select||!I(this.el))return;let e=this.select,t=W(this.el);t&&!e.multiple||(this.el.selected=!t,e.multiple||w(e,d.CHANGE))}toggleRadioButtonOrCheckbox(e){}focusOnElement(){let e=ee(this.el,o=>!!o&&E(o)&&Ke(o),!0);e=e||this.el;let t=te(e);return e===t?!1:(t&&typeof t.blur=="function"&&(E(t,"BODY")||t.blur()),typeof e.focus=="function"?(e.focus(),!0):!1)}submitForm(e){if(!We(e))throw new T(12,"Element is not a form, so could not submit.");w(e,d.SUBMIT)&&(E(e.submit)?HTMLFormElement.prototype.submit.call(e):e.submit())}};function Dt(n){return navigator.userAgent.includes(n)}function Te(){return Dt("Macintosh")}function P(n){try{return typeof n.selectionStart=="number"}catch{return!1}}function x(n,e){P(n)&&(n.selectionStart=e)}function z(n){return F(n)[0]}function K(n,e){P(n)&&(n.selectionEnd=e)}function be(n){return F(n)[1]}function F(n){let e=0,t=0;if(P(n)){let o=n;e=o.selectionStart,t=o.selectionEnd}return[e,t]}function J(n,e){if(P(n)){let t=n;t.selectionStart=e,t.selectionEnd=e}}function re(n,e){if(!P(n))throw new Error("Cannot set the selection end");let t=n,o=t.value,s=t.selectionStart,i=o.slice(0,s),u=o.slice(t.selectionEnd);t.value=i+e+u,t.selectionStart=s,t.selectionEnd=s+e.length}function ie(n){if(!P(n))throw new Error("Element does not support selection")}function Ne(n){return P(n)}var Oe=class{code;character;shiftChar;constructor(e,t,o){this.code=e,this.character=t||null,this.shiftChar=o||this.character}},le=new Map;function r(n,e,t){let o=new Oe(n,e,t);return e&&(!le.has(e)||t)&&(le.set(e,{key:o,shift:!1}),t&&le.set(t,{key:o,shift:!0})),o}var l={BACKSPACE:r(8),TAB:r(9),ENTER:r(13),SHIFT:r(16),CONTROL:r(17),ALT:r(18),PAUSE:r(19),CAPS_LOCK:r(20),ESC:r(27),SPACE:r(32," "),PAGE_UP:r(33),PAGE_DOWN:r(34),END:r(35),HOME:r(36),LEFT:r(37),UP:r(38),RIGHT:r(39),DOWN:r(40),PRINT_SCREEN:r(44),INSERT:r(45),DELETE:r(46),ZERO:r(48,"0",")"),ONE:r(49,"1","!"),TWO:r(50,"2","@"),THREE:r(51,"3","#"),FOUR:r(52,"4","$"),FIVE:r(53,"5","%"),SIX:r(54,"6","^"),SEVEN:r(55,"7","&"),EIGHT:r(56,"8","*"),NINE:r(57,"9","("),A:r(65,"a","A"),B:r(66,"b","B"),C:r(67,"c","C"),D:r(68,"d","D"),E:r(69,"e","E"),F:r(70,"f","F"),G:r(71,"g","G"),H:r(72,"h","H"),I:r(73,"i","I"),J:r(74,"j","J"),K:r(75,"k","K"),L:r(76,"l","L"),M:r(77,"m","M"),N:r(78,"n","N"),O:r(79,"o","O"),P:r(80,"p","P"),Q:r(81,"q","Q"),R:r(82,"r","R"),S:r(83,"s","S"),T:r(84,"t","T"),U:r(85,"u","U"),V:r(86,"v","V"),W:r(87,"w","W"),X:r(88,"x","X"),Y:r(89,"y","Y"),Z:r(90,"z","Z"),META:r(91),META_RIGHT:r(Te()?93:92),CONTEXT_MENU:r(Te()?0:null),NUM_ZERO:r(96,"0"),NUM_ONE:r(97,"1"),NUM_TWO:r(98,"2"),NUM_THREE:r(99,"3"),NUM_FOUR:r(100,"4"),NUM_FIVE:r(101,"5"),NUM_SIX:r(102,"6"),NUM_SEVEN:r(103,"7"),NUM_EIGHT:r(104,"8"),NUM_NINE:r(105,"9"),NUM_MULTIPLY:r(106,"*"),NUM_PLUS:r(107,"+"),NUM_MINUS:r(109,"-"),NUM_PERIOD:r(110,"."),NUM_DIVISION:r(111,"/"),NUM_LOCK:r(144),F1:r(112),F2:r(113),F3:r(114),F4:r(115),F5:r(116),F6:r(117),F7:r(118),F8:r(119),F9:r(120),F10:r(121),F11:r(122),F12:r(123),EQUALS:r(187,"=","+"),SEPARATOR:r(108,","),HYPHEN:r(189,"-","_"),COMMA:r(188,",","<"),PERIOD:r(190,".",">"),SLASH:r(191,"/","?"),BACKTICK:r(192,"`","~"),OPEN_BRACKET:r(219,"[","{"),BACKSLASH:r(220,"\\","|"),CLOSE_BRACKET:r(221,"]","}"),SEMICOLON:r(186,";",":"),APOSTROPHE:r(222,"'",'"')};function Ye(n){if(n.length!==1)throw new T(13,`Argument not a single character: ${n}`);let e=le.get(n);if(!e){let t=n.toUpperCase(),o=t.charCodeAt(0),s=r(o,n.toLowerCase(),t);e={key:s,shift:n!==s.character}}return e}var q=[l.ALT,l.CONTROL,l.META,l.SHIFT],xt=new Map([[l.SHIFT.code,1],[l.CONTROL.code,2],[l.ALT.code,4],[l.META.code,8]]),ge=`
 `,Z=class extends D{editable;currentPos=0;pressed=new Set;constructor(e){if(super(),this.editable=G(this.getElement()),e){for(let t of e.pressed)this.setKeyPressed(t,!0);this.currentPos=e.currentPos||0}}isPressed(e){return this.pressed.has(e)}setKeyPressed(e,t){if(q.includes(e)){let o=xt.get(e.code);this.modifiersState.setPressed(o,t)}t?this.pressed.add(e):this.pressed.delete(e)}pressKey(e){if(q.includes(e)&&this.isPressed(e))throw new T(13,"Cannot press a modifier key that is already pressed.");e.code!==null&&this.fireKeyEventInternal(d.KEYDOWN,e)&&(!this.requiresKeyPress(e)||this.fireKeyEventInternal(d.KEYPRESS,e,!1))&&(this.maybeSubmitForm(e),this.editable&&this.maybeEditText(e)),this.setKeyPressed(e,!0)}requiresKeyPress(e){return!!e.character||e===l.ENTER}maybeSubmitForm(e){if(e!==l.ENTER||!E(this.getElement(),"INPUT"))return;let t=pe(this.getElement());if(t){let o=t.getElementsByTagName("input");([...o].some(i=>Xe(i))||o.length===1)&&this.submitForm(t)}}maybeEditText(e){if(e.character)this.updateOnCharacter(e);else switch(e){case l.ENTER:this.updateOnEnter();break;case l.BACKSPACE:case l.DELETE:this.updateOnBackspaceOrDelete(e);break;case l.LEFT:case l.RIGHT:this.updateOnLeftOrRight(e);break;case l.HOME:case l.END:this.updateOnHomeOrEnd(e);break}}releaseKey(e){if(!this.isPressed(e))throw new T(13,`Cannot release a key that is not pressed. (${e.code})`);e.code!==null&&this.fireKeyEventInternal(d.KEYUP,e),this.setKeyPressed(e,!1)}getChar(e){if(!e.character)throw new T(13,"not a character key");return this.isPressed(l.SHIFT)?e.shiftChar:e.character}updateOnCharacter(e){let t=this.getChar(e),o=z(this.getElement())+1;Ne(this.getElement())?(re(this.getElement(),t),x(this.getElement(),o)):this.getElement().value+=t,this.fireHtmlEvent(d.TEXTINPUT),this.fireHtmlEvent(d.INPUT),this.updateCurrentPos(o)}updateOnEnter(){if(this.fireHtmlEvent(d.TEXTINPUT),E(this.getElement(),"TEXTAREA")){let e=z(this.getElement())+ge.length;Ne(this.getElement())?(re(this.getElement(),ge),x(this.getElement(),e)):this.getElement().value+=ge,this.fireHtmlEvent(d.INPUT),this.updateCurrentPos(e)}}updateOnBackspaceOrDelete(e){ie(this.getElement());let t=F(this.getElement());t[0]===t[1]&&(e===l.BACKSPACE?(x(this.getElement(),t[1]-1),K(this.getElement(),t[1])):K(this.getElement(),t[1]+1)),t=F(this.getElement());let o=this.getElement().value,s=!(t[0]===o.length||t[1]===0);re(this.getElement(),""),s&&this.fireHtmlEvent(d.INPUT),t=F(this.getElement()),this.updateCurrentPos(t[1])}updateOnLeftOrRight(e){ie(this.getElement());let t=this.getElement(),o=z(t),s=be(t),i,u=0,a=0;e===l.LEFT?this.isPressed(l.SHIFT)?this.currentPos===o?(u=Math.max(o-1,0),a=s,i=u):(u=o,a=s-1,i=a):i=o===s?Math.max(o-1,0):o:this.isPressed(l.SHIFT)?this.currentPos===s?(u=o,a=Math.min(s+1,t.value.length),i=a):(u=o+1,a=s,i=u):i=o===s?Math.min(s+1,t.value.length):s,this.isPressed(l.SHIFT)?(x(t,u),K(t,a)):J(t,i),this.updateCurrentPos(i)}updateOnHomeOrEnd(e){ie(this.getElement());let t=this.getElement(),o=z(t),s=be(t);if(e===l.HOME){if(this.isPressed(l.SHIFT)){x(t,0);let i=this.currentPos===o?s:o;K(t,i)}else J(t,0);this.updateCurrentPos(0)}else{let i=t.value.length;this.isPressed(l.SHIFT)?(this.currentPos===o&&x(t,s),K(t,i)):J(t,i),this.updateCurrentPos(i)}}updateCurrentPos(e){this.currentPos=e}fireKeyEventInternal(e,t,o=!1){if(t.code===null)throw new T(13,"Key must have a keycode to be fired.");let s={altKey:this.isPressed(l.ALT),ctrlKey:this.isPressed(l.CONTROL),metaKey:this.isPressed(l.META),shiftKey:this.isPressed(l.SHIFT),keyCode:t.code,charCode:t.character&&e===d.KEYPRESS?this.getChar(t).charCodeAt(0):0,preventDefault:o};return this.fireKeyboardEvent(e,s)}moveCursor(e){this.setElement(e),this.editable=G(e);let t=this.focusOnElement();if(this.editable&&t){let o=e.value.length;J(e,o),this.updateCurrentPos(o)}}getState(){return{pressed:[...this.pressed],currentPos:this.currentPos}}getModifiersState(){return this.modifiersState}};var k=new Map([[d.CLICK,[0,1,2,null]],[d.CONTEXTMENU,[null,null,2,null]],[d.MOUSEUP,[0,1,2,null]],[d.MOUSEOUT,[0,1,2,0]],[d.MOUSEMOVE,[0,1,2,0]]]);k.set(d.DBLCLICK,k.get(d.CLICK));k.set(d.MOUSEDOWN,k.get(d.MOUSEUP));k.set(d.MOUSEOVER,k.get(d.MOUSEOUT));function Ft(n){if(!I(n))throw new T(12,"Element is not currently interactable and may not be manipulated")}function Ge(n,e,t,o){n!==document.activeElement&&(Ft(n),ye(n));let s=t||new Z;s.moveCursor(n);function i(u){if(typeof u=="string")for(let a of u.split("")){let f=Ye(a),b=s.isPressed(l.SHIFT);f.shift&&!b&&s.pressKey(l.SHIFT),s.pressKey(f.key),s.releaseKey(f.key),f.shift&&!b&&s.releaseKey(l.SHIFT)}else q.includes(u)?s.isPressed(u)?s.releaseKey(u):s.pressKey(u):(s.pressKey(u),s.releaseKey(u))}if(n.type==="date"){let u=Array.isArray(e)?e=e.join(""):e,a=/\d{4}-\d{2}-\d{2}/,f=typeof u=="string"?u.match(a):null;if(f){w(n,d.TOUCHSTART),w(n,d.TOUCHEND),w(n,d.FOCUS),n.value=f[0],w(n,d.CHANGE),w(n,d.BLUR);return}}if(Array.isArray(e)?e.forEach(i):i(e),!o)for(let u of q)s.isPressed(u)&&s.releaseKey(u)}function ye(n,e){let t=L(n,e);if(t!=="scroll")return t==="none";if(n.scrollIntoView&&(n.scrollIntoView(),L(n,e)==="none"))return!0;let o=oe(n,e);for(let s=C(n);s;s=C(s))kt(o,s);return L(n,e)==="none"}function kt(n,e){let t=M(e),o=Ht(e),s=n.left-t.left-o.left,i=n.top-t.top-o.top,u=e.clientWidth+n.left-n.right,a=e.clientHeight+n.top-n.bottom;e.scrollLeft+=Math.min(s,Math.max(s-u,0)),e.scrollTop+=Math.min(i,Math.max(i-a,0))}function Ht(n){let e=n.ownerDocument.defaultView?.getComputedStyle(n),t=parseFloat(e?.borderLeftWidth||"0"),o=parseFloat(e?.borderRightWidth||"0"),s=parseFloat(e?.borderTopWidth||"0"),i=parseFloat(e?.borderBottomWidth||"0");return new X(s,o,i,t)}var ve="ELEMENT",Se="element-6066-11e4-a52e-4f735466cecf",Re="WINDOW";function Bt(n){return typeof n=="object"&&n!==null||typeof n=="function"}function Vt(n){return Array.isArray(n)?!0:typeof n=="object"&&n!==null&&typeof n.length=="number"}function Wt(n){function e(t,o){if(t==null)return null;switch(typeof t){case"string":case"number":case"boolean":return t;case"function":return t.toString();case"object":break;default:return null}let s=t;if(o.includes(s))throw new T(17,"Recursive object cannot be transferred");if("nodeType"in s&&(s.nodeType===Node.ELEMENT_NODE||s.nodeType===Node.DOCUMENT_NODE)){let u=$e(s);return{[ve]:u,[Se]:u}}if("document"in s)return{[Re]:$e(s)};if(Array.isArray(t))return t.map(u=>e(u,o));if(o=[...o,s],Vt(t))return Array.from(t).map(u=>e(u,o));let i={};for(let u of Object.keys(s))i[u]=e(s[u],o);return i}return e(n,[])}function ue(n,e){if(Array.isArray(n))return n.map(t=>ue(t,e));if(Bt(n)){if(typeof n=="function")return n;let t=n;if(ve in t)return Q(t[ve],e);if(Se in t)return Q(t[Se],e);if(Re in t)return Q(t[Re],e);let o={};for(let s of Object.keys(t))o[s]=ue(t[s],e);return o}return n}function ze(n){return{status:0,value:Wt(n)}}function Je(n){let e="code"in n&&typeof n.code=="number"?n.code:13;return{status:e,value:{message:n.message,error:ce(e)}}}var je="$wdc_",Xt=":wdc:";function Ce(n){let e=n||document,t=e[je];return t||(t=e[je]={},t.nextId=Date.now()),t.nextId||(t.nextId=Date.now()),t}var Yt=200;function qe(n,e){if("setInterval"in e)return!!e.closed;let t=e;for(;t;){if(t===n.documentElement)return!1;t.host&&t.nodeType===Node.DOCUMENT_FRAGMENT_NODE&&(t=t.host),t=t.parentNode}return!0}function Gt(n){let e=Ce(n);for(let t of Object.keys(e)){let o=e[t];!o||typeof o!="object"||qe(n,o)&&delete e[t]}}function $e(n){let e=n.ownerDocument,t=Ce(e);Object.keys(t).length-1>=Yt&&Gt(e||document);let o;for(let s of Object.keys(t))if(t[s]===n){o=s;break}return o||(o=Xt+t.nextId++,t[o]=n),o}function Q(n,e){n=decodeURIComponent(n);let t=e||document,o=Ce(t);if(!(n in o))throw new T(10,"Element does not exist in cache");let s=o[n];if(!qe(t,s))return s;throw delete o[n],"setInterval"in s?new T(23,"Window has been closed."):new T(10,"Element is no longer attached to the DOM")}var c={NULL:"\uE000",CANCEL:"\uE001",HELP:"\uE002",BACK_SPACE:"\uE003",TAB:"\uE004",CLEAR:"\uE005",RETURN:"\uE006",ENTER:"\uE007",SHIFT:"\uE008",CONTROL:"\uE009",ALT:"\uE00A",PAUSE:"\uE00B",ESCAPE:"\uE00C",SPACE:"\uE00D",PAGE_UP:"\uE00E",PAGE_DOWN:"\uE00F",END:"\uE010",HOME:"\uE011",ARROW_LEFT:"\uE012",LEFT:"\uE012",ARROW_UP:"\uE013",UP:"\uE013",ARROW_RIGHT:"\uE014",RIGHT:"\uE014",ARROW_DOWN:"\uE015",DOWN:"\uE015",INSERT:"\uE016",DELETE:"\uE017",SEMICOLON:"\uE018",EQUALS:"\uE019",NUMPAD0:"\uE01A",NUMPAD1:"\uE01B",NUMPAD2:"\uE01C",NUMPAD3:"\uE01D",NUMPAD4:"\uE01E",NUMPAD5:"\uE01F",NUMPAD6:"\uE020",NUMPAD7:"\uE021",NUMPAD8:"\uE022",NUMPAD9:"\uE023",MULTIPLY:"\uE024",ADD:"\uE025",SEPARATOR:"\uE026",SUBTRACT:"\uE027",DECIMAL:"\uE028",DIVIDE:"\uE029",F1:"\uE031",F2:"\uE032",F3:"\uE033",F4:"\uE034",F5:"\uE035",F6:"\uE036",F7:"\uE037",F8:"\uE038",F9:"\uE039",F10:"\uE03A",F11:"\uE03B",F12:"\uE03C",COMMAND:"\uE03D",META:"\uE03D"};function jt(n){return n>="\uE000"&&n<="\uE03D"}var $t=new Map([[c.NULL,null],[c.BACK_SPACE,l.BACKSPACE],[c.TAB,l.TAB],[c.RETURN,l.ENTER],[c.ENTER,l.ENTER],[c.SHIFT,l.SHIFT],[c.CONTROL,l.CONTROL],[c.ALT,l.ALT],[c.PAUSE,l.PAUSE],[c.ESCAPE,l.ESC],[c.SPACE,l.SPACE],[c.PAGE_UP,l.PAGE_UP],[c.PAGE_DOWN,l.PAGE_DOWN],[c.END,l.END],[c.HOME,l.HOME],[c.LEFT,l.LEFT],[c.UP,l.UP],[c.RIGHT,l.RIGHT],[c.DOWN,l.DOWN],[c.INSERT,l.INSERT],[c.DELETE,l.DELETE],[c.SEMICOLON,l.SEMICOLON],[c.EQUALS,l.EQUALS],[c.NUMPAD0,l.NUM_ZERO],[c.NUMPAD1,l.NUM_ONE],[c.NUMPAD2,l.NUM_TWO],[c.NUMPAD3,l.NUM_THREE],[c.NUMPAD4,l.NUM_FOUR],[c.NUMPAD5,l.NUM_FIVE],[c.NUMPAD6,l.NUM_SIX],[c.NUMPAD7,l.NUM_SEVEN],[c.NUMPAD8,l.NUM_EIGHT],[c.NUMPAD9,l.NUM_NINE],[c.MULTIPLY,l.NUM_MULTIPLY],[c.ADD,l.NUM_PLUS],[c.SUBTRACT,l.NUM_MINUS],[c.DECIMAL,l.NUM_PERIOD],[c.DIVIDE,l.NUM_DIVISION],[c.SEPARATOR,l.SEPARATOR],[c.F1,l.F1],[c.F2,l.F2],[c.F3,l.F3],[c.F4,l.F4],[c.F5,l.F5],[c.F6,l.F6],[c.F7,l.F7],[c.F8,l.F8],[c.F9,l.F9],[c.F10,l.F10],[c.F11,l.F11],[c.F12,l.F12],[c.META,l.META]]);function Ie(n,e,t,o){let s=!!o;function i(){return{persist:s,keys:[]}}let u=[],a=i();u.push(a);for(let f of e)for(let b of f.split(""))if(jt(b)){let h=$t.get(b);if(h===null)a=i(),u.push(a),s&&(a.persist=!1,a=i(),u.push(a));else if(h!==void 0)a.keys.push(h);else throw new Error(`Unsupported WebDriver key: \\u${b.charCodeAt(0).toString(16)}`)}else switch(b){case`
 `:a.keys.push(l.ENTER);break;case"	":a.keys.push(l.TAB);break;case"\b":a.keys.push(l.BACKSPACE);break;default:a.keys.push(b);break}for(let f of u)Ge(n,f.keys,t,f.persist)}function et(n){return n?Q(n.WINDOW):window}function qt(n,e,t){let o;try{let s=et(t),i=ue(e,s.document),u=n.apply(null,i);o=ze(u)}catch(s){o=Je(s)}return JSON.stringify(o)}function tt(n,e,t){return qt(Ie,[n,e],t)}return at(Zt);})();
